@@ -2,6 +2,7 @@
 """General testing utilities for CdEDB2 testsuite"""
 
 import collections.abc
+import contextlib
 import copy
 import datetime
 import decimal
@@ -23,14 +24,15 @@ import tempfile
 import unittest
 import urllib.parse
 from typing import (
-    Any, AnyStr, Callable, ClassVar, Dict, Iterable, List, Mapping, MutableMapping,
-    NamedTuple, Optional, Pattern, Sequence, Set, Tuple, Type, TypeVar, Union, cast,
-    no_type_check,
+    Any, AnyStr, Callable, ClassVar, Dict, Generator, Iterable, List, Mapping,
+    MutableMapping, NamedTuple, Optional, Pattern, Sequence, Set, Tuple, Type, TypeVar,
+    Union, cast, no_type_check,
 )
 
 import PIL.Image
 import webtest
 import webtest.utils
+
 from cdedb.backend.assembly import AssemblyBackend
 from cdedb.backend.cde import CdEBackend
 from cdedb.backend.common import AbstractBackend
@@ -40,8 +42,9 @@ from cdedb.backend.ml import MlBackend
 from cdedb.backend.past_event import PastEventBackend
 from cdedb.backend.session import SessionBackend
 from cdedb.common import (
-    ADMIN_VIEWS_COOKIE_NAME, ALL_ADMIN_VIEWS, CdEDBObject, CdEDBObjectMap, PathLike,
-    PrivilegeError, RequestState, nearly_now, now, roles_to_db_role, merge_dicts,
+    ADMIN_VIEWS_COOKIE_NAME, ALL_ADMIN_VIEWS, ANTI_CSRF_TOKEN_NAME,
+    ANTI_CSRF_TOKEN_PAYLOAD, CdEDBLog, CdEDBObject, CdEDBObjectMap, PathLike,
+    PrivilegeError, RequestState, merge_dicts, nearly_now, now, roles_to_db_role,
 )
 from cdedb.config import BasicConfig, Config, SecretsConfig
 from cdedb.database import DATABASE_ROLES
@@ -49,11 +52,11 @@ from cdedb.database.connection import connection_pool_factory
 from cdedb.frontend.application import Application
 from cdedb.frontend.common import AbstractFrontend, Worker, setup_translations
 from cdedb.frontend.cron import CronFrontend
+from cdedb.frontend.paths import CDEDB_PATHS
 from cdedb.query import QueryOperators
-from cdedb.script import setup
+from cdedb.script import Script
 
 _BASICCONF = BasicConfig()
-_SECRETSCONF = SecretsConfig()
 
 # TODO: use TypedDict to specify UserObject.
 UserObject = Mapping[str, Any]
@@ -146,7 +149,7 @@ def _make_backend_shim(backend: B, internal: bool = False) -> B:
     secrets = SecretsConfig(backend.conf._configpath)
     connpool = connection_pool_factory(
         backend.conf["CDB_DATABASE_NAME"], DATABASE_ROLES,
-        secrets, backend.conf["DB_PORT"])
+        secrets, backend.conf["DB_HOST"], backend.conf["DB_PORT"])
     translations = setup_translations(backend.conf)
 
     def setup_requeststate(key: Optional[str], ip: str = "127.0.0.0"
@@ -217,7 +220,11 @@ def _make_backend_shim(backend: B, internal: bool = False) -> B:
             @functools.wraps(attr)
             def wrapper(key: Optional[str], *args: Any, **kwargs: Any) -> Any:
                 rs = setup_requeststate(key)
-                return attr(rs, *args, **kwargs)
+                try:
+                    return attr(rs, *args, **kwargs)
+                except FileNotFoundError as e:
+                    raise RuntimeError("Did you forget to add a `@storage` decorator to"
+                                       " the test?") from e
 
             return wrapper
 
@@ -252,11 +259,17 @@ class BasicTest(unittest.TestCase):
             shutil.rmtree(self.storage_dir)
 
     @staticmethod
-    def get_sample_data(table: str, ids: Iterable[int],
-                        keys: Iterable[str]) -> CdEDBObjectMap:
+    def get_sample_data(table: str, ids: Iterable[int] = None,
+                        keys: Iterable[str] = None) -> CdEDBObjectMap:
         """This mocks a select request against the sample data.
 
         "SELECT <keys> FROM <table> WHERE id = ANY(<ids>)"
+
+        if `keys` is None:
+        "SELECT * FROM <table> WHERE id = ANY(<ids>)"
+
+        if `ids` is None:
+        "SELECT <keys> FROM <table>"
 
         For some fields of some tables we perform a type conversion. These
         should be added as necessary to ease comparison against backend results.
@@ -271,42 +284,46 @@ class BasicTest(unittest.TestCase):
                 return nearly_now()
             return datetime.datetime.fromisoformat(s)
 
+        if keys is None:
+            try:
+                keys = next(iter(_SAMPLE_DATA[table].values())).keys()
+            except StopIteration:
+                return {}
+        if ids is None:
+            ids = _SAMPLE_DATA[table].keys()
         # Turn Iterator into Collection and ensure consistent order.
         keys = tuple(keys)
         ret = {}
         for anid in ids:
-            if keys:
-                r = {}
-                for k in keys:
-                    r[k] = copy.deepcopy(_SAMPLE_DATA[table][anid][k])
-                    if table == 'core.personas':
-                        if k == 'balance':
-                            r[k] = decimal.Decimal(r[k])
-                        if k == 'birthday':
-                            r[k] = datetime.date.fromisoformat(r[k])
-                    if k in {'ctime', 'atime', 'vote_begin', 'vote_end',
-                             'vote_extension_end'}:
-                        r[k] = parse_datetime(r[k])
-                ret[anid] = r
-            else:
-                ret[anid] = copy.deepcopy(_SAMPLE_DATA[table][anid])
+            r = {}
+            for k in keys:
+                r[k] = copy.deepcopy(_SAMPLE_DATA[table][anid][k])
+                if table == 'core.personas':
+                    if k == 'balance' and r[k]:
+                        r[k] = decimal.Decimal(r[k])
+                    if k == 'birthday' and r[k]:
+                        r[k] = datetime.date.fromisoformat(r[k])
+                if k in {'ctime', 'atime', 'vote_begin', 'vote_end',
+                         'vote_extension_end', 'signup_end'} and r[k]:
+                    r[k] = parse_datetime(r[k])
+            ret[anid] = r
         return ret
 
     def get_sample_datum(self, table: str, id_: int) -> CdEDBObject:
-        return self.get_sample_data(table, [id_], [])[id_]
+        return self.get_sample_data(table, [id_])[id_]
 
 
 class CdEDBTest(BasicTest):
     """Reset the DB for every test."""
+    longMessage = False
 
     def setUp(self) -> None:
-        with setup(
+        with Script(
             persona_id=-1,
             dbuser="cdb",
-            dbpassword=_SECRETSCONF["CDB_DATABASE_ROLES"]["cdb"],
             dbname=self.conf["CDB_DATABASE_NAME"],
             check_system_user=False,
-        )().conn as conn:
+        ).rs().conn as conn:
             conn.set_session(autocommit=True)
             with conn.cursor() as curr:
                 with open("tests/ancillary_files/clean_data.sql") as f:
@@ -374,10 +391,50 @@ class BackendTest(CdEDBTest):
         self.key = ANONYMOUS
         self.user = USER_DICT["anonymous"]
 
+    @contextlib.contextmanager
+    def switch_user(self, new_user: UserIdentifier) -> Generator[None, None, None]:
+        """This method can be used as a context manager to temporarily switch users."""
+        old_user = self.user
+        self.logout()
+        self.login(new_user)
+        yield
+        self.logout()
+        self.login(old_user)
+
     def user_in(self, *identifiers: UserIdentifier) -> bool:
         """Check whether the current user is any of the given users."""
         users = {get_user(i)["id"] for i in identifiers}
         return self.user.get("id", -1) in users
+
+    def assertLogEqual(self, log_expectation: Sequence[CdEDBObject], *,
+                       realm: str = None,
+                       log_retriever: Callable[..., CdEDBLog] = None,
+                       **kwargs: Any) -> None:
+        """Helper to compare a log expectation to the actual thing."""
+        if realm and not log_retriever:
+            log_retriever = getattr(self, realm).retrieve_log
+        if log_retriever:
+            _, log = log_retriever(self.key, **kwargs)
+        else:
+            raise ValueError("No method of log retrieval provided.")
+
+        for real, exp in zip(log, log_expectation):
+            if 'id' not in exp:
+                del real['id']
+            if 'ctime' not in exp:
+                exp['ctime'] = nearly_now()
+            if 'submitted_by' not in exp:
+                exp['submitted_by'] = self.user['id']
+            for k in ('event_id', 'assembly_id', 'mailinglist_id'):
+                if k in kwargs and k not in exp:
+                    exp[k] = kwargs[k]
+            for k in ('persona_id', 'change_note'):
+                if k not in exp:
+                    exp[k] = None
+            for k in ('total', 'delta', 'new_balance'):
+                if exp.get(k):
+                    exp[k] = decimal.Decimal(exp[k])
+        self.assertEqual(log, tuple(log_expectation))
 
     @staticmethod
     def initialize_raw_backend(backendcls: Type[SessionBackend]
@@ -613,6 +670,16 @@ USER_DICT: Dict[str, UserObject] = {
         'family_name': "Finanzvorstand",
         'default_name_format': "Farin Finanzvorstand",
     },
+    "katarina": {
+        'id': 37,
+        'DB-ID': "DB-37-X",
+        'username': "katarina@example.cde",
+        'password': "secret",
+        'diplay_name': "Katarina",
+        'given_names': "Katarina",
+        'family_name': "Kassenprüfer",
+        'default_name_format': "Katarina Kassenprüfer",
+    },
     "viktor": {
         'id': 48,
         'DB-ID': "DB-48-5",
@@ -797,13 +864,13 @@ class FrontendTest(BackendTest):
     def _log_generation_time(self, response: webtest.TestResponse = None) -> None:
         if response is None:
             response = self.response
-        if _BASICCONF["TIMING_LOG"]:
-            with open(_BASICCONF["TIMING_LOG"], 'a') as f:
-                output = "{} {} {} {}\n".format(
-                    response.request.path, response.request.method,
-                    response.headers.get('X-Generation-Time'),
-                    response.request.query_string)
-                f.write(output)
+        # record performance information during test runs
+        with open(_BASICCONF["LOG_DIR"] / "cdedb-timing.log", 'a') as f:
+            output = "{} {} {} {}\n".format(
+                response.request.path, response.request.method,
+                response.headers.get('X-Generation-Time'),
+                response.request.query_string)
+            f.write(output)
 
     def get(self, url: str, *args: Any, verbose: bool = False, **kwargs: Any) -> None:
         """Navigate directly to a given URL using GET."""
@@ -818,11 +885,36 @@ class FrontendTest(BackendTest):
         if self.response != oldresponse:
             self._log_generation_time(oldresponse)
 
-    def post(self, url: str, *args: Any, verbose: bool = False, **kwargs: Any) -> None:
+    def assertRedirect(self, url: str, *args: Any, target_url: str,
+                       verbose: bool = False, **kwargs: Any) -> webtest.TestResponse:
+        """Checck that a GET-request to the url returns a redirect to the target url."""
+        response: webtest.TestResponse = self.app.get(url, *args, **kwargs)
+        self.assertLessEqual(300, response.status_int)
+        self.assertGreater(400, response.status_int)
+        self.assertIn("You should be redirected", response)
+        self.assertIn(target_url, response)
+        return response
+
+    def post(self, url: str, params: Dict[str, Any], *args: Any, verbose: bool = False,
+             evade_anti_csrf: bool = True, csrf_token_name: str = ANTI_CSRF_TOKEN_NAME,
+             csrf_token_payload: str = ANTI_CSRF_TOKEN_PAYLOAD, **kwargs: Any) -> None:
         """Directly send a POST-request.
 
-        Note that most of our POST-handlers require a CSRF-token."""
-        self.response = self.app.post(url, *args, **kwargs)
+        Note that most of our POST-handlers require an Anti-CSRF token,
+        which is forged here by default.
+
+        :param params: This is a restriction of self.app.post, but enforces a general
+            style and simplifies processing here.
+        :param evade_anti_csrf: Do CSRF, forging the Anti-CSRF token.
+        """
+        if evade_anti_csrf:
+            urlmap = CDEDB_PATHS
+            urls = urlmap.bind(self.app_extra_environ["HTTP_HOST"])
+            endpoint, _ = urls.match(url, method="POST")
+            params[csrf_token_name] = self.app.app.encode_anti_csrf_token(
+                endpoint, csrf_token_name, csrf_token_payload,
+                persona_id=self.user['id'])
+        self.response = self.app.post(url, params, *args, **kwargs)
         self.follow()
         self.basic_validate(verbose=verbose)
 
@@ -851,18 +943,17 @@ class FrontendTest(BackendTest):
             if "formmethod" in tmp_button.attrs:
                 form.method = tmp_button.attrs["formmethod"]
         method = form.method
+        if value and not button:
+            raise ValueError(
+                "Cannot specify button value without specifying button name.")
         self.response = form.submit(button, value=value)
         self.follow()
         self.basic_validate(verbose=verbose)
         if method == "POST" and check_notification:
             # check that we acknowledged the POST with a notification
-            success_str = "alert alert-success"
-            target = self.response.text
-            if verbose:
-                self.assertIn(success_str, target)
-            elif success_str not in target:
-                raise AssertionError(
-                    "Post request did not produce success notification.")
+            self.assertNotification(ntype='success',
+                                    msg=("No success notification found in"
+                                         + self.response.text if verbose else None))
 
     def traverse(self, *links: LinkIdentifier, verbose: bool = False) -> None:
         """Follow a sequence of links, described by their kwargs.
@@ -924,6 +1015,16 @@ class FrontendTest(BackendTest):
         self.submit(f, check_notification=False, verbose=verbose)
         self.key = ANONYMOUS
         self.user = USER_DICT["anonymous"]
+
+    @contextlib.contextmanager
+    def switch_user(self, new_user: UserIdentifier) -> Generator[None, None, None]:
+        """context manager to temporarily switch users - frontend variant
+
+        This restores the original response after the original user logged in again"""
+        saved_response = self.response
+        with super().switch_user(new_user):
+            yield
+        self.response = saved_response
 
     def admin_view_profile(self, user: UserIdentifier, check: bool = True,
                            verbose: bool = False) -> None:
@@ -999,6 +1100,7 @@ class FrontendTest(BackendTest):
         return body.get_content()
 
     def fetch_link(self, index: int = 0, num: int = 1) -> str:
+        """Extract the <num>th link out of the <index>th mail just sent."""
         for line in self.fetch_mail_content(index).splitlines():
             if line.startswith(f'[{num}] '):
                 return line.split(maxsplit=1)[-1]
@@ -1035,6 +1137,14 @@ class FrontendTest(BackendTest):
             self.fail("No valid html document.")
         if self.response.lxml.xpath("//*[@id='{}']".format(div)):
             self.fail("Element with id {} found".format(div))
+
+    def assertInputHasAttr(self, input_field: webtest.forms.Field, attr: str) -> None:
+        """Assert that the form input has a specific HTML DOM attribute.
+
+        This is no big logic, but should make this slightly internal feature of webtest
+        more easy to use.
+        """
+        self.assertIn(attr, input_field.attrs)
 
     def assertCheckbox(self, status: bool, anid: str) -> None:
         """Assert that the checkbox with the given id is checked (or not)."""
@@ -1094,6 +1204,34 @@ class FrontendTest(BackendTest):
             else:
                 self.assertNotIn(s.strip(), content.text_content())
 
+    def assertNotification(self, ntext: str = None, ntype: str = None, *,
+                           static: bool = False, msg: str = None) -> None:
+        """Check for a notification containing `ntext` under all `ntype` notifications.
+
+        :param ntext: Substring to be present in the notification's message.
+            If not given, only check for notification type.
+        :param ntype: type of notification. Can be any of bootstraps possible alert
+            contextes or 'error', which will expect a 'danger' alert.
+        :param static: whether to search for a static notification
+        :param msg: Custom message on assertion failure.
+        """
+        if ntype == 'error':  # allow this for convenience
+            ntype = 'danger'
+
+        div = 'static-notifications' if static else 'notifications'
+        alert_type_class = f" alert-{ntype}" if ntype is not None else ""
+        # source: https://devhints.io/xpath#string-functions
+        notifications = self.response.lxml.xpath(
+                f"//div[@id='{div}']/div[starts-with(@class,'alert{alert_type_class}')]"
+                "/span[@class='notificationMessage']")
+        self.assertTrue(notifications,
+                        msg=(f"No{alert_type_class} notification found."
+                             if msg is None else msg))
+        if ntext is not None:
+            # joining them this way is useful for meaningful failure message
+            all_texts = " | ".join(n.text_content().strip() for n in notifications)
+            self.assertIn(ntext, all_texts, msg=msg)
+
     def assertLogin(self, name: str) -> None:
         """Assert that a user is logged in by checking their display name."""
         span = self.response.lxml.xpath("//span[@id='displayname']")[0]
@@ -1113,17 +1251,52 @@ class FrontendTest(BackendTest):
             specify which one should be checked.
         :param message: The expected error message displayed below the input
         :param notification: The expected notification displayed at the top of the page
-            This can be a regex. If this is None, skip the notification check.
+            If this is None, skip the notification check.
         :raise AssertionError: If field is not found, field is not within
             .has-error container or error message is not found
         """
+        self._assertValidationComplaint(
+            kind="error", fieldname=fieldname, message=message, index=index,
+            notification=notification)
+
+    def assertValidationWarning(
+            self, fieldname: str, message: str = "", index: int = None,
+            notification: Optional[str] = "Eingaben scheinen fehlerhaft") -> None:
+        """
+        Check for a specific form input field to be highlighted as .has-warning
+        and a specific warning message to be shown near the field. Also check that an
+        .alert-warning notification (with the given text) is indicating validation
+        warning.
+
+        :param fieldname: The field's 'name' attribute
+        :param index: If more than one field with the given name exists,
+            specify which one should be checked.
+        :param message: The expected warning message displayed below the input
+        :param notification: The expected notification displayed at the top of the page
+            If this is None, skip the notification check.
+        :raise AssertionError: If field is not found, field is not within
+            .has-warning container or error message is not found
+        """
+        self._assertValidationComplaint(
+            kind="warning", fieldname=fieldname, message=message, index=index,
+            notification=notification)
+
+    def _assertValidationComplaint(
+            self, kind: str, fieldname: str, message: str, index: Optional[int],
+            notification: Optional[str]) -> None:
+        """Common helper for assertValidationError and assertValidationWarning."""
+        if kind == "error":
+            alert_type = "danger"
+        elif kind == "warning":
+            alert_type = "warning"
+        else:
+            raise NotImplementedError
+
         if notification is not None:
-            self.assertIn("alert alert-danger", self.response.text)
-            self.assertPresence(notification, div="notifications",
-                                regex=True)
+            self.assertNotification(notification, alert_type)
 
         nodes = self.response.lxml.xpath(
-            '(//input|//select|//textarea)[@name="{}"]'.format(fieldname))
+            f'(//input|//select|//textarea)[@name="{fieldname}"]')
         f = fieldname
         if index is None:
             if len(nodes) == 1:
@@ -1144,10 +1317,10 @@ class FrontendTest(BackendTest):
         # From https://devhints.io/xpath#class-check
         container = node.xpath(
             "ancestor::*[contains(concat(' ',normalize-space(@class),' '),"
-            "' has-error ')]")
+            f"' has-{kind} ')]")
         if not container:
             raise AssertionError(
-                f"Input with name {f!r} is not contained in an .has-error box")
+                f"Input with name {f!r} is not contained in an .has-{kind} box")
         msg = f"Expected error message not found near input with name {f!r}."
         self.assertIn(message, container[0].text_content(), msg)
 
@@ -1355,7 +1528,8 @@ class FrontendTest(BackendTest):
         self.submit(f)
         self.assertTitle("Zelda Zeruda-Hime")
         for key, value in data.items():
-            if key not in {'birthday', 'telephone', 'mobile', 'country', 'country2'}:
+            if key not in {'birthday', 'telephone', 'mobile', 'country', 'country2',
+                           'gender'}:
                 # Omit values with heavy formatting in the frontend here
                 self.assertPresence(value)
         # Now test archival
@@ -1410,11 +1584,13 @@ class FrontendTest(BackendTest):
         :return: The button element to perform further checks.
             Is actually of type `bs4.BeautifulSoup`.
         """
+        if isinstance(label, str):
+            label = re.compile(label)
         f = self.response.forms['adminviewstoggleform']
-        button = self.response.html\
-            .find(id="adminviewstoggleform")\
-            .find(text=label)\
-            .parent
+        button = self.response.html.find(id="adminviewstoggleform").find(text=label)
+        if not button:
+            raise KeyError(f"Admin view toggle with label {label!r} not found.")
+        button = button.parent
         if current_state is not None:
             if current_state:
                 self.assertIn("active", button['class'])
