@@ -16,10 +16,10 @@ import subprocess
 import sys
 from typing import Collection
 
-from psycopg2.extras import Json, DictCursor
+from psycopg2.extras import DictCursor, Json
 
 from cdedb.common import CdEDBObject
-from cdedb.script import setup
+from cdedb.script import Script
 
 # This is 'secret' the hashed
 PHASH = ("$6$rounds=60000$uvCUTc5OULJF/kT5$CNYWFoGXgEwhrZ0nXmbw0jlWvqi/"
@@ -118,7 +118,7 @@ def work(args: argparse.Namespace) -> None:
     with open(args.data_path, encoding='UTF-8') as infile:
         data = json.load(infile)
 
-    if data.get("EVENT_SCHEMA_VERSION") != [15, 3]:
+    if data.get("EVENT_SCHEMA_VERSION") != [15, 5]:
         raise RuntimeError("Version mismatch -- aborting.")
     if data["kind"] != "full":
         raise RuntimeError("Not a full export -- aborting.")
@@ -184,13 +184,11 @@ def work(args: argparse.Namespace) -> None:
         'event.course_choices', 'event.questionnaire_rows', 'event.log')
 
     print("Connect to database")
-    conn = setup(
-        persona_id=-1,
+    conn = Script(
         dbuser="cdb_admin",
-        dbpassword="9876543210abcdefghijklmnopqrst",
         dbname=db_name,
         check_system_user=False,
-    )().conn
+    )._conn
 
     with conn as con:
         with conn.cursor() as cur:
