@@ -10,11 +10,12 @@ from typing import Tuple, Type
 
 import cdedb.database.constants as const
 from cdedb.common import (
-    Accounts, AgeClasses, CourseChoiceToolActions, CourseFilterPositions,
-    LineResolutions, LodgementsSortkeys, TransactionType,
+    Accounts, AgeClasses, ConfidenceLevel, CourseChoiceToolActions,
+    CourseFilterPositions, GenesisDecision, LineResolutions, LodgementsSortkeys,
+    TransactionType,
 )
-from cdedb.subman.machine import SubscriptionAction, SubscriptionPolicy
 from cdedb.query import QueryOperators, QueryScope
+from cdedb.subman.machine import SubscriptionAction, SubscriptionPolicy
 
 #: The list of normal enums
 ALL_ENUMS: Tuple[Type[Enum], ...] = (
@@ -39,15 +40,18 @@ ALL_ENUMS: Tuple[Type[Enum], ...] = (
     const.MailinglistTypes,
     const.MailinglistDomain,
     const.QuestionnaireUsages,
+    const.EventPartGroupType,
     QueryOperators,
     QueryScope,
     AgeClasses,
     LineResolutions,
+    GenesisDecision,
     SubscriptionAction,
     SubscriptionPolicy,
     LodgementsSortkeys,
     Accounts,
     TransactionType,
+    ConfidenceLevel,
 )
 
 #: The list of infinite enums
