@@ -351,135 +351,105 @@ CDEDB_PATHS = werkzeug.routing.Map((
                      endpoint="archived_user_search"),)),
             rule("/registration/select", methods=_GET,
                  endpoint="select_registration"),
+            rule("/offline/partial", methods=_GET,
+                 endpoint="download_quick_partial_export"),
             rule("/user/create", methods=_GET,
                  endpoint="create_user_form"),
             rule("/user/create", methods=_POST,
                  endpoint="create_user"),
-            rule("/event/list", methods=_GET,
-                 endpoint="list_events"),
-            rule("/event/create", methods=_GET,
-                 endpoint="create_event_form"),
-            rule("/event/create", methods=_POST,
-                 endpoint="create_event"),
-            rule("/event/log", methods=_GET,
-                 endpoint="view_log"),
-            rule("/offline/partial", methods=_GET,
-                 endpoint="download_quick_partial_export"),
+            sub('/event', (
+                rule("/list", methods=_GET,
+                     endpoint="list_events"),
+                rule("/create", methods=_GET,
+                     endpoint="create_event_form"),
+                rule("/create", methods=_POST,
+                     endpoint="create_event"),
+                rule("/log", methods=_GET,
+                     endpoint="view_log"),
+            )),
             sub('/event/<int:event_id>', (
                 rule("/show", methods=_GET,
                      endpoint="show_event"),
                 rule("/log", methods=_GET,
                      endpoint="view_event_log"),
-                rule("/change", methods=_GET,
-                     endpoint="change_event_form"),
-                rule("/change", methods=_POST,
-                     endpoint="change_event"),
-                rule("/minorform", methods=_GET,
-                     endpoint="get_minor_form"),
-                rule("/orga/add", methods=_POST,
-                     endpoint="add_orga"),
-                rule("/orga/remove", methods=_POST,
-                     endpoint="remove_orga"),
-                rule("/minorform/change", methods=_POST,
-                     endpoint="change_minor_form"),
                 rule("/ml/create", methods=_POST,
                      endpoint="create_event_mailinglist"),
-                rule("/part/summary", methods=_GET,
-                     endpoint="part_summary"),
-                rule("/part/summary", methods=_POST,
-                     endpoint="part_summary"),
-                rule("/course/list", methods=_GET,
-                     endpoint="course_list"),
-                rule("/course/stats", methods=_GET,
-                     endpoint="course_stats"),
-                rule("/course/create", methods=_GET,
-                     endpoint="create_course_form"),
-                rule("/course/create", methods=_POST,
-                     endpoint="create_course"),
-                rule("/course/query", methods=_GET,
-                     endpoint="course_query"),
                 rule("/stats", methods=_GET,
                      endpoint="stats"),
-                rule("/course/choices", methods=_GET,
-                     endpoint="course_choices_form"),
-                rule("/course/choices", methods=_POST,
-                     endpoint="course_choices"),
-                rule("/course/checks", methods=_GET,
-                     endpoint="course_assignment_checks"),
+                rule("/notes", methods=_GET,
+                     endpoint="participant_info"),
                 rule("/export", methods=_GET,
                      endpoint="download_export"),
                 rule("/lock", methods=_POST,
                      endpoint="lock_event"),
                 rule("/unlock", methods=_POST,
                      endpoint="unlock_event"),
-                rule("/batchfees", methods=_GET,
-                     endpoint="batch_fees_form"),
-                rule("/batchfees", methods=_POST,
-                     endpoint="batch_fees"),
-                rule("/register", methods=_GET,
-                     endpoint="register_form"),
-                rule("/register", methods=_POST,
-                     endpoint="register"),
-                rule("/register/config", methods=_GET,
-                     endpoint="configure_registration_form"),
-                rule("/register/config", methods=_POST,
-                     endpoint="configure_registration"),
-                rule("/questionnaire/config", methods=_GET,
-                     endpoint="configure_additional_questionnaire_form"),
-                rule("/questionnaire/config", methods=_POST,
-                     endpoint="configure_additional_questionnaire"),
-                rule("/questionnaire/reorder", methods=_GET,
-                     endpoint="reorder_questionnaire_form"),
-                rule("/questionnaire/reorder", methods=_POST,
-                     endpoint="reorder_questionnaire"),
-                rule("/registration/add", methods=_GET,
-                     endpoint="add_registration_form"),
-                rule("/registration/add", methods=_POST,
-                     endpoint="add_registration"),
-                rule("/notes", methods=_GET,
-                     endpoint="participant_info"),
-                rule("/lodgement/overview", methods=_GET,
-                     endpoint="lodgements"),
-                rule("/lodgement/create", methods=_GET,
-                     endpoint="create_lodgement_form"),
-                rule("/lodgement/create", methods=_POST,
-                     endpoint="create_lodgement"),
-                rule("/lodgement/group/summary", methods=_GET,
-                     endpoint="lodgement_group_summary_form"),
-                rule("/lodgement/group/summary", methods=_POST,
-                     endpoint="lodgement_group_summary"),
-                rule("/lodgement/query", methods=_GET,
-                     endpoint="lodgement_query"),
-                rule("/lodgement/graph/show", methods=_GET,
-                     endpoint="lodgement_wishes_graph"),
-                rule("/lodgement/graph/form", methods=_GET,
-                     endpoint="lodgement_wishes_graph_form"),
-                rule("/registration/query", methods=_GET,
-                     endpoint="registration_query"),
-                rule("/checkin", methods=_GET,
-                     endpoint="checkin_form"),
-                rule("/checkin", methods=_POST,
-                     endpoint="checkin"),
-                rule("/field/setselect", methods=_GET,
-                     endpoint="field_set_select"),
-                rule("/field/summary", methods=_GET,
-                     endpoint="field_summary_form"),
-                rule("/field/summary", methods=_POST,
-                     endpoint="field_summary"),
-                rule("/query/store", methods=_POST,
-                     endpoint="store_event_query"),
-                rule("/query/delete", methods=_POST,
-                     endpoint="delete_event_query"),
-                rule("/download", methods=_GET,
-                     endpoint="downloads"),
-                rule("/import", methods=_GET,
-                     endpoint="partial_import_form"),
-                rule("/import", methods=_POST,
-                     endpoint="partial_import"),
                 rule("/archive", methods=_POST,
                      endpoint="archive_event"),
                 rule("/delete", methods=_POST,
                      endpoint="delete_event"),
+                rule("/change", methods=_GET,
+                     endpoint="change_event_form"),
+                rule("/change", methods=_POST,
+                     endpoint="change_event"),
+                rule("/batchfees", methods=_GET,
+                     endpoint="batch_fees_form"),
+                rule("/batchfees", methods=_POST,
+                     endpoint="batch_fees"),
+                rule("/checkin", methods=_GET,
+                     endpoint="checkin_form"),
+                rule("/checkin", methods=_POST,
+                     endpoint="checkin"),
+                sub('/minorform', (
+                    rule("/get", methods=_GET,
+                         endpoint="get_minor_form"),
+                    rule("/change", methods=_POST,
+                         endpoint="change_minor_form"),
+                )),
+                sub('/orga', (
+                    rule("/add", methods=_POST,
+                         endpoint="add_orga"),
+                    rule("/remove", methods=_POST,
+                         endpoint="remove_orga"),
+                )),
+                sub('/query', (
+                    rule("/store", methods=_POST,
+                         endpoint="store_event_query"),
+                    rule("/delete", methods=_POST,
+                         endpoint="delete_event_query"),
+                )),
+                rule("/register", methods=_GET,
+                     endpoint="register_form"),
+                rule("/register", methods=_POST,
+                     endpoint="register"),
+                sub('/register', (
+                    rule("/config", methods=_GET,
+                         endpoint="configure_registration_form"),
+                    rule("/config", methods=_POST,
+                         endpoint="configure_registration"),
+                )),
+                sub('/questionnaire', (
+                    rule("/config", methods=_GET,
+                         endpoint="configure_additional_questionnaire_form"),
+                    rule("/config", methods=_POST,
+                         endpoint="configure_additional_questionnaire"),
+                    rule("/reorder", methods=_GET,
+                         endpoint="reorder_questionnaire_form"),
+                    rule("/reorder", methods=_POST,
+                         endpoint="reorder_questionnaire"),
+                )),
+                sub("/import", (
+                    rule("/partial", methods=_GET,
+                         endpoint="partial_import_form"),
+                    rule("/partial", methods=_POST,
+                         endpoint="partial_import"),
+                    rule("/questionnaire", methods=_GET,
+                         endpoint="questionnaire_import_form"),
+                    rule("/questionnaire", methods=_POST,
+                         endpoint="questionnaire_import"),
+                )),
+                rule("/download", methods=_GET,
+                     endpoint="downloads"),
                 sub("/download", (
                     rule("/nametag", methods=_GET,
                          endpoint="download_nametags"),
@@ -505,31 +475,76 @@ CDEDB_PATHS = werkzeug.routing.Map((
                          endpoint="download_csv_lodgements"),
                     rule("/csv_registrations", methods=_GET,
                          endpoint="download_csv_registrations"),)),
-                rule("/part/add", methods=_GET,
-                     endpoint="add_part_form"),
-                rule("/part/add", methods=_POST,
-                     endpoint="add_part"),
-                sub('/part/<int:part_id>', (
-                    rule("/change", methods=_GET,
-                         endpoint="change_part_form"),
-                    rule("/change", methods=_POST,
-                         endpoint="change_part"),
-                    rule("/delete", methods=_POST,
-                         endpoint="delete_part"),)),
-                sub('/course/<int:course_id>', (
-                    rule("/show", methods=_GET,
-                         endpoint="show_course"),
-                    rule("/change", methods=_GET,
-                         endpoint="change_course_form"),
-                    rule("/change", methods=_POST,
-                         endpoint="change_course"),
-                    rule("/delete", methods=_POST,
-                         endpoint="delete_course"),
-                    rule("/manage", methods=_GET,
-                         endpoint="manage_attendees_form"),
-                    rule("/manage", methods=_POST,
-                         endpoint="manage_attendees"),)),
+                sub('/part', (
+                    rule("/summary", methods=_GET,
+                         endpoint="part_summary"),
+                    rule("/add", methods=_GET,
+                         endpoint="add_part_form"),
+                    rule("/add", methods=_POST,
+                         endpoint="add_part"),
+                    sub('/<int:part_id>', (
+                        rule("/change", methods=_GET,
+                             endpoint="change_part_form"),
+                        rule("/change", methods=_POST,
+                             endpoint="change_part"),
+                        rule("/delete", methods=_POST,
+                             endpoint="delete_part"),)),
+                    sub('/group', (
+                        rule("/summary", methods=_GET,
+                             endpoint="part_group_summary"),
+                        rule("/add", methods=_GET,
+                             endpoint="add_part_group_form"),
+                        rule("/add", methods=_POST,
+                             endpoint="add_part_group"),
+                        sub('/<int:part_group_id>', (
+                            rule("/change", methods=_GET,
+                                 endpoint="change_part_group_form"),
+                            rule("/change", methods=_POST,
+                                 endpoint="change_part_group"),
+                            rule("/delete", methods=_POST,
+                                 endpoint="delete_part_group"),
+                        )),
+                    )),
+                )),
+                sub('/course', (
+                    rule("/list", methods=_GET,
+                         endpoint="course_list"),
+                    rule("/stats", methods=_GET,
+                         endpoint="course_stats"),
+                    rule("/create", methods=_GET,
+                         endpoint="create_course_form"),
+                    rule("/create", methods=_POST,
+                         endpoint="create_course"),
+                    rule("/query", methods=_GET,
+                         endpoint="course_query"),
+                    rule("/choices", methods=_GET,
+                         endpoint="course_choices_form"),
+                    rule("/choices", methods=_POST,
+                         endpoint="course_choices"),
+                    rule("/checks", methods=_GET,
+                         endpoint="course_assignment_checks"),
+                    sub('/<int:course_id>', (
+                        rule("/show", methods=_GET,
+                             endpoint="show_course"),
+                        rule("/change", methods=_GET,
+                             endpoint="change_course_form"),
+                        rule("/change", methods=_POST,
+                             endpoint="change_course"),
+                        rule("/delete", methods=_POST,
+                             endpoint="delete_course"),
+                        rule("/manage", methods=_GET,
+                             endpoint="manage_attendees_form"),
+                        rule("/manage", methods=_POST,
+                             endpoint="manage_attendees"),
+                    )),
+                )),
                 sub('/registration', (
+                    rule("/add", methods=_GET,
+                         endpoint="add_registration_form"),
+                    rule("/add", methods=_POST,
+                         endpoint="add_registration"),
+                    rule("/query", methods=_GET,
+                         endpoint="registration_query"),
                     rule("/status", methods=_GET,
                          endpoint="registration_status"),
                     rule("/amend", methods=_GET,
@@ -547,37 +562,73 @@ CDEDB_PATHS = werkzeug.routing.Map((
                     rule("/multiedit", methods=_GET,
                          endpoint="change_registrations_form"),
                     rule("/multiedit", methods=_POST,
-                         endpoint="change_registrations"),)),
-                sub('/registration/<int:registration_id>', (
-                    rule("/show", methods=_GET,
-                         endpoint="show_registration"),
-                    rule("/change", methods=_GET,
-                         endpoint="change_registration_form"),
-                    rule("/change", methods=_POST,
-                         endpoint="change_registration"),
-                    rule("/delete", methods=_POST,
-                         endpoint="delete_registration"),)),
-                sub('/lodgement/<int:lodgement_id>', (
-                    rule("/change", methods=_GET,
-                         endpoint="change_lodgement_form"),
-                    rule("/change", methods=_POST,
-                         endpoint="change_lodgement"),
-                    rule("/delete", methods=_POST,
-                         endpoint="delete_lodgement"),
-                    rule("/manage", methods=_GET,
-                         endpoint="manage_inhabitants_form"),
-                    rule("/manage", methods=_POST,
-                         endpoint="manage_inhabitants"),
-                    rule("/swap", methods=_POST,
-                         endpoint="swap_inhabitants"),
-                    rule("/show", methods=_GET,
-                         endpoint="show_lodgement"),)),
-                sub('/field/<int:field_id>', (
-                    rule("/set", methods=_GET,
-                         endpoint="field_set_form"),
-                    rule("/set", methods=_POST,
-                         endpoint="field_set"),)),)),
-        )),)),
+                         endpoint="change_registrations"),
+                    sub('/<int:registration_id>', (
+                        rule("/show", methods=_GET,
+                             endpoint="show_registration"),
+                        rule("/change", methods=_GET,
+                             endpoint="change_registration_form"),
+                        rule("/change", methods=_POST,
+                             endpoint="change_registration"),
+                        rule("/delete", methods=_POST,
+                             endpoint="delete_registration"),
+                    )),
+                )),
+                sub('/lodgement', (
+                    rule("/overview", methods=_GET,
+                         endpoint="lodgements"),
+                    rule("/create", methods=_GET,
+                         endpoint="create_lodgement_form"),
+                    rule("/create", methods=_POST,
+                         endpoint="create_lodgement"),
+                    rule("/query", methods=_GET,
+                         endpoint="lodgement_query"),
+                    sub('/group', (
+                        rule("/summary", methods=_GET,
+                             endpoint="lodgement_group_summary_form"),
+                        rule("/summary", methods=_POST,
+                             endpoint="lodgement_group_summary"),
+                    )),
+                    sub('/graph', (
+                        rule("/show", methods=_GET,
+                             endpoint="lodgement_wishes_graph"),
+                        rule("/form", methods=_GET,
+                             endpoint="lodgement_wishes_graph_form"),
+                    )),
+                    sub('/<int:lodgement_id>', (
+                        rule("/change", methods=_GET,
+                             endpoint="change_lodgement_form"),
+                        rule("/change", methods=_POST,
+                             endpoint="change_lodgement"),
+                        rule("/delete", methods=_POST,
+                             endpoint="delete_lodgement"),
+                        rule("/manage", methods=_GET,
+                             endpoint="manage_inhabitants_form"),
+                        rule("/manage", methods=_POST,
+                             endpoint="manage_inhabitants"),
+                        rule("/swap", methods=_POST,
+                             endpoint="swap_inhabitants"),
+                        rule("/show", methods=_GET,
+                             endpoint="show_lodgement"),
+                    )),
+                )),
+                sub('/field', (
+                    rule("/setselect", methods=_GET,
+                         endpoint="field_set_select"),
+                    rule("/summary", methods=_GET,
+                         endpoint="field_summary_form"),
+                    rule("/summary", methods=_POST,
+                         endpoint="field_summary"),
+                    sub('/<int:field_id>', (
+                        rule("/set", methods=_GET,
+                             endpoint="field_set_form"),
+                        rule("/set", methods=_POST,
+                             endpoint="field_set"),
+                    )),
+                )),
+            )),
+        )),
+    )),
     werkzeug.routing.EndpointPrefix('assembly/', (
         sub('/assembly', (
             rule("/", methods=_GET,
@@ -632,29 +683,32 @@ CDEDB_PATHS = werkzeug.routing.Map((
                     rule("/add", methods=_POST,
                          endpoint="add_attachment"),
                     sub("/<int:attachment_id>", (
-                        rule("/get", methods=_GET,
+                        rule("/latest", methods=_GET,
                              endpoint="get_attachment"),
-                        rule("/show", methods=_GET,
-                             endpoint="show_attachment"),
-                        rule("/change", methods=_GET,
-                             endpoint="change_attachment_link_form"),
-                        rule("/change", methods=_POST,
-                             endpoint="change_attachment_link"),
+                        # allow this simple link for convenience
+                        rule("/", endpoint="",
+                             redirect_to='assembly/assembly/<assembly_id>/attachment'
+                                         '/<attachment_id>/latest'),
+                        # this is kept as legacy support
+                        rule("/get", endpoint="",
+                             redirect_to='assembly/assembly/<assembly_id>/attachment'
+                                         '/<attachment_id>/latest'),
                         rule("/delete", methods=_POST,
                              endpoint="delete_attachment"),
                         rule("/add", methods=_GET,
-                             endpoint="add_attachment_form"),
+                             endpoint="add_attachment_version_form"),
                         rule("/add", methods=_POST,
-                             endpoint="add_attachment"),
-                        sub("/version/<int:version>", (
-                            rule("/get", methods=_GET,
-                                 endpoint="get_attachment"),
+                             endpoint="add_attachment_version"),
+                        sub("/version/<int:version_nr>", (
+                            rule("/", methods=_GET,
+                                 endpoint="get_attachment_version"),
+                            # this is kept as legacy support
+                            rule("/get", endpoint="",
+                                 redirect_to="assembly/assembly/<assembly_id>"
+                                             "/attachment/<attachment_id>/version"
+                                             "/<version_nr>"),
                             rule("/delete", methods=_POST,
-                                 endpoint="delete_attachment"),
-                            rule("/edit", methods=_GET,
-                                 endpoint="edit_attachment_version_form"),
-                            rule("/edit", methods=_POST,
-                                 endpoint="edit_attachment_version"),
+                                 endpoint="delete_attachment_version"),
                         )),
                     )),
                 )),
@@ -673,6 +727,10 @@ CDEDB_PATHS = werkzeug.routing.Map((
                          endpoint="change_ballot_form"),
                     rule("/change", methods=_POST,
                          endpoint="change_ballot"),
+                    rule("/comment", methods=_GET,
+                         endpoint="comment_concluded_ballot_form"),
+                    rule("/comment", methods=_POST,
+                         endpoint="comment_concluded_ballot"),
                     rule("/delete", methods=_POST,
                          endpoint="delete_ballot"),
                     rule("/start", methods=_POST,
@@ -686,34 +744,17 @@ CDEDB_PATHS = werkzeug.routing.Map((
                     rule("/result/download", methods=_GET,
                          endpoint="get_result"),
                     sub("/attachment", (
-                        rule("/add", methods=_GET,
-                             endpoint="add_attachment_form"),
-                        rule("/add", methods=_POST,
-                             endpoint="add_attachment"),
                         sub("/<int:attachment_id>", (
-                            rule("/get", methods=_GET,
-                                 endpoint="get_attachment"),
-                            rule("/show", methods=_GET,
-                                 endpoint="show_attachment"),
-                            rule("/change", methods=_GET,
-                                 endpoint="change_attachment_link_form"),
-                            rule("/change", methods=_POST,
-                                 endpoint="change_attachment_link"),
-                            rule("/delete", methods=_POST,
-                                 endpoint="delete_attachment"),
-                            rule("/add", methods=_GET,
-                                 endpoint="add_attachment_form"),
-                            rule("/add", methods=_POST,
-                                 endpoint="add_attachment"),
+                            # this is kept as legacy support
+                            rule("/get", endpoint="",
+                                 redirect_to='assembly/assembly/<assembly_id>'
+                                             '/attachment/<attachment_id>/latest'),
                             sub("/version/<int:version>", (
-                                rule("/get", methods=_GET,
-                                     endpoint="get_attachment"),
-                                rule("/delete", methods=_POST,
-                                     endpoint="delete_attachment"),
-                                rule("/edit", methods=_GET,
-                                     endpoint="edit_attachment_version_form"),
-                                rule("/edit", methods=_POST,
-                                     endpoint="edit_attachment_version"),
+                                # this is kept as legacy support
+                                rule("/get", endpoint="",
+                                     redirect_to="assembly/assembly/<assembly_id>"
+                                                 "/attachment/<attachment_id>/version"
+                                                 "/<version>"),
                             )),
                         )),
                     )),
