@@ -12,15 +12,17 @@ up an environment for passing a query from frontend to backend.
 import collections
 import copy
 import enum
+import itertools
 import re
-from typing import Any, Collection, Dict, Tuple
-
-from cdedb.common import (
-    CdEDBObject, CdEDBObjectMap, RequestState, EntitySorter, n_,
-    get_localized_country_codes, xsorted,
+from typing import (
+    Any, Callable, Collection, Dict, List, Mapping, NamedTuple, Optional, Tuple, Union,
 )
+
 import cdedb.database.constants as const
-from cdedb.filter import enum_entries_filter, keydictsort_filter
+from cdedb.common import (
+    ADMIN_KEYS, CdEDBObject, CdEDBObjectMap, EntitySorter, RequestState, n_, xsorted,
+)
+from cdedb.filter import keydictsort_filter
 
 
 @enum.unique
@@ -95,8 +97,40 @@ MULTI_VALUE_OPERATORS = {_ops.oneof, _ops.otherthan, _ops.containsall,
 NO_VALUE_OPERATORS = {_ops.empty, _ops.nonempty}
 
 
+# A query constraint translates to (part of) a WHERE clause. All constraints are
+# conjugated.
 QueryConstraint = Tuple[str, QueryOperators, Any]
+# A query order translate to an ORDER BY clause. The bool decides whether the sorting
+# is ASC (i.e. True -> ASC, False -> DESC).
 QueryOrder = Tuple[str, bool]
+
+QueryChoices = Mapping[Union[int, str, enum.Enum], str]
+
+
+class QuerySpecEntry(NamedTuple):
+    type: str
+    title_base: str
+    title_prefix: str = ""
+    title_params: Dict[str, str] = {}
+    choices: QueryChoices = {}
+
+    # Mask gettext so pybabel doesn't try to extract the f-string.
+    def get_title(self, g: Callable[[str], str]) -> str:
+        ret_prefix = g(f"{self.title_prefix}: ") if self.title_prefix else ""
+        ret_title = g(self.title_base).format(**self.title_params)
+        return ret_prefix + ret_title
+
+    def replace_choices(self, choices: QueryChoices) -> "QuerySpecEntry":
+        return self.__class__(
+            type=self.type,
+            title_base=self.title_base,
+            title_prefix=self.title_prefix,
+            title_params=self.title_params,
+            choices=choices,
+        )
+
+
+QuerySpec = Dict[str, QuerySpecEntry]
 
 
 class QueryScope(enum.IntEnum):
@@ -142,7 +176,9 @@ class QueryScope(enum.IntEnum):
             return ret.split(".", 1)[1]
         return ret
 
-    def get_spec(self, *, event: CdEDBObject = None) -> Dict[str, str]:
+    def get_spec(self, *, event: CdEDBObject = None, courses: CdEDBObjectMap = None,
+                 lodgements: CdEDBObjectMap = None,
+                 lodgement_groups: CdEDBObjectMap = None) -> QuerySpec:
         """Return the query spec for this scope.
 
         These may be enriched by ext-fields. Order is important for UI purposes.
@@ -153,6 +189,9 @@ class QueryScope(enum.IntEnum):
         :param event: For some scopes, the spec is dependent on specific event data.
             For these scopes (see `event_spec_map` below) this must be provided.
             The format should be like the return of `EventBackend.get_event()`.
+        :param courses: Same as `event`.
+        :param lodgements: Same as `event`.
+        :param lodgement_groups: Same as `event`.
         """
         event_spec_map = {
             QueryScope.registration: make_registration_query_spec,
@@ -164,7 +203,7 @@ class QueryScope(enum.IntEnum):
                 raise ValueError(n_("Constructing the query spec for %(scope)s"
                                     " requires additional event information."),
                                  {"scope": self})
-            return event_spec_map[self](event)
+            return event_spec_map[self](event, courses, lodgements, lodgement_groups)
 
         return copy.deepcopy(_QUERY_SPECS[self])
 
@@ -173,7 +212,7 @@ class QueryScope(enum.IntEnum):
         return self in {QueryScope.registration, QueryScope.lodgement,
                         QueryScope.event_course}
 
-    def get_target(self, *, prepend_realm: bool = True) -> str:
+    def get_target(self, *, redirect: bool = True) -> str:
         """For scopes that support storing, where to redirect to after storing."""
         if self == QueryScope.registration:
             realm, target = "event", "registration_query"
@@ -183,9 +222,7 @@ class QueryScope(enum.IntEnum):
             realm, target = "event", "course_query"
         else:
             realm, target = "", ""
-        if prepend_realm:
-            return f"{realm}/{target}"
-        return target
+        return f"{realm if redirect else 'query'}/{target}"
 
     def mangle_query_input(self, rs: RequestState, defaults: CdEDBObject = None,
                            ) -> Dict[str, str]:
@@ -275,229 +312,221 @@ PRIMARY_KEYS = {
 
 # See QueryScope.get_spec().
 _QUERY_SPECS = {
-    QueryScope.persona:  # query for a persona without past event infos
-        collections.OrderedDict([
-            ("personas.id", "id"),
-            ("given_names", "str"),
-            ("family_name", "str"),
-            ("username", "str"),
-            ("display_name", "str"),
-            ("is_active", "bool"),
-            ("is_ml_realm", "bool"),
-            ("is_event_realm", "bool"),
-            ("is_assembly_realm", "bool"),
-            ("is_cde_realm", "bool"),
-            ("is_member", "bool"),
-            ("is_searchable", "bool"),
-            ("is_ml_admin", "bool"),
-            ("is_event_admin", "bool"),
-            ("is_assembly_admin", "bool"),
-            ("is_cde_admin", "bool"),
-            ("is_core_admin", "bool"),
-            ("is_meta_admin", "bool"),
-            ("notes", "str"),
-            ("fulltext", "str"),
-        ]),
-    QueryScope.core_user:  # query for a general user including past event infos
-        collections.OrderedDict([
-            ("personas.id", "id"),
-            ("given_names", "str"),
-            ("family_name", "str"),
-            ("username", "str"),
-            ("display_name", "str"),
-            ("is_active", "bool"),
-            ("is_ml_realm", "bool"),
-            ("is_assembly_realm", "bool"),
-            ("is_event_realm", "bool"),
-            ("is_cde_realm", "bool"),
-            ("is_member", "bool"),
-            ("is_searchable", "bool"),
-            ("is_ml_admin", "bool"),
-            ("is_event_admin", "bool"),
-            ("is_assembly_admin", "bool"),
-            ("is_cde_admin", "bool"),
-            ("is_core_admin", "bool"),
-            ("is_meta_admin", "bool"),
-            ("is_ml_admin,is_event_admin,is_assembly_admin,is_cde_admin,"
-             "is_core_admin,is_meta_admin", "bool"),
-            ("pevent_id", "id"),
-            ("notes", "str"),
-            ("fulltext", "str"),
-        ]),
+    QueryScope.persona:
+        {
+            "personas.id": QuerySpecEntry("id", n_("ID")),
+            "given_names": QuerySpecEntry("str", n_("Given Names")),
+            "family_name": QuerySpecEntry("str", n_("Family Name")),
+            "username": QuerySpecEntry("str", n_("E-Mail")),
+            "display_name": QuerySpecEntry("str", n_("Known as (Forename)")),
+            "birth_name": QuerySpecEntry("str", n_("Birth Name")),
+            "gender": QuerySpecEntry("int", n_("Gender")),
+            "birthday": QuerySpecEntry("date", n_("Birthday")),
+            "is_active": QuerySpecEntry("bool", n_("Active Account")),
+            "is_ml_realm": QuerySpecEntry("bool", n_("Mailinglists"), n_("Realm")),
+            "is_event_realm": QuerySpecEntry("bool", n_("Events"), n_("Realm")),
+            "is_assembly_realm": QuerySpecEntry("bool", n_("Assemblies"), n_("Realm")),
+            "is_cde_realm": QuerySpecEntry("bool", n_("cde_realm"), n_("Realm")),
+            "is_member": QuerySpecEntry("bool", n_("CdE-Member")),
+            "is_searchable": QuerySpecEntry("bool", n_("Searchable")),
+            **{
+                k: QuerySpecEntry("bool", k, n_("Admin"))
+                for k in ADMIN_KEYS
+            },
+            ",".join(ADMIN_KEYS): QuerySpecEntry("bool", n_("Any"), n_("Admin")),
+            "notes": QuerySpecEntry("str", n_("Admin-Notes")),
+            "fulltext": QuerySpecEntry("str", n_("Fulltext")),
+        },
     QueryScope.cde_user:
-        collections.OrderedDict([
-            ("personas.id", "id"),
-            ("given_names", "str"),
-            ("family_name", "str"),
-            ("username", "str"),
-            ("display_name", "str"),
-            ("title", "str"),
-            ("name_supplement", "str"),
-            ("birth_name", "str"),
-            ("gender", "int"),
-            ("birthday", "date"),
-            ("telephone", "str"),
-            ("mobile", "str"),
-            ("address", "str"),
-            ("address_supplement", "str"),
-            ("postal_code", "str"),
-            ("location", "str"),
-            ("country", "str"),
-            ("address2", "str"),
-            ("address_supplement2", "str"),
-            ("postal_code2", "str"),
-            ("location2", "str"),
-            ("country2", "str"),
-            ("is_active", "bool"),
-            ("is_member", "bool"),
-            ("trial_member", "bool"),
-            ("paper_expuls", "bool"),
-            ("is_searchable", "bool"),
-            ("decided_search", "bool"),
-            ("balance", "float"),
-            ("is_ml_admin", "bool"),
-            ("is_event_admin", "bool"),
-            ("is_assembly_admin", "bool"),
-            ("is_cde_admin", "bool"),
-            ("is_core_admin", "bool"),
-            ("is_meta_admin", "bool"),
-            ("weblink", "str"),
-            ("specialisation", "str"),
-            ("affiliation", "str"),
-            ("timeline", "str"),
-            ("interests", "str"),
-            ("free_form", "str"),
-            ("pevent_id", "id"),
-            ("pcourse_id", "id"),
-            ("notes", "str"),
-            ("fulltext", "str"),
-            ("lastschrift.granted_at", "datetime"),
-            ("lastschrift.revoked_at", "datetime"),
-            ("lastschrift.active_lastschrift", "bool"),
-            ("lastschrift.amount", "float"),
-        ]),
+        {
+            "personas.id": QuerySpecEntry("id", n_("ID")),
+            "given_names": QuerySpecEntry("str", n_("Given Names")),
+            "family_name": QuerySpecEntry("str", n_("Family Name")),
+            "username": QuerySpecEntry("str", n_("E-Mail")),
+            "display_name": QuerySpecEntry("str", n_("Known as (Forename)")),
+            "title": QuerySpecEntry("str", n_("Title_[[of a persona]]")),
+            "name_supplement": QuerySpecEntry("str", n_("Name Affix")),
+            "birth_name": QuerySpecEntry("str", n_("Birth Name")),
+            "gender": QuerySpecEntry("int", n_("Gender")),
+            "birthday": QuerySpecEntry("date", n_("Birthday")),
+            "telephone": QuerySpecEntry("str", n_("Phone")),
+            "mobile": QuerySpecEntry("str", n_("Mobile Phone")),
+            "address": QuerySpecEntry("str", n_("Address")),
+            "address_supplement": QuerySpecEntry("str", n_("Address Supplement")),
+            "postal_code": QuerySpecEntry("str", n_("ZIP")),
+            "location": QuerySpecEntry("str", n_("City")),
+            "country": QuerySpecEntry("str", n_("Country")),
+            "address2": QuerySpecEntry("str", n_("Address (2)")),
+            "address_supplement2": QuerySpecEntry("str", n_("Address Supplement (2)")),
+            "postal_code2": QuerySpecEntry("str", n_("ZIP (2)")),
+            "location2": QuerySpecEntry("str", n_("City (2)")),
+            "country2": QuerySpecEntry("str", n_("Country (2)")),
+            "is_active": QuerySpecEntry("bool", n_("Active Account")),
+            "is_member": QuerySpecEntry("bool", n_("CdE-Member")),
+            "trial_member": QuerySpecEntry("bool", n_("Trial Member")),
+            "paper_expuls": QuerySpecEntry("bool", n_("Printed exPuls")),
+            "is_searchable": QuerySpecEntry("bool", n_("Searchable")),
+            "decided_search": QuerySpecEntry("bool", n_("Searchability Decided")),
+            "balance": QuerySpecEntry("float", n_("Membership-Fee Balance")),
+            **{
+                k: QuerySpecEntry("bool", k, n_("Admin"))
+                for k in ADMIN_KEYS
+            },
+            ",".join(ADMIN_KEYS): QuerySpecEntry("bool", n_("Any"), n_("Admin")),
+            "weblink": QuerySpecEntry("str", n_("WWW")),
+            "specialisation": QuerySpecEntry("str", n_("Specialisation")),
+            "affiliation": QuerySpecEntry("str", n_("School, University, …")),
+            "timeline": QuerySpecEntry("str", n_("Year(s) of Graduation")),
+            "interests": QuerySpecEntry("str", n_("Interests")),
+            "free_form": QuerySpecEntry("str", n_("Miscellaneous")),
+            "pevent_id": QuerySpecEntry("id", n_("Past Event")),
+            "pcourse_id": QuerySpecEntry("id", n_("Past Course")),
+            "lastschrift.granted_at": QuerySpecEntry(
+                "datetime", n_("Lastschrift Granted")),
+            "lastschrift.revoked_at": QuerySpecEntry(
+                "datetime", n_("Lastschrift Revoked")),
+            "lastschrift.active_lastschrift": QuerySpecEntry(
+                "bool", n_("Active Lastschrift")),
+            "lastschrift.amount": QuerySpecEntry("float", n_("Lastschrift Amount")),
+            "notes": QuerySpecEntry("str", n_("Admin-Notes")),
+        },
     QueryScope.event_user:
-        collections.OrderedDict([
-            ("personas.id", "id"),
-            ("given_names", "str"),
-            ("family_name", "str"),
-            ("username", "str"),
-            ("display_name", "str"),
-            ("title", "str"),
-            ("name_supplement", "str"),
-            ("gender", "int"),
-            ("birthday", "date"),
-            ("telephone", "str"),
-            ("mobile", "str"),
-            ("address", "str"),
-            ("address_supplement", "str"),
-            ("postal_code", "str"),
-            ("location", "str"),
-            ("country", "str"),
-            ("is_active", "bool"),
-            ("is_cde_realm", "bool"),
-            ("is_member", "bool"),
-            ("is_searchable", "bool"),
-            ("is_event_admin", "bool"),
-            ("is_ml_admin", "bool"),
-            ("notes", "str"),
-            ("fulltext", "str"),
-        ]),
+        {
+            "personas.id": QuerySpecEntry("id", n_("ID")),
+            "given_names": QuerySpecEntry("str", n_("Given Names")),
+            "family_name": QuerySpecEntry("str", n_("Family Name")),
+            "username": QuerySpecEntry("str", n_("E-Mail")),
+            "display_name": QuerySpecEntry("str", n_("Known as (Forename)")),
+            "title": QuerySpecEntry("str", n_("Title_[[of a persona]]")),
+            "name_supplement": QuerySpecEntry("str", n_("Name Affix")),
+            "gender": QuerySpecEntry("int", n_("Gender")),
+            "birthday": QuerySpecEntry("date", n_("Birthday")),
+            "telephone": QuerySpecEntry("str", n_("Phone")),
+            "mobile": QuerySpecEntry("str", n_("Mobile Phone")),
+            "address": QuerySpecEntry("str", n_("Address")),
+            "address_supplement": QuerySpecEntry("str", n_("Address Supplement")),
+            "postal_code": QuerySpecEntry("str", n_("ZIP")),
+            "location": QuerySpecEntry("str", n_("City")),
+            "country": QuerySpecEntry("str", n_("Country")),
+            "is_active": QuerySpecEntry("bool", n_("Active Account")),
+            "is_member": QuerySpecEntry("bool", n_("CdE-Member")),
+            "is_searchable": QuerySpecEntry("bool", n_("Searchable")),
+            **{
+                k: QuerySpecEntry("bool", k, n_("Admin"))
+                for k in ADMIN_KEYS
+            },
+            ",".join(ADMIN_KEYS): QuerySpecEntry("bool", n_("Any"), n_("Admin")),
+            "pevent_id": QuerySpecEntry("id", n_("Past Event")),
+            "pcourse_id": QuerySpecEntry("id", n_("Past Course")),
+            "notes": QuerySpecEntry("str", n_("Admin-Notes")),
+        },
     QueryScope.past_event_user:
-        collections.OrderedDict([
-            ("personas.id", "id"),
-            ("given_names", "str"),
-            ("family_name", "str"),
-            ("username", "str"),
-            ("display_name", "str"),
-            ("title", "str"),
-            ("name_supplement", "str"),
-            ("birthday", "date"),
-            ("telephone", "str"),
-            ("mobile", "str"),
-            ("address", "str"),
-            ("address_supplement", "str"),
-            ("postal_code", "str"),
-            ("location", "str"),
-            ("country", "str"),
-            ("pevent_id", "id"),
-            ("pcourse_id", "id"),
-        ]),
+        {
+            "personas.id": QuerySpecEntry("id", n_("ID")),
+            "given_names": QuerySpecEntry("str", n_("Given Names")),
+            "family_name": QuerySpecEntry("str", n_("Family Name")),
+            "username": QuerySpecEntry("str", n_("E-Mail")),
+            "display_name": QuerySpecEntry("str", n_("Known as (Forename)")),
+            "birth_name": QuerySpecEntry("str", n_("Birth Name")),
+            "title": QuerySpecEntry("str", n_("Title_[[of a persona]]")),
+            "name_supplement": QuerySpecEntry("str", n_("Name Affix")),
+            "birthday": QuerySpecEntry("date", n_("Birthday")),
+            "telephone": QuerySpecEntry("str", n_("Phone")),
+            "mobile": QuerySpecEntry("str", n_("Mobile Phone")),
+            "address": QuerySpecEntry("str", n_("Address")),
+            "address_supplement": QuerySpecEntry("str", n_("Address Supplement")),
+            "postal_code": QuerySpecEntry("str", n_("ZIP")),
+            "location": QuerySpecEntry("str", n_("City")),
+            "country": QuerySpecEntry("str", n_("Country")),
+            "is_cde_realm": QuerySpecEntry("bool", n_("cde_realm"), n_("Realm")),
+            "pevent_id": QuerySpecEntry("id", n_("Past Event")),
+            "pcourse_id": QuerySpecEntry("id", n_("Past Course")),
+            "notes": QuerySpecEntry("str", n_("Admin-Notes")),
+        },
     QueryScope.archived_persona:
-        collections.OrderedDict([
-            ("personas.id", "id"),
-            ("given_names", "str"),
-            ("family_name", "str"),
-            ("display_name", "str"),
-            ("notes", "str"),
-        ]),
+        {
+            "personas.id": QuerySpecEntry("id", n_("ID")),
+            "given_names": QuerySpecEntry("str", n_("Given Names")),
+            "family_name": QuerySpecEntry("str", n_("Family Name")),
+            "display_name": QuerySpecEntry("str", n_("Known as (Forename)")),
+            "birth_name": QuerySpecEntry("str", n_("Birth Name")),
+            "gender": QuerySpecEntry("int", n_("Gender")),
+            "birthday": QuerySpecEntry("date", n_("Birthday")),
+            "is_ml_realm": QuerySpecEntry("bool", n_("Mailinglists"), n_("Realm")),
+            "is_event_realm": QuerySpecEntry("bool", n_("Events"), n_("Realm")),
+            "is_assembly_realm": QuerySpecEntry("bool", n_("Assemblies"), n_("Realm")),
+            "is_cde_realm": QuerySpecEntry("bool", n_("cde_realm"), n_("Realm")),
+            "pevent_id": QuerySpecEntry("id", n_("Past Event")),
+            "notes": QuerySpecEntry("str", n_("Admin-Notes")),
+        },
     QueryScope.archived_core_user:
-        collections.OrderedDict([
-            ("personas.id", "id"),
-            ("given_names", "str"),
-            ("family_name", "str"),
-            ("display_name", "str"),
-            ("birth_name", "str"),
-            ("gender", "int"),
-            ("birthday", "date"),
-            ("pevent_id", "id"),
-            ("notes", "str"),
-            ("is_ml_realm", "bool"),
-            ("is_event_realm", "bool"),
-            ("is_assembly_realm", "bool"),
-            ("is_cde_realm", "bool"),
-        ]),
+        {
+            "personas.id": QuerySpecEntry("id", n_("ID")),
+            "given_names": QuerySpecEntry("str", n_("Given Names")),
+            "family_name": QuerySpecEntry("str", n_("Family Name")),
+            "display_name": QuerySpecEntry("str", n_("Known as (Forename)")),
+            "is_ml_realm": QuerySpecEntry("bool", n_("Mailinglists"), n_("Realm")),
+            "is_event_realm": QuerySpecEntry("bool", n_("Events"), n_("Realm")),
+            "is_assembly_realm": QuerySpecEntry("bool", n_("Assemblies"), n_("Realm")),
+            "is_cde_realm": QuerySpecEntry("bool", n_("cde_realm"), n_("Realm")),
+            "notes": QuerySpecEntry("str", n_("Admin-Notes")),
+        },
     QueryScope.archived_past_event_user:
-        collections.OrderedDict([
-            ("personas.id", "id"),
-            ("given_names", "str"),
-            ("family_name", "str"),
-            ("display_name", "str"),
-            ("birth_name", "str"),
-            ("gender", "int"),
-            ("birthday", "date"),
-            ("pevent_id", "id"),
-            ("notes", "str"),
-        ]),
+        {
+            "personas.id": QuerySpecEntry("id", n_("ID")),
+            "given_names": QuerySpecEntry("str", n_("Given Names")),
+            "family_name": QuerySpecEntry("str", n_("Family Name")),
+            "display_name": QuerySpecEntry("str", n_("Known as (Forename)")),
+            "birth_name": QuerySpecEntry("str", n_("Birth Name")),
+            "birthday": QuerySpecEntry("date", n_("Birthday")),
+            "is_cde_realm": QuerySpecEntry("bool", n_("cde_realm"), n_("Realm")),
+            "pevent_id": QuerySpecEntry("id", n_("Past Event")),
+            "pcourse_id": QuerySpecEntry("id", n_("Past Course")),
+            "notes": QuerySpecEntry("str", n_("Admin-Notes")),
+        },
     QueryScope.cde_member:
-        collections.OrderedDict([
-            ("personas.id", "id"),
-            ("given_names,display_name", "str"),
-            ("family_name,birth_name", "str"),
-            ("username", "str"),
-            ("address,address_supplement,address2,address_supplement2", "str"),
-            ("postal_code,postal_code2", "str"),
-            ("telephone,mobile", "str"),
-            ("location,location2", "str"),
-            ("country,country2", "str"),
-            ("weblink,specialisation,affiliation,timeline,interests,free_form",
-             "str"),
-            ("pevent_id", "id"),
-            ("pcourse_id", "id"),
-            ("fulltext", "str"),
-        ]),
+        {
+            "personas.id": QuerySpecEntry("id", n_("ID")),
+            "given_names,display_name": QuerySpecEntry("str", n_("Given Names")),
+            "family_name,birth_name": QuerySpecEntry("str", n_("Family Name")),
+            "username": QuerySpecEntry("str", n_("E-Mail")),
+            "address,address_supplement,address2,address_supplement2":
+                QuerySpecEntry("str", n_("Address")),
+            "postal_code,postal_code2": QuerySpecEntry("str", n_("ZIP")),
+            "location,location2": QuerySpecEntry("str", n_("City")),
+            "country,country2": QuerySpecEntry("str", n_("Country")),
+            "telephone,mobile": QuerySpecEntry("str", n_("Phone")),
+            "weblink,specialisation,affiliation,timeline,interests,free_form":
+                QuerySpecEntry("str", n_("Interests")),
+            "pevent_id": QuerySpecEntry("id", n_("Past Event")),
+            "pcourse_id": QuerySpecEntry("id", n_("Past Course")),
+            "fulltext": QuerySpecEntry("str", n_("Fulltext")),
+        },
     QueryScope.quick_registration:
-        collections.OrderedDict([
-            ("registrations.id", "id"),
-            ("given_names", "str"),
-            ("family_name", "str"),
-            ("username", "str"),
-            ("display_name", "str"),
-            ("title", "str"),
-            ("name_supplement", "str"),
-        ]),
+        {
+            "registrations.id": QuerySpecEntry("id", n_("ID")),
+            "given_names": QuerySpecEntry("str", n_("Given Names")),
+            "family_name": QuerySpecEntry("str", n_("Family Name")),
+            "username": QuerySpecEntry("str", n_("E-Mail")),
+            "display_name": QuerySpecEntry("str", n_("Known as (Forename)")),
+            "title": QuerySpecEntry("str", n_("Title_[[of a persona]]")),
+            "name_supplement": QuerySpecEntry("str", n_("Name Affix")),
+        },
     QueryScope.past_event_course:
-        collections.OrderedDict([
-            ("courses.id", "id"),
-            ("courses.pcourse_id", "id"),
-            ("courses.pevent_id", "id"),
-            ("courses.nr", "str"),
-            ("courses.title", "str"),
-            ("courses.description", "str"),
-            ("events.title", "str"),
-            ("events.tempus", "date")
-        ]),
+        {
+            "courses.id": QuerySpecEntry("id", n_("course ID")),
+            "courses.pcourse_id": QuerySpecEntry("id", n_("course")),
+            "courses.pevent_id": QuerySpecEntry("id", n_("Past Event")),
+            "courses.nr": QuerySpecEntry("str", n_("course nr")),
+            "courses.title": QuerySpecEntry("str", n_("course title")),
+            "courses.description": QuerySpecEntry("str", n_("course description")),
+            "events.title": QuerySpecEntry(
+                "str", n_("Title_[[name of an entity]]"), n_("Past Event")),
+            "events.tempus": QuerySpecEntry(
+                "date", n_("Cutoff date"), n_("Past Event")),
+        },
 }
+_QUERY_SPECS[QueryScope.core_user] = _QUERY_SPECS[QueryScope.persona]
 _QUERY_SPECS[QueryScope.ml_user] = _QUERY_SPECS[QueryScope.persona]
 _QUERY_SPECS[QueryScope.assembly_user] = _QUERY_SPECS[QueryScope.persona]
 
@@ -524,7 +553,7 @@ class Query:
     everything.
     """
 
-    def __init__(self, scope: QueryScope, spec: CdEDBObject,
+    def __init__(self, scope: QueryScope, spec: QuerySpec,
                  fields_of_interest: Collection[str],
                  constraints: Collection[QueryConstraint],
                  order: Collection[QueryOrder],
@@ -545,7 +574,7 @@ class Query:
             the id of the entry in the corresponding table.
         """
         self.scope = scope
-        self.spec = spec
+        self.spec = dict(spec)
         self.fields_of_interest = list(fields_of_interest)
         self.constraints = list(constraints)
         self.order = list(order)
@@ -624,11 +653,11 @@ class Query:
         return params
 
     def get_field_format_spec(self, field: str) -> QueryResultEntryFormat:
-        if self.spec[field] == "date":
+        if self.spec[field].type == "date":
             return QueryResultEntryFormat.date
-        if self.spec[field] == "datetime":
+        if self.spec[field].type == "datetime":
             return QueryResultEntryFormat.datetime
-        if self.spec[field] == "bool":
+        if self.spec[field].type == "bool":
             return QueryResultEntryFormat.bool
         if self.scope == QueryScope.registration:
             if field == "persona.id":
@@ -656,704 +685,476 @@ class Query:
         return QueryResultEntryFormat.other
 
 
-def make_registration_query_spec(event: CdEDBObject) -> Dict[str, str]:
+def _sort_event_fields(fields: CdEDBObjectMap
+                       ) -> Dict[const.FieldAssociations, List[CdEDBObject]]:
+    """Helper to sort event fields and group them by association."""
+    sorted_fields: Dict[const.FieldAssociations, List[CdEDBObject]] = {
+        association: [] for association in const.FieldAssociations}
+    for field in xsorted(fields.values(), key=EntitySorter.event_field):
+        sorted_fields[field['association']].append(field)
+    return sorted_fields
+
+
+def _combine_specs(spec_map: Dict[int, QuerySpec], entity_ids: Collection[int],
+                   prefix: str) -> QuerySpec:
+    """Helper to create combined spec entries for specified entities.
+
+    Entries are grouped by their position in the individual spec. Thus the individual
+    specs need to be ordered in the same way. They need not have the same length.
+    If the spec for one entity is shorter, that entity will simple be ignored when
+    creating the combinations.
+    """
+    ret: QuerySpec = {}
+    entity_ids = xsorted(entity_ids)
+    if len(entity_ids) <= 1:
+        return ret
+
+    # Choose the "longest" spec to serve as a reference.
+    reference_spec = max(spec_map.values(), key=len)
+
+    # Create a two dimensional grid of spec keys. First dimension are the given entites,
+    # second dimension are the keys.
+    all_keys = tuple(tuple(spec_map[id_].keys()) for id_ in entity_ids)
+    for i, k in enumerate(reference_spec):
+        # Exclude keys that are not present in all specs.
+        relevant_keys = tuple(keys[i] for keys in all_keys if len(keys) > i)
+        # Do not combine entries consisting of only one key.
+        if len(relevant_keys) <= 1:
+            continue
+        key = ",".join(relevant_keys)
+        entry = reference_spec[k]
+        ret[key] = QuerySpecEntry(
+            type=entry.type, title_base=entry.title_base, title_prefix=prefix,
+            title_params=entry.title_params, choices=entry.choices,
+        )
+    return ret
+
+
+def _get_course_choices(courses: Optional[CdEDBObjectMap]) -> QueryChoices:
+    if courses is None:
+        return {}
+    course_identifier = lambda c: "{}. {}".format(c["nr"], c["shortname"])
+    return dict((c_id, course_identifier(c))
+                for c_id, c in keydictsort_filter(courses, EntitySorter.course))
+
+
+def _get_lodgement_choices(lodgements: Optional[CdEDBObjectMap]) -> QueryChoices:
+    if lodgements is None:
+        return {}
+    lodge_identifier = lambda l: l["title"]
+    return dict(
+        (l_id, lodge_identifier(l))
+        for l_id, l in keydictsort_filter(lodgements, EntitySorter.lodgement))
+
+
+def _get_lodgement_group_choices(lodgement_groups: Optional[CdEDBObjectMap]
+                                 ) -> QueryChoices:
+    if lodgement_groups is None:
+        return {}
+    lodgement_group_identifier = lambda g: g["title"]
+    return dict(
+        (g_id, lodgement_group_identifier(g))
+        for g_id, g in keydictsort_filter(
+            lodgement_groups, EntitySorter.lodgement_group))
+
+
+def make_registration_query_spec(event: CdEDBObject, courses: CdEDBObjectMap = None,
+                                 lodgements: CdEDBObjectMap = None,
+                                 lodgement_groups: CdEDBObjectMap = None) -> QuerySpec:
     """Helper to generate ``QueryScope.registration``'s spec.
 
     Since each event has dynamic columns for parts and extra fields we
     have amend the query spec on the fly.
     """
 
-    tracks = event['tracks']
-    spec = collections.OrderedDict([
-        ("reg.id", "id"),
-        ("persona.id", "id"),
-        ("persona.given_names", "str"),
-        ("persona.family_name", "str"),
-        ("persona.username", "str"),
-        ("persona.is_member", "bool"),
-        ("persona.display_name", "str"),
-        ("persona.title", "str"),
-        ("persona.name_supplement", "str"),
-        ("persona.gender", "int"),
-        ("persona.birthday", "date"),
-        ("persona.telephone", "str"),
-        ("persona.mobile", "str"),
-        ("persona.address", "str"),
-        ("persona.address_supplement", "str"),
-        ("persona.postal_code", "str"),
-        ("persona.location", "str"),
-        ("persona.country", "str"),
-        ("reg.payment", "date"),
-        ("reg.amount_paid", "float"),
-        ("reg.amount_owed", "float"),
-        ("reg.parental_agreement", "bool"),
-        ("reg.mixed_lodging", "bool"),
-        ("reg.list_consent", "bool"),
-        ("reg.notes", "str"),
-        ("reg.orga_notes", "str"),
-        ("reg.checkin", "datetime"),
-        ("ctime.creation_time", "datetime"),
-        ("mtime.modification_time", "datetime"),
-    ])
-    # note that spec is an ordered dict and we should respect the order
-    for part_id, part in keydictsort_filter(event['parts'],
-                                            EntitySorter.event_part):
-        spec["part{0}.status".format(part_id)] = "int"
-        spec["part{0}.is_camping_mat".format(part_id)] = "bool"
-        spec["part{0}.lodgement_id".format(part_id)] = "id"
-        spec["lodgement{0}.id".format(part_id)] = "id"
-        spec["lodgement{0}.group_id".format(part_id)] = "id"
-        spec["lodgement{0}.title".format(part_id)] = "str"
-        spec["lodgement{0}.notes".format(part_id)] = "str"
-        for f in xsorted(event['fields'].values(),
-                            key=EntitySorter.event_field):
-            if f['association'] == const.FieldAssociations.lodgement:
-                temp = "lodgement{0}.xfield_{1}"
-                kind = const.FieldDatatypes(f['kind']).name
-                spec[temp.format(part_id, f['field_name'])] = kind
-        spec["lodgement_group{0}.id".format(part_id)] = "id"
-        spec["lodgement_group{0}.title".format(part_id)] = "str"
-        ordered_tracks = keydictsort_filter(
-            part['tracks'], EntitySorter.course_track)
-        for track_id, track in ordered_tracks:
-            spec["track{0}.is_course_instructor".format(track_id)] \
-                = "bool"
-            spec["track{0}.course_id".format(track_id)] = "int"
-            spec["track{0}.course_instructor".format(track_id)] = "int"
-            for temp in ("course", "course_instructor",):
-                spec["{1}{0}.id".format(track_id, temp)] = "id"
-                spec["{1}{0}.nr".format(track_id, temp)] = "str"
-                spec["{1}{0}.title".format(track_id, temp)] = "str"
-                spec["{1}{0}.shortname".format(track_id, temp)] = "str"
-                spec["{1}{0}.notes".format(track_id, temp)] = "str"
-                for f in xsorted(event['fields'].values(),
-                                    key=EntitySorter.event_field):
-                    if f['association'] == const.FieldAssociations.course:
-                        key = "{1}{0}.xfield_{2}".format(
-                            track_id, temp, f['field_name'])
-                        kind = const.FieldDatatypes(f['kind']).name
-                        spec[key] = kind
-            for i in range(track['num_choices']):
-                spec[f"course_choices{track_id}.rank{i}"] = "int"
-            if track['num_choices'] > 1:
-                spec[",".join(f"course_choices{track_id}.rank{i}"
-                                for i in range(track['num_choices']))] = "int"
-    if len(event['parts']) > 1:
-        spec[",".join("part{0}.status".format(part_id)
-                        for part_id in event['parts'])] = "int"
-        spec[",".join("part{0}.is_camping_mat".format(part_id)
-                        for part_id in event['parts'])] = "bool"
-        spec[",".join("part{0}.lodgement_id".format(part_id)
-                        for part_id in event['parts'])] = "id"
-        spec[",".join("lodgement{0}.id".format(part_id)
-                        for part_id in event['parts'])] = "id"
-        spec[",".join("lodgement{0}.group_id".format(part_id)
-                        for part_id in event['parts'])] = "id"
-        spec[",".join("lodgement{0}.title".format(part_id)
-                        for part_id in event['parts'])] = "str"
-        spec[",".join("lodgement{0}.notes".format(part_id)
-                        for part_id in event['parts'])] = "str"
-        spec[",".join("lodgement_group{0}.id".format(part_id)
-                        for part_id in event['parts'])] = "id"
-        spec[",".join("lodgement_group{0}.title".format(part_id)
-                        for part_id in event['parts'])] = "str"
-        for f in xsorted(event['fields'].values(),
-                            key=EntitySorter.event_field):
-            if f['association'] == const.FieldAssociations.lodgement:
-                key = ",".join(
-                    "lodgement{0}.xfield_{1}".format(
-                        part_id, f['field_name'])
-                    for part_id in event['parts'])
-                kind = const.FieldDatatypes(f['kind']).name
-                spec[key] = kind
-    if len(tracks) > 1:
-        spec[",".join("track{0}.is_course_instructor".format(track_id)
-                        for track_id in tracks)] = "bool"
-        spec[",".join("track{0}.course_id".format(track_id)
-                        for track_id in tracks)] = "bool"
-        spec[",".join("track{0}.course_instructor".format(track_id)
-                        for track_id in tracks)] = "int"
-        for temp in ("course", "course_instructor",):
-            spec[",".join("{1}{0}.id".format(track_id, temp)
-                            for track_id in tracks)] = "id"
-            spec[",".join("{1}{0}.nr".format(track_id, temp)
-                            for track_id in tracks)] = "str"
-            spec[",".join("{1}{0}.title".format(track_id, temp)
-                            for track_id in tracks)] = "str"
-            spec[",".join("{1}{0}.shortname".format(track_id, temp)
-                            for track_id in tracks)] = "str"
-            spec[",".join("{1}{0}.notes".format(track_id, temp)
-                            for track_id in tracks)] = "str"
-            for f in xsorted(event['fields'].values(),
-                                key=EntitySorter.event_field):
-                if f['association'] == const.FieldAssociations.course:
-                    key = ",".join("{1}{0}.xfield_{2}".format(
-                        track_id, temp, f['field_name'])
-                                    for track_id in tracks)
-                    kind = const.FieldDatatypes(f['kind']).name
-                    spec[key] = kind
-        if sum(track['num_choices'] for track in tracks.values()) > 1:
-            spec[",".join(f"course_choices{track_id}.rank{i}"
-                            for track_id, track in tracks.items()
-                            for i in range(track['num_choices']))] = "int"
-    for f in xsorted(event['fields'].values(),
-                        key=EntitySorter.event_field):
-        if f['association'] == const.FieldAssociations.registration:
-            kind = const.FieldDatatypes(f['kind']).name
-            spec["reg_fields.xfield_{}".format(f['field_name'])] = kind
+    sorted_fields = _sort_event_fields(event['fields'])
+    field_choices = {
+        field['field_name']: dict(field['entries']) if field['entries'] else {}
+        for field in event['fields'].values()
+    }
+    course_choices = _get_course_choices(courses)
+    lodgement_choices = _get_lodgement_choices(lodgements)
+    lodgement_group_choices = _get_lodgement_group_choices(lodgement_groups)
+    spec: QuerySpec = {
+        "reg.id": QuerySpecEntry("id", n_("ID")),
+        "persona.id": QuerySpecEntry("id", n_("CdEDB-ID")),
+        "persona.given_names": QuerySpecEntry("str", n_("Given Names")),
+        "persona.family_name": QuerySpecEntry("str", n_("Family Name")),
+        "persona.username": QuerySpecEntry("str", n_("E-Mail")),
+        "persona.is_member": QuerySpecEntry("bool", n_("CdE-Member")),
+        "persona.display_name": QuerySpecEntry("str", n_("Known as (Forename)")),
+        "persona.title": QuerySpecEntry("str", n_("Title_[[of a persona]]")),
+        "persona.name_supplement": QuerySpecEntry("str", n_("Name Affix")),
+        # Choices for the gender will be manually set when displaying the result.
+        "persona.gender": QuerySpecEntry("int", n_("Gender"), choices=None),  # type: ignore[arg-type]
+        "persona.birthday": QuerySpecEntry("date", n_("Birthday")),
+        "persona.telephone": QuerySpecEntry("str", n_("Phone")),
+        "persona.mobile": QuerySpecEntry("str", n_("Mobile Phone")),
+        "persona.address": QuerySpecEntry("str", n_("Address")),
+        "persona.address_supplement": QuerySpecEntry("str", n_("Address Supplement")),
+        "persona.postal_code": QuerySpecEntry("str", n_("ZIP")),
+        "persona.location": QuerySpecEntry("str", n_("City")),
+        # Choices for the country will be manually set when displaying the result.
+        "persona.country": QuerySpecEntry("id", n_("Country"), choices=None),  # type: ignore[arg-type]
+        "reg.payment": QuerySpecEntry("date", n_("Payment")),
+        "reg.amount_paid": QuerySpecEntry("float", n_("Amount Paid")),
+        "reg.amount_owed": QuerySpecEntry("float", n_("Amount Owed")),
+        "reg.parental_agreement": QuerySpecEntry("bool", n_("Parental Consent")),
+        "reg.mixed_lodging": QuerySpecEntry("bool", n_("Mixed Lodging")),
+        "reg.list_consent": QuerySpecEntry("bool", n_("Participant List Consent")),
+        "reg.notes": QuerySpecEntry("str", n_("Notes")),
+        "reg.orga_notes": QuerySpecEntry("str", n_("Orga-Notes")),
+        "reg.checkin": QuerySpecEntry("datetime", n_("Checkin")),
+        "ctime.creation_time": QuerySpecEntry("datetime", n_("Registration Time")),
+        "mtime.modification_time":
+            QuerySpecEntry("datetime", n_("Last Modification Time")),
+    }
+
+    def get_part_spec(part: CdEDBObject) -> QuerySpec:
+        part_id = part['id']
+        prefix = "" if len(event['parts']) <= 1 else f"{part['shortname']}: "
+        return {
+            # Choices for the status will be manually set.
+            f"part{part_id}.status": QuerySpecEntry(
+                "int", n_("registration status"), prefix, choices=None),  # type: ignore[arg-type]
+            f"part{part_id}.is_camping_mat": QuerySpecEntry(
+                "bool", n_("camping mat user"), prefix),
+            f"part{part_id}.lodgement_id": QuerySpecEntry(
+                "id", n_("lodgement"), prefix, choices=lodgement_choices),
+            f"lodgement{part_id}.id": QuerySpecEntry("id", n_("lodgement ID"), prefix),
+            f"lodgement{part_id}.group_id": QuerySpecEntry(
+                "id", n_("lodgement group"), prefix, choices=lodgement_group_choices),
+            f"lodgement{part_id}.title": QuerySpecEntry(
+                "str", n_("lodgement title"), prefix),
+            f"lodgement{part_id}.notes": QuerySpecEntry(
+                "str", n_("lodgement notes"), prefix),
+            **{
+                f"lodgement{part_id}.xfield_{f['field_name']}": QuerySpecEntry(
+                    f['kind'].name, n_("lodgement {field}"), prefix,
+                    {'field': f['field_name']})
+                for f in sorted_fields[const.FieldAssociations.lodgement]
+            },
+            f"lodgement_group{part_id}.id": QuerySpecEntry(
+                "id", n_("lodgement group ID"), prefix),
+            f"lodgement_group{part_id}.title": QuerySpecEntry(
+                "str", n_("lodgement group title"), prefix),
+        }
+
+    def get_track_spec(track: CdEDBObject) -> QuerySpec:
+        track_id = track['id']
+        prefix = "" if len(event['tracks']) <= 1 else f"{track['shortname']}: "
+        return {
+            f"track{track_id}.is_course_instructor": QuerySpecEntry(
+                "bool", n_("instructs their course"), prefix),
+            f"track{track_id}.course_id": QuerySpecEntry(
+                "id", n_("course"), prefix, choices=course_choices),
+            f"track{track_id}.course_instructor": QuerySpecEntry(
+                "id", n_("instructed course"), prefix, choices=course_choices),
+            f"course{track_id}.id": QuerySpecEntry("id", n_("course ID"), prefix),
+            f"course{track_id}.nr": QuerySpecEntry("str", n_("course nr"), prefix),
+            f"course{track_id}.title": QuerySpecEntry(
+                "str", n_("course title"), prefix),
+            f"course{track_id}.shortname": QuerySpecEntry(
+                "str", n_("course shortname"), prefix),
+            f"course{track_id}.notes": QuerySpecEntry(
+                "str", n_("course notes"), prefix),
+            **{
+                f"course{track_id}.xfield_{f['field_name']}": QuerySpecEntry(
+                    f['kind'].name, n_("course {field}"), prefix,
+                    {'field': f['field_name']}, choices=field_choices[f['field_name']],
+                )
+                for f in sorted_fields[const.FieldAssociations.course]
+            },
+            f"course_instructor{track_id}.id": QuerySpecEntry(
+                "id", n_("instructed course ID"), prefix),
+            f"course_instructor{track_id}.nr": QuerySpecEntry(
+                "str", n_("instructed course nr"), prefix),
+            f"course_instructor{track_id}.title": QuerySpecEntry(
+                "str", n_("instructed course title"), prefix),
+            f"course_instructor{track_id}.shortname": QuerySpecEntry(
+                "str", n_("instructed course shortname"), prefix),
+            f"course_instructor{track_id}.notes": QuerySpecEntry(
+                "str", n_("instructed course notes"), prefix),
+            **{
+                f"course_instructor{track_id}.xfield_{f['field_name']}": QuerySpecEntry(
+                    f['kind'].name, n_("instructed course {field}"), prefix,
+                    {'field': f['field_name']}, choices=field_choices[f['field_name']],
+                )
+                for f in sorted_fields[const.FieldAssociations.course]
+            },
+        }
+
+    def get_course_choice_spec(track: CdEDBObject) -> QuerySpec:
+        track_id = track['id']
+        prefix = "" if len(event['tracks']) <= 1 else f"{track['shortname']}: "
+        return {
+            f"course_choices{track_id}.rank{i}": QuerySpecEntry(
+                "id", n_("{rank}. Choice"), prefix, {'rank': str(i + 1)},
+                choices=course_choices,
+            )
+            for i in range(track['num_choices'])
+        }
+
+    # Presort part specs, so we can iterate over them in order.
+    part_specs = {
+        part_id: get_part_spec(part)
+        for part_id, part in keydictsort_filter(event['parts'], EntitySorter.event_part)
+    }
+    track_specs = {
+        track_id: get_track_spec(track)
+        for track_id, track in keydictsort_filter(
+            event['tracks'], EntitySorter.course_track)
+    }
+    course_choice_specs = {
+        track_id: get_course_choice_spec(track)
+        for track_id, track in keydictsort_filter(
+            event['tracks'], EntitySorter.course_track)
+    }
+
+    # Add entries for individual parts and tracks in those parts.
+    for part_id, part_spec in part_specs.items():
+        part = event['parts'][part_id]
+        spec.update(part_spec)
+
+        # Add entries for individual tracks.
+        for track_id, track in keydictsort_filter(part['tracks'],
+                                                  EntitySorter.course_track):
+            spec.update(track_specs[track_id])
+
+            course_choice_spec = course_choice_specs[track_id]
+            # If there are course choices for the track, add an entry for any choice.
+            if key := ",".join(course_choice_spec.keys()):
+                # Don't overwrite a potential existing spec.
+                # This happens if there is exactly one choice.
+                if key not in course_choice_spec:
+                    prefix = "" if len(event['tracks']) <= 1 else track['shortname']
+                    spec[key] = QuerySpecEntry(
+                        "id", n_("Any Choice"), prefix, choices=course_choices)
+            spec.update(course_choice_spec)
+
+        # Add Entries for all tracks in this part.
+        spec.update(_combine_specs(
+            track_specs, part['tracks'], prefix=part['shortname']))
+        spec.update(_combine_specs(
+            course_choice_specs, part['tracks'], prefix=part['shortname']))
+
+    # Add entries for groups of parts and tracks in those parts.
+    sorted_part_groups = xsorted(
+        event['part_groups'].values(), key=EntitySorter.event_part_group)
+    sorted_part_groups.append({'part_ids': event['parts'].keys(), 'shortname': None})
+    for part_group in sorted_part_groups:
+        if constraint := part_group.get('constraint_type'):
+            if constraint != const.EventPartGroupType.Statistic:
+                continue
+        part_ids = part_group['part_ids']
+        prefix = part_group['shortname']
+        spec.update(_combine_specs(
+            part_specs, part_ids, prefix=prefix or n_("any part")))
+        # Add entries for track combinations.
+        track_ids = tuple(itertools.chain.from_iterable(
+            event['parts'][part_id]['tracks'].keys() for part_id in part_ids))
+        spec.update(_combine_specs(
+            track_specs, track_ids, prefix=prefix or n_("any track")))
+        spec.update(_combine_specs(
+            course_choice_specs, track_ids, prefix=prefix or n_("any track")))
+
+    spec.update({
+        f"reg_fields.xfield_{f['field_name']}": QuerySpecEntry(
+            f['kind'].name, f['field_name'], choices=field_choices[f['field_name']])
+        for f in sorted_fields[const.FieldAssociations.registration]
+    })
     return spec
 
 
-# TODO specify return type as OrderedDict.
-def make_registration_query_aux(
-    rs: RequestState, event: CdEDBObject, courses: CdEDBObjectMap,
-    lodgements: CdEDBObjectMap, lodgement_groups: CdEDBObjectMap,
-    fixed_gettext: bool = False
-) -> Tuple[Dict[str, Dict[int, str]], Dict[str, str]]:
-    """Un-inlined code to prepare input for template.
-    :param fixed_gettext: whether or not to use a fixed translation
-        function. True means static, False means localized.
-    :returns: Choices for select inputs and titles for columns.
-    """
-    tracks = event['tracks']
-
-    if fixed_gettext:
-        gettext = rs.default_gettext
-        enum_gettext = lambda x: x.name
-    else:
-        gettext = rs.gettext
-        enum_gettext = rs.gettext
-
-    course_identifier = lambda c: "{}. {}".format(c["nr"], c["shortname"])
-    course_choices = collections.OrderedDict(
-        (c_id, course_identifier(c))
-        for c_id, c in keydictsort_filter(courses, EntitySorter.course))
-    lodge_identifier = lambda l: l["title"]
-    lodgement_choices = collections.OrderedDict(
-        (l_id, lodge_identifier(l))
-        for l_id, l in keydictsort_filter(lodgements,
-                                          EntitySorter.lodgement))
-    lodgement_group_identifier = lambda g: g["title"]
-    lodgement_group_choices = collections.OrderedDict(
-        (g_id, lodgement_group_identifier(g))
-        for g_id, g in keydictsort_filter(lodgement_groups,
-                                          EntitySorter.lodgement_group))
-    # First we construct the choices
-    choices: Dict[str, Dict[Any, str]] = {
-        # Genders enum
-        'persona.gender': collections.OrderedDict(
-            enum_entries_filter(
-                const.Genders, enum_gettext, raw=fixed_gettext)),
-        'persona.country': collections.OrderedDict(get_localized_country_codes(rs)),
-    }
-
-    # Precompute some choices
-    reg_part_stati_choices = collections.OrderedDict(
-        enum_entries_filter(
-            const.RegistrationPartStati, enum_gettext, raw=fixed_gettext))
-    lodge_fields = {
-        field_id: field for field_id, field in event['fields'].items()
-        if field['association'] == const.FieldAssociations.lodgement
-        }
-    course_fields = {
-        field_id: field for field_id, field in event['fields'].items()
-        if field['association'] == const.FieldAssociations.course
-        }
-    reg_fields = {
-        field_id: field for field_id, field in event['fields'].items()
-        if field['association'] == const.FieldAssociations.registration
-        }
-
-    for part_id in event['parts']:
-        choices.update({
-            # RegistrationPartStati enum
-            "part{0}.status".format(part_id): reg_part_stati_choices,
-            # Lodgement choices for the JS selector
-            "part{0}.lodgement_id".format(part_id): lodgement_choices,
-            "lodgement{0}.group_id".format(part_id): lodgement_group_choices,
-        })
-        if not fixed_gettext:
-            # Lodgement fields value -> description
-            key = "lodgement{0}.xfield_{1}"
-            choices.update({
-                key.format(part_id, field['field_name']):
-                    collections.OrderedDict(field['entries'])
-                for field in lodge_fields.values() if field['entries']
-            })
-    for track_id, track in tracks.items():
-        choices.update({
-            # Course choices for the JS selector
-            "track{0}.course_id".format(track_id): course_choices,
-            "track{0}.course_instructor".format(track_id): course_choices,
-        })
-        for i in range(track['num_choices']):
-            choices[f"course_choices{track_id}.rank{i}"] = course_choices
-        if track['num_choices'] > 1:
-            choices[",".join(
-                f"course_choices{track_id}.rank{i}"
-                for i in range(track['num_choices']))] = course_choices
-        if not fixed_gettext:
-            # Course fields value -> description
-            for temp in ("course", "course_instructor"):
-                for field in course_fields.values():
-                    key = f"{temp}{track_id}.xfield_{field['field_name']}"
-                    if field['entries']:
-                        choices[key] = collections.OrderedDict(field['entries'])
-    if len(event['parts']) > 1:
-        choices.update({
-            # RegistrationPartStati enum
-            ",".join("part{0}.status".format(part_id)
-                     for part_id in event['parts']): reg_part_stati_choices,
-            ",".join("part{0}.lodgement_id".format(part_id)
-                     for part_id in event['parts']): lodgement_choices,
-            ",".join("lodgement{0}.group_id".format(part_id)
-                     for part_id in event['parts']): lodgement_group_choices,
-        })
-    if len(tracks) > 1:
-        choices[",".join(f"course_choices{track_id}.rank{i}"
-                for track_id, track in tracks.items()
-                for i in range(track['num_choices']))] = course_choices
-    if not fixed_gettext:
-        # Registration fields value -> description
-        choices.update({
-            "reg_fields.xfield_{}".format(field['field_name']):
-                collections.OrderedDict(field['entries'])
-            for field in reg_fields.values() if field['entries']
-        })
-
-    # Second we construct the titles
-    titles: Dict[str, str] = {
-        "reg_fields.xfield_{}".format(field['field_name']):
-            field['field_name']
-        for field in reg_fields.values()
-    }
-    for track_id, track in tracks.items():
-        if len(tracks) > 1:
-            prefix = "{shortname}: ".format(shortname=track['shortname'])
-        else:
-            prefix = ""
-        titles.update({
-            "track{0}.is_course_instructor".format(track_id):
-                prefix + gettext("instructs their course"),
-            "track{0}.course_id".format(track_id):
-                prefix + gettext("course"),
-            "track{0}.course_instructor".format(track_id):
-                prefix + gettext("instructed course"),
-            "course{0}.id".format(track_id):
-                prefix + gettext("course ID"),
-            "course{0}.nr".format(track_id):
-                prefix + gettext("course nr"),
-            "course{0}.title".format(track_id):
-                prefix + gettext("course title"),
-            "course{0}.shortname".format(track_id):
-                prefix + gettext("course shortname"),
-            "course{0}.notes".format(track_id):
-                prefix + gettext("course notes"),
-            "course_instructor{0}.id".format(track_id):
-                prefix + gettext("instructed course ID"),
-            "course_instructor{0}.nr".format(track_id):
-                prefix + gettext("instructed course nr"),
-            "course_instructor{0}.title".format(track_id):
-                prefix + gettext("instructed course title"),
-            "course_instructor{0}.shortname".format(track_id):
-                prefix + gettext("instructed course shortname"),
-            "course_instructor{0}.notes".format(track_id):
-                prefix + gettext("instructed courese notes"),
-        })
-        key = "course{0}.xfield_{1}"
-        titles.update({
-            key.format(track_id, field['field_name']):
-                prefix + gettext("course {field}").format(
-                    field=field['field_name'])
-            for field in course_fields.values()
-        })
-        key = "course_instructor{0}.xfield_{1}"
-        titles.update({
-            key.format(track_id, field['field_name']):
-                prefix + gettext("instructed course {field}").format(
-                    field=field['field_name'])
-            for field in course_fields.values()
-        })
-        for i in range(track['num_choices']):
-            titles[f"course_choices{track_id}.rank{i}"] = \
-                prefix + gettext("%s. Choice") % (i + 1)
-        if track['num_choices'] > 1:
-            titles[",".join(f"course_choices{track_id}.rank{i}"
-                            for i in range(track['num_choices']))] = \
-                prefix + gettext("Any Choice")
-    if len(event['tracks']) > 1:
-        titles.update({
-            ",".join("track{0}.is_course_instructor".format(track_id)
-                     for track_id in tracks):
-                gettext("any track: instructs their course"),
-            ",".join("track{0}.course_id".format(track_id)
-                     for track_id in tracks):
-                gettext("any track: course"),
-            ",".join("track{0}.course_instructor".format(track_id)
-                     for track_id in tracks):
-                gettext("any track: instructed course"),
-            ",".join("course{0}.id".format(track_id)
-                     for track_id in tracks):
-                gettext("any track: course ID"),
-            ",".join("course{0}.nr".format(track_id)
-                     for track_id in tracks):
-                gettext("any track: course nr"),
-            ",".join("course{0}.title".format(track_id)
-                     for track_id in tracks):
-                gettext("any track: course title"),
-            ",".join("course{0}.shortname".format(track_id)
-                     for track_id in tracks):
-                gettext("any track: course shortname"),
-            ",".join("course{0}.notes".format(track_id)
-                     for track_id in tracks):
-                gettext("any track: course notes"),
-            ",".join("course_instructor{0}.id".
-                     format(track_id) for track_id in tracks):
-                gettext("any track: instructed course ID"),
-            ",".join("course_instructor{0}.nr".
-                     format(track_id) for track_id in tracks):
-                gettext("any track: instructed course nr"),
-            ",".join("course_instructor{0}.title".
-                     format(track_id) for track_id in tracks):
-                gettext("any track: instructed course title"),
-            ",".join("course_instructor{0}.shortname".
-                     format(track_id) for track_id in tracks):
-                gettext("any track: instructed course shortname"),
-            ",".join("course_instructor{0}.notes".format(track_id)
-                     for track_id in tracks):
-                gettext("any track: instructed course notes"),
-        })
-        key = "course{0}.xfield_{1}"
-        titles.update({
-            ",".join(key.format(track_id, field['field_name'])
-                     for track_id in tracks):
-                gettext("any track: course {field}").format(
-                    field=field['field_name'])
-            for field in course_fields.values()
-        })
-        key = "course_instructor{0}.xfield_{1}"
-        titles.update({
-            ",".join(key.format(track_id, field['field_name'])
-                     for track_id in tracks):
-                gettext("any track: instructed course {field}").format(
-                    field=field['field_name'])
-            for field in course_fields.values()
-        })
-        key = ",".join(f"course_choices{track_id}.rank{i}"
-                       for track_id, track in tracks.items()
-                       for i in range(track['num_choices']))
-        titles[key] = gettext("any track: Any Choice")
-    for part_id, part in event['parts'].items():
-        if len(event['parts']) > 1:
-            prefix = "{shortname}: ".format(shortname=part['shortname'])
-        else:
-            prefix = ""
-        titles.update({
-            "part{0}.status".format(part_id):
-                prefix + gettext("registration status"),
-            "part{0}.is_camping_mat".format(part_id):
-                prefix + gettext("camping mat user"),
-            "part{0}.lodgement_id".format(part_id):
-                prefix + gettext("lodgement"),
-            "lodgement{0}.id".format(part_id):
-                prefix + gettext("lodgement ID"),
-            "lodgement{0}.group_id".format(part_id):
-                prefix + gettext("lodgement group"),
-            "lodgement{0}.title".format(part_id):
-                prefix + gettext("lodgement title"),
-            "lodgement{0}.notes".format(part_id):
-                prefix + gettext("lodgement notes"),
-            "lodgement_group{0}.id".format(part_id):
-                prefix + gettext("lodgement group ID"),
-            "lodgement_group{0}.title".format(part_id):
-                prefix + gettext("lodgement group title"),
-        })
-        key = "lodgement{0}.xfield_{1}"
-        titles.update({
-            key.format(part_id, field['field_name']):
-                prefix + gettext("lodgement {field}").format(
-                    field=field['field_name'])
-            for field in lodge_fields.values()
-        })
-    if len(event['parts']) > 1:
-        titles.update({
-            ",".join("part{0}.status".format(part_id)
-                     for part_id in event['parts']):
-                gettext("any part: registration status"),
-            ",".join("part{0}.is_camping_mat".format(part_id)
-                     for part_id in event['parts']):
-                gettext("any part: camping mat user"),
-            ",".join("part{0}.lodgement_id".format(part_id)
-                     for part_id in event['parts']):
-                gettext("any part: lodgement"),
-            ",".join("lodgement{0}.id".format(part_id)
-                     for part_id in event['parts']):
-                gettext("any part: lodgement ID"),
-            ",".join("lodgement{0}.group_id".format(part_id)
-                     for part_id in event['parts']):
-                gettext("any part: lodgement group"),
-            ",".join("lodgement{0}.title".format(part_id)
-                     for part_id in event['parts']):
-                gettext("any part: lodgement title"),
-            ",".join("lodgement{0}.notes".format(part_id)
-                     for part_id in event['parts']):
-                gettext("any part: lodgement notes"),
-            ",".join("lodgement_group{0}.id".format(part_id)
-                     for part_id in event['parts']):
-                gettext("any part: lodgement group ID"),
-            ",".join("lodgement_group{0}.title".format(part_id)
-                     for part_id in event['parts']):
-                gettext("any part: lodgement group title"),
-        })
-        key = "lodgement{0}.xfield_{1}"
-        titles.update({
-            ",".join(key.format(part_id, field['field_name'])
-                     for part_id in event['parts']):
-                gettext("any part: lodgement {field}").format(
-                    field=field['field_name'])
-            for field in lodge_fields.values()
-        })
-    return choices, titles
-
-
-def make_course_query_spec(event: CdEDBObject) -> Dict[str, str]:
+def make_course_query_spec(event: CdEDBObject, courses: CdEDBObjectMap = None,
+                           lodgements: CdEDBObjectMap = None,
+                           lodgement_groups: CdEDBObjectMap = None) -> QuerySpec:
     """Helper to generate ``QueryScope.event_course``'s spec.
 
-    Since each event has custom course fields we have to amend the query
-    spec on the fly.
+    Since each event has custom course fields and an arbitrary number
+    of course tracks we have to extend this spec on the fly.
     """
-    tracks = event['tracks']
-    course_fields = {
-        field_id: field for field_id, field in event['fields'].items()
-        if field['association'] == const.FieldAssociations.course
+    sorted_tracks = keydictsort_filter(event['tracks'], EntitySorter.course_track)
+    sorted_course_fields = _sort_event_fields(event['fields'])[
+        const.FieldAssociations.course]
+    field_choices = {
+        field['field_name']: dict(field['entries']) if field['entries'] else {}
+        for field in sorted_course_fields
     }
 
-    # This is an OrderedDict, so order should be respected.
-    spec = collections.OrderedDict([
-                    ("course.id", "id"),
-                    ("course.course_id", "id"),
-                    ("course.nr", "str"),
-                    ("course.title", "str"),
-                    ("course.description", "str"),
-                    ("course.shortname", "str"),
-                    ("course.instructors", "str"),
-                    ("course.min_size", "int"),
-                    ("course.max_size", "int"),
-                    ("course.notes", "str"),
-                    # This will be augmented with additional fields in the fly.
-                ])
-    spec.update({
-        "course_fields.xfield_{0}".format(field['field_name']):
-            const.FieldDatatypes(field['kind']).name
-        for field in course_fields.values()
-    })
+    course_choices = _get_course_choices(courses)
 
-    for track_id, track in tracks.items():
-        spec["track{0}.is_offered".format(track_id)] = "bool"
-        spec["track{0}.takes_place".format(track_id)] = "bool"
-        spec["track{0}.attendees".format(track_id)] = "int"
-        spec["track{0}.instructors".format(track_id)] = "int"
-        for rank in range(track['num_choices']):
-            spec["track{0}.num_choices{1}".format(track_id, rank)] = "int"
-
-    return spec
-
-
-# TODO specify return type as OrderedDict.
-def make_course_query_aux(rs: RequestState, event: CdEDBObject,
-                          courses: CdEDBObjectMap,
-                          fixed_gettext: bool = False
-                          ) -> Tuple[Dict[str, Dict[int, str]],
-                                     Dict[str, str]]:
-    """Un-inlined code to prepare input for template.
-
-    :param fixed_gettext: whether or not to use a fixed translation
-        function. True means static, False means localized.
-    :returns: Choices for select inputs and titles for columns.
-    """
-
-    tracks = event['tracks']
-    gettext = rs.default_gettext if fixed_gettext else rs.gettext
-
-    # Construct choices.
-    course_identifier = lambda c: "{}. {}".format(c["nr"], c["shortname"])
-    course_choices = collections.OrderedDict(
-        xsorted((c["id"], course_identifier(c)) for c in courses.values()))
-    choices: Dict[str, Dict[int, str]] = {
-        "course.course_id": course_choices
+    spec = {
+        "course.id": QuerySpecEntry("id", n_("course id")),
+        "course.course_id": QuerySpecEntry("id", n_("course"), choices=course_choices),
+        "course.nr": QuerySpecEntry("str", n_("course nr")),
+        "course.title": QuerySpecEntry("str", n_("course title")),
+        "course.description": QuerySpecEntry("str", n_("course description")),
+        "course.shortname": QuerySpecEntry("str", n_("course shortname")),
+        "course.instructors": QuerySpecEntry("str", n_("course instructors")),
+        "course.min_size": QuerySpecEntry("int", n_("course min size")),
+        "course.max_size": QuerySpecEntry("int", n_("course max size")),
+        "course.notes": QuerySpecEntry("str", n_("course notes")),
+        # This will be augmented with additional fields in the fly.
     }
-    course_fields = {
-        field_id: field for field_id, field in event['fields'].items()
-        if field['association'] == const.FieldAssociations.course
+
+    def get_track_spec(track: CdEDBObject) -> QuerySpec:
+        track_id = track['id']
+        prefix = "" if len(event['tracks']) <= 1 else f"{track['shortname']}: "
+        return {
+            f"track{track_id}.is_offered": QuerySpecEntry(
+                "bool", n_("is offered"), prefix),
+            f"track{track_id}.takes_place": QuerySpecEntry(
+                "bool", n_("takes place"), prefix),
+            f"track{track_id}.attendees": QuerySpecEntry(
+                "int", n_("attendees"), prefix),
+            f"track{track_id}.instructors": QuerySpecEntry(
+                "int", n_("instructors"), prefix),
         }
-    if not fixed_gettext:
-        # Course fields value -> description
-        choices.update({
-            "course_fields.xfield_{0}".format(field['field_name']):
-                collections.OrderedDict(field['entries'])
-            for field in course_fields.values() if field['entries']
-        })
 
-    # Construct titles.
-    titles: Dict[str, str] = {
-        "course.id": gettext("course id"),
-        "course.course_id": gettext("course"),
-        "course.nr": gettext("course nr"),
-        "course.title": gettext("course title"),
-        "course.description": gettext("course description"),
-        "course.shortname": gettext("course shortname"),
-        "course.instructors": gettext("course instructors"),
-        "course.min_size": gettext("course min size"),
-        "course.max_size": gettext("course max size"),
-        "course.notes": gettext("course notes"),
+    def get_course_choice_spec(track: CdEDBObject) -> QuerySpec:
+        track_id = track['id']
+        prefix = "" if len(event['tracks']) <= 1 else f"{track['shortname']}: "
+        return {
+            f"track{track_id}.num_choices{i}": QuerySpecEntry(
+                "int", n_("{rank}. choices"), prefix, {'rank': str(i + 1)})
+            for i in range(track['num_choices'])
+        }
+
+    track_specs = {
+        track_id: get_track_spec(track)
+        for track_id, track in sorted_tracks
     }
-    titles.update({
-        "course_fields.xfield_{}".format(field['field_name']):
-            field['field_name']
-        for field in course_fields.values()
-    })
-    for track_id, track in tracks.items():
-        if len(tracks) > 1:
-            prefix = "{shortname}: ".format(shortname=track['shortname'])
-        else:
-            prefix = ""
-        titles.update({
-            "track{0}.takes_place".format(track_id):
-                prefix + gettext("takes place"),
-            "track{0}.is_offered".format(track_id):
-                prefix + gettext("is offered"),
-            "track{0}.attendees".format(track_id):
-                prefix + gettext("attendees"),
-            "track{0}.instructors".format(track_id):
-                prefix + gettext("instructors"),
-        })
-        for rank in range(track['num_choices']):
-            titles.update({
-                "track{0}.num_choices{1}".format(track_id, rank):
-                    prefix + gettext("{}. choices").format(
-                        rank+1),
-            })
-
-    return choices, titles
-
-
-def make_lodgement_query_spec(event: CdEDBObject) -> Dict[str, str]:
-    parts = event["parts"]
-    lodgement_fields = {
-        field_id: field for field_id, field in event['fields'].items()
-        if field['association'] == const.FieldAssociations.lodgement
+    course_choice_specs = {
+        track_id: get_course_choice_spec(track)
+        for track_id, track in sorted_tracks
     }
+    # Add entries for individual tracks.
+    for track_id, track_spec in track_specs.items():
+        spec.update(track_spec)
 
-    # This is an OrderedDcit, so order should be respected.
-    spec = collections.OrderedDict([
-                    ("lodgement.id", "id"),
-                    ("lodgement.lodgement_id", "id"),
-                    ("lodgement.title", "str"),
-                    ("lodgement.regular_capacity", "int"),
-                    ("lodgement.camping_mat_capacity", "int"),
-                    ("lodgement.notes", "str"),
-                    ("lodgement.group_id", "int"),
-                    ("lodgement_group.title", "int"),
-                    # This will be augmented with additional fields in the fly.
-                ])
+        course_choice_spec = course_choice_specs[track_id]
+        # If there are course choices for the track, add an entry for any choice.
+        if key := ",".join(course_choice_spec.keys()):
+            # Don't overwrite a potential existing spec.
+            # This happens if there is exactly one choice.
+            if key not in course_choice_spec:
+                prefix = ("" if len(event['tracks']) <= 1
+                          else event['tracks'][track_id]['shortname'])
+                spec[key] = QuerySpecEntry("id", n_("Any Choice"), prefix)
+        spec.update(course_choice_spec)
+
+    # Add entries for groups of tracks.
+    sorted_parts = xsorted(event['parts'].values(), key=EntitySorter.event_part)
+    sorted_part_groups = xsorted(
+        event['part_groups'].values(), key=EntitySorter.event_part_group)
+    track_groups = (
+        {'track_ids': event['tracks'].keys(), 'shortname': n_("any track")},
+        *(
+            {'track_ids': part['tracks'].keys(), 'shortname': part['shortname']}
+            for part in sorted_parts
+        ),
+        *(
+            {
+                'track_ids': tuple(itertools.chain.from_iterable(
+                    event['parts'][part_id]['tracks'].keys()
+                    for part_id in part_group['part_ids'])),
+                'shortname': part_group['shortname'],
+            }
+            for part_group in sorted_part_groups
+        )
+    )
+    for track_group in track_groups:
+        track_ids = track_group['track_ids']
+        prefix = track_group['shortname'] or n_("any track")
+        spec.update(_combine_specs(track_specs, track_ids, prefix))
+        spec.update(_combine_specs(course_choice_specs, track_ids, prefix))
+
     spec.update({
-        f"lodgement_fields.xfield_{field['field_name']}":
-            const.FieldDatatypes(field['kind']).name
-        for field in lodgement_fields.values()
+        f"course_fields.xfield_{field['field_name']}": QuerySpecEntry(
+            field['kind'].name, field['field_name'],
+            choices=field_choices[field['field_name']])
+        for field in sorted_course_fields
     })
-
-    for part_id, part in parts.items():
-        spec[f"part{part_id}.regular_inhabitants"] = "int"
-        spec[f"part{part_id}.camping_mat_inhabitants"] = "int"
-        spec[f"part{part_id}.total_inhabitants"] = "int"
-        spec[f"part{part_id}.group_regular_inhabitants"] = "int"
-        spec[f"part{part_id}.group_camping_mat_inhabitants"] = "int"
-        spec[f"part{part_id}.group_total_inhabitants"] = "int"
 
     return spec
 
 
-def make_lodgement_query_aux(rs: RequestState, event: CdEDBObject,
-                             lodgements: CdEDBObjectMap,
-                             lodgement_groups: CdEDBObjectMap,
-                             fixed_gettext: bool = False
-                             ) -> Tuple[Dict[str, Dict[int, str]],
-                                        Dict[str, str]]:
-    """Un-inlined code to prepare input for template.
+def make_lodgement_query_spec(event: CdEDBObject, courses: CdEDBObjectMap = None,
+                              lodgements: CdEDBObjectMap = None,
+                              lodgement_groups: CdEDBObjectMap = None) -> QuerySpec:
+    """Helper to generate ``QueryScope.lodgement``'s spec.
 
-    :param fixed_gettext: whether or not to use a fixed translation
-        function. True means static, False means localized.
-    :returns: Choices for select inputs and titles for columns.
+    Since each event has custom lodgement fields and an arbitrary number
+    of event parts, we have to expand this spec on the fly.
     """
-
-    parts = event['parts']
-    gettext = rs.default_gettext if fixed_gettext else rs.gettext
-
-    # Construct choices.
-    lodgement_choices = collections.OrderedDict(
-        (l_id, l['title'])
-        for l_id, l in keydictsort_filter(lodgements,
-                                          EntitySorter.lodgement))
-    lodgement_group_choices = collections.OrderedDict({-1: gettext(n_("--no group--"))})
-    lodgement_group_choices.update(
-        [(lg_id, lg['title']) for lg_id, lg in keydictsort_filter(
-            lodgement_groups, EntitySorter.lodgement_group)])
-    choices: Dict[str, Dict[int, str]] = {
-        "lodgement.lodgement_id": lodgement_choices,
-        "lodgement_group.id": lodgement_group_choices,
+    sorted_parts = keydictsort_filter(event['parts'], EntitySorter.event_part)
+    sorted_lodgement_fields = _sort_event_fields(event['fields'])[
+        const.FieldAssociations.lodgement]
+    field_choices = {
+        field['field_name']: dict(field['entries']) if field['entries'] else {}
+        for field in sorted_lodgement_fields
     }
-    lodgement_fields = {
-        field_id: field for field_id, field in event['fields'].items()
-        if field['association'] == const.FieldAssociations.lodgement
-    }
-    if not fixed_gettext:
-        # Lodgement fields value -> description
-        choices.update({
-            f"lodgement_fields.xfield_{field['field_name']}":
-                collections.OrderedDict(field['entries'])
-            for field in lodgement_fields.values() if field['entries']
-        })
+    lodgement_choices = _get_lodgement_choices(lodgements)
+    lodgement_group_choices = _get_lodgement_group_choices(lodgement_groups)
 
-    # Construct titles.
-    titles: Dict[str, str] = {
-        "lodgement.id": gettext(n_("Lodgement ID")),
-        "lodgement.lodgement_id": gettext(n_("Lodgement")),
-        "lodgement.title": gettext(n_("Title_[[name of an entity]]")),
-        "lodgement.regular_capacity": gettext(n_("Regular Capacity")),
-        "lodgement.camping_mat_capacity":
-            gettext(n_("Camping Mat Capacity")),
-        "lodgement.notes": gettext(n_("Lodgement Notes")),
-        "lodgement.group_id": gettext(n_("Lodgement Group ID")),
-        "lodgement_group.tmp_id": gettext(n_("Lodgement Group")),
-        "lodgement_group.title": gettext(n_("Lodgement Group Title")),
-        "lodgement_group.regular_capacity":
-            gettext(n_("Lodgement Group Regular Capacity")),
-        "lodgement_group.camping_mat_capacity":
-            gettext(n_("Lodgement Group Camping Mat Capacity")),
+    spec = {
+        "lodgement.id": QuerySpecEntry("id", n_("lodgement ID")),
+        "lodgement.lodgement_id": QuerySpecEntry(
+            "id", n_("lodgement"), choices=lodgement_choices),
+        "lodgement.title": QuerySpecEntry("str", n_("Title_[[name of an entity]]")),
+        "lodgement.regular_capacity": QuerySpecEntry("int", n_("Regular Capacity")),
+        "lodgement.camping_mat_capacity": QuerySpecEntry(
+            "int", n_("Camping Mat Capacity")),
+        "lodgement.notes": QuerySpecEntry("str", n_("Lodgement Notes")),
+        "lodgement.group_id": QuerySpecEntry(
+            "int", n_("Lodgement Group"), choices=lodgement_group_choices),
+        "lodgement_group.id": QuerySpecEntry("int", n_("Lodgement Group ID")),
+        "lodgement_group.title": QuerySpecEntry("int", n_("Lodgement Group Title")),
+        # This will be augmented with additional fields in the fly.
     }
-    titles.update({
-        f"lodgement_fields.xfield_{field['field_name']}":
-            field['field_name']
-        for field in lodgement_fields.values()
+
+    def get_part_spec(part: CdEDBObject) -> QuerySpec:
+        part_id = part['id']
+        prefix = "" if len(event['parts']) <= 1 else f"{part['shortname']}: "
+        return {
+            f"part{part_id}.regular_inhabitants": QuerySpecEntry(
+                "int", n_("Regular Inhabitants"), prefix),
+            f"part{part_id}.camping_mat_inhabitants": QuerySpecEntry(
+                "int", n_("Camping Mat Inhabitants"), prefix),
+            f"part{part_id}.total_inhabitants": QuerySpecEntry(
+                "int", n_("Total Inhabitants"), prefix),
+            f"part{part_id}.group_regular_inhabitants": QuerySpecEntry(
+                "int", n_("Group Regular Inhabitants"), prefix),
+            f"part{part_id}.group_camping_mat_inhabitants": QuerySpecEntry(
+                "int", n_("Group Camping Mat Inhabitants"), prefix),
+            f"part{part_id}.group_total_inhabitants": QuerySpecEntry(
+                "int", n_("Group Total Inhabitants"), prefix),
+        }
+
+    # Presort part specs so we can iterate over them in order.
+    part_specs = {
+        part_id: get_part_spec(part)
+        for part_id, part in sorted_parts
+    }
+
+    # Add entries for individual parts.
+    for part_id, part_spec in part_specs.items():
+        spec.update(part_spec)
+
+    # Add entries for groups of parts.
+    sorted_part_groups = xsorted(
+        event['part_groups'].values(), key=EntitySorter.event_part_group)
+    sorted_part_groups.append({'part_ids': event['parts'].keys(), 'shortname': None})
+    for part_group in sorted_part_groups:
+        part_ids = part_group['part_ids']
+        prefix = part_group['shortname'] or n_("any part")
+        spec.update(_combine_specs(part_specs, part_ids, prefix=prefix))
+
+    spec.update({
+        f"lodgement_fields.xfield_{f['field_name']}": QuerySpecEntry(
+            f['kind'].name, f['field_name'], choices=field_choices[f['field_name']])
+        for f in sorted_lodgement_fields
     })
-    for part_id, part in parts.items():
-        if len(parts) > 1:
-            prefix = f"{part['shortname']}: "
-        else:
-            prefix = ""
-        titles.update({
-            f"part{part_id}.regular_inhabitants":
-                prefix + gettext(n_("Regular Inhabitants")),
-            f"part{part_id}.camping_mat_inhabitants":
-                prefix + gettext(n_("Reserve Inhabitants")),
-            f"part{part_id}.total_inhabitants":
-                prefix + gettext(n_("Total Inhabitants")),
-            f"part{part_id}.group_regular_inhabitants":
-                prefix + gettext(n_("Group Regular Inhabitants")),
-            f"part{part_id}.group_camping_mat_inhabitants":
-                prefix + gettext(n_("Group Reserve Inhabitants")),
-            f"part{part_id}.group_total_inhabitants":
-                prefix + gettext(n_("Group Total Inhabitants")),
-        })
 
-    return choices, titles
+    return spec
