@@ -2,11 +2,13 @@
 
 import datetime
 import decimal
-from typing import Any, Dict, List, NewType as _NewType, Type
+from typing import Any, Dict, List, Mapping, NewType as _NewType, Type
 
-from cdedb.common import CdEDBObject
-from cdedb.subman import SubscriptionState
+from cdedb.common import CdEDBObject, CdEDBOptionalMap
 from cdedb.query import Query
+from cdedb.subman import SubscriptionState
+
+TypeMapping = Mapping[str, Type[Any]]
 
 # SIMPLE/PRIMITIVE/ATOMIC TYPES
 
@@ -27,10 +29,13 @@ EmptyList = _NewType("EmptyList", List[Any])
 
 Realm = _NewType("Realm", str)
 StringType = _NewType("StringType", str)
+Shortname = _NewType("Shortname", str)
+ShortnameIdentifier = _NewType("ShortnameIdentifier", Shortname)
+ShortnameRestrictiveIdentifier = _NewType(
+    "ShortnameRestrictiveIdentifier", ShortnameIdentifier)
+LegacyShortname = _NewType("LegacyShortname", str)
 PrintableASCIIType = _NewType("PrintableASCIIType", str)
 PrintableASCII = _NewType("PrintableASCII", str)  # TODO make these subtypes?
-Alphanumeric = _NewType("Alphanumeric", str)
-CSVAlphanumeric = _NewType("CSVAlphanumeric", str)
 Identifier = _NewType("Identifier", str)
 RestrictiveIdentifier = _NewType("RestrictiveIdentifier", str)
 CSVIdentifier = _NewType("CSVIdentifier", str)
@@ -83,6 +88,8 @@ Institution = _NewType("Institution", CdEDBObject)
 PastEvent = _NewType("PastEvent", CdEDBObject)
 Event = _NewType("Event", CdEDBObject)
 EventPart = _NewType("EventPart", CdEDBObject)
+EventPartGroup = _NewType("EventPartGroup", CdEDBObject)
+EventPartGroupSetter = _NewType("EventPartGroupSetter", CdEDBOptionalMap)
 EventTrack = _NewType("EventTrack", CdEDBObject)
 EventField = _NewType("EventField", CdEDBObject)
 EventFeeModifier = _NewType("EventFeeModifier", CdEDBObject)
@@ -95,13 +102,18 @@ EventAssociatedFields = _NewType("EventAssociatedFields", CdEDBObject)
 FeeBookingEntry = _NewType("FeeBookingEntry", CdEDBObject)
 LodgementGroup = _NewType("LodgementGroup", CdEDBObject)
 Lodgement = _NewType("Lodgement", CdEDBObject)
-Questionnaire = _NewType("Questionnaire", Dict[int, Any])  # TODO maybe cast keys to str
+QuestionnaireRow = _NewType("QuestionnaireRow", CdEDBObject)
+# TODO maybe cast keys to str
+Questionnaire = _NewType("Questionnaire", Dict[int, List[QuestionnaireRow]])
 
 SerializedEvent = _NewType("SerializedEvent", CdEDBObject)
 SerializedEventUpload = _NewType("SerializedEventUpload", SerializedEvent)
 SerializedPartialEvent = _NewType("SerializedPartialEvent", CdEDBObject)
 SerializedPartialEventUpload = _NewType(
     "SerializedPartialEventUpload", SerializedPartialEvent)
+SerializedEventQuestionnaire = _NewType("SerializedEventQuestionnaire", CdEDBObject)
+SerializedEventQuestionnaireUpload = _NewType(
+    "SerializedEventQuestionnaireUpload", SerializedEventQuestionnaire)
 
 PartialCourse = _NewType("PartialCourse", CdEDBObject)
 PartialLodgementGroup = _NewType("PartialLodgementGroup", CdEDBObject)
