@@ -8,8 +8,8 @@ import pytz
 
 import cdedb.database.constants as const
 from cdedb.common import (
-    PERSONA_CDE_FIELDS, PERSONA_CORE_FIELDS, PERSONA_EVENT_FIELDS, QuotaException,
-    CdEDBLog
+    PERSONA_CDE_FIELDS, PERSONA_CORE_FIELDS, PERSONA_EVENT_FIELDS, CdEDBLog,
+    QuotaException,
 )
 from cdedb.query import Query, QueryOperators, QueryScope
 from tests.common import USER_DICT, BackendTest, as_users, nearly_now
@@ -121,7 +121,8 @@ class TestCdEBackend(BackendTest):
                 ("country,country2", QueryOperators.empty, None)],
             order=(("family_name,birth_name", True),),)
         result = self.cde.submit_general_query(self.key, query)
-        self.assertEqual({6, 9, 15, 100}, {e['id'] for e in result})
+        expectation = {6, 9, 15, 100}
+        self.assertEqual({e['id'] for e in result}, expectation)
 
     @as_users("vera")
     def test_user_search(self) -> None:
@@ -134,7 +135,7 @@ class TestCdEBackend(BackendTest):
                 ("birthday", QueryOperators.less, datetime.datetime.now())],
             order=(("family_name", True),),)
         result = self.cde.submit_general_query(self.key, query)
-        self.assertEqual({2, 3, 4, 6, 7, 13, 15, 16, 22, 23, 27, 32, 100},
+        self.assertEqual({2, 3, 4, 6, 7, 13, 15, 16, 22, 23, 27, 32, 37, 100},
                          {e['id'] for e in result})
 
     @as_users("vera")
