@@ -6,7 +6,7 @@ import datetime
 import pytz
 
 import cdedb.database.constants as const
-from cdedb.common import xsorted, nearly_now
+from cdedb.common import nearly_now, xsorted
 from tests.common import BackendTest, as_users
 
 
