@@ -9,16 +9,16 @@ from typing import List
 import freezegun
 import webtest
 
-import cdedb.database.constants as const
 from cdedb.common import (
-    CdEDBObject, ADMIN_VIEWS_COOKIE_NAME, ASSEMBLY_BAR_SHORTNAME, now, NearlyNow
+    ADMIN_VIEWS_COOKIE_NAME, ANTI_CSRF_TOKEN_NAME, ASSEMBLY_BAR_SHORTNAME, CdEDBObject,
+    NearlyNow, now,
 )
 from cdedb.frontend.common import datetime_filter
 from cdedb.query import QueryOperators
 from cdedb.validation import parse_datetime
 from tests.common import (
-    FrontendTest, MultiAppFrontendTest, UserIdentifier, USER_DICT, as_users,
-    get_user, storage,
+    USER_DICT, FrontendTest, MultiAppFrontendTest, UserIdentifier, as_users, get_user,
+    storage,
 )
 
 
@@ -68,7 +68,7 @@ class AssemblyTestHelpers(FrontendTest):
                 'signup_end': "2222-4-1 00:00:00",
                 'description': "Wir werden alle Häretiker exkommunizieren.",
                 'notes': "Nur ein Aprilscherz",
-                'presider_ids': "DB-23-X",
+                'presider_ids': USER_DICT['werner']['DB-ID'],
             }
         else:
             adata = adata.copy()
@@ -77,7 +77,7 @@ class AssemblyTestHelpers(FrontendTest):
         self.traverse({'description': 'Versammlungen'},
                       {'description': 'Versammlung anlegen'})
         self.assertTitle("Versammlung anlegen")
-        f = self.response.forms['createassemblyform']
+        f = self.response.forms['configureassemblyform']
         for k, v in adata.items():
             if isinstance(f[k], webtest.forms.Checkbox):
                 f[k].checked = bool(v)
@@ -118,7 +118,7 @@ class AssemblyTestHelpers(FrontendTest):
         """
         self.traverse({"description": "Abstimmungen"},
                       {"description": "Abstimmung anlegen"})
-        f = self.response.forms["createballotform"]
+        f = self.response.forms["configureballotform"]
         for k, v in bdata.items():
             f[k] = v
         self.submit(f)
@@ -269,14 +269,11 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertNoLink('assembly/assembly/1/change')
 
         # Test Presider Controls Admin View
-        self.traverse({'href': '/assembly/assembly/1/show'})
-        self.assertNoLink('/assembly/assembly/1/attachment/add')
         self.traverse({'href': '/assembly/assembly/1/ballot/list'})
         self.assertNoLink('/assembly/assembly/1/ballot/2/change')
         self.assertNoLink('/assembly/assembly/1/ballot/create')
         self.traverse({'href': '/assembly/assembly/1/ballot/2/show'})
         self.assertNoLink('/assembly/assembly/1/ballot/2/change')
-        self.assertNoLink('/assembly/assembly/1/ballot/2/attachment/add')
         self.assertNotIn('removecandidateform6', self.response.forms)
         self.assertNotIn('addcandidateform', self.response.forms)
         self.assertNotIn('deleteballotform', self.response.forms)
@@ -284,21 +281,34 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertNotIn('addattendeeform', self.response.forms)
         self.assertNonPresence("TeX-Liste")
 
+        # check attachments in Archiv-Versammlung
+        self.traverse("Versammlungs-Übersicht", "Archiv-Sammlung", "Dateien")
+        self.assertNotIn('deleteattachmentform1', self.response.forms)
+        self.assertNotIn('removeattachmentversionform2_3', self.response.forms)
+        self.assertNoLink('/assembly/assembly/3/attachment/add')
+        self.assertNoLink('/assembly/assembly/3/attachment/1/add')
+
         self._click_admin_view_button(re.compile(r"Versammlungs-Administration"),
                                       current_state=True)
         self._click_admin_view_button(re.compile(r"Versammlungsleitung-Schaltflächen"),
                                       current_state=False)
+
+        self.traverse("Versammlungs-Übersicht", "Archiv-Sammlung", "Dateien")
+        self.assertIn('deleteattachmentform1', self.response.forms)
+        self.assertIn('removeattachmentversionform2_3', self.response.forms)
+        self.traverse({'href': '/assembly/assembly/3/attachment/add'},
+                      {'href': '/assembly/assembly/3/attachments'},
+                      {'href': '/assembly/assembly/3/attachment/1/add'})
+
+        # go back to Internationaler Kongress
+        self.traverse("Versammlungs-Übersicht", "Internationaler Kongress")
         self.traverse({'href': '/assembly/assembly/1/show'},
-                      {'href': '/assembly/assembly/1/attachment/add'},
-                      {'href': '/assembly/assembly/1/show'},
                       {'href': '/assembly/assembly/1/log'},
                       {'href': '/assembly/assembly/1/ballot/list'},
                       {'href': '/assembly/assembly/1/ballot/2/change'},
                       {'href': '/assembly/assembly/1/ballot/list'},
                       {'href': '/assembly/assembly/1/ballot/create'},
                       {'href': '/assembly/assembly/1/ballot/list'},
-                      {'href': '/assembly/assembly/1/ballot/2/show'},
-                      {'href': '/assembly/assembly/1/ballot/2/attachment/add'},
                       {'href': '/assembly/assembly/1/ballot/2/show'})
         self.assertIn('candidatessummaryform', self.response.forms)
         self.assertIn('deleteballotform', self.response.forms)
@@ -316,18 +326,15 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
             self.assertNonPresence("Internationaler Kongress")
 
             # now, sign them up
-            self.logout()
-            self.login('werner')
-            self.traverse({'description': 'Versammlungen'},
-                          {'description': 'Internationaler Kongress'})
-            self._external_signup(user)
-            self.logout()
-            self.login(user)
+            with self.switch_user('werner'):
+                self.traverse({'description': 'Versammlungen'},
+                              {'description': 'Internationaler Kongress'})
+                self._external_signup(user)
             self.traverse({'description': 'Versammlungen'})
 
         self.traverse({'description': 'Internationaler Kongress'})
         attendee = {"Versammlungs-Übersicht", "Übersicht", "Teilnehmer",
-                    "Abstimmungen", "Zusammenfassung", "Datei-Übersicht"}
+                    "Abstimmungen", "Zusammenfassung", "Dateien"}
         admin = {"Konfiguration", "Log"}
 
         # not assembly admins
@@ -349,7 +356,7 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
                       {'description': 'Internationaler Kongress'},)
         self.assertTitle("Internationaler Kongress")
         self.traverse({'description': 'Konfiguration'},)
-        f = self.response.forms['changeassemblyform']
+        f = self.response.forms['configureassemblyform']
         f['title'] = 'Drittes CdE-Konzil'
         f['description'] = "Wir werden alle Häretiker exkommunizieren."
         f['presider_address'] = "drittes konzil@example.cde"
@@ -361,7 +368,7 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertTitle("Drittes CdE-Konzil")
         self.assertPresence("Häretiker", div='description')
         self.traverse({'description': 'Konfiguration'},)
-        f = self.response.forms['changeassemblyform']
+        f = self.response.forms['configureassemblyform']
         self.assertEqual(f['presider_address'].value, 'konzil@example.cde')
 
     @as_users("werner")
@@ -369,7 +376,7 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.traverse({'description': 'Versammlungen'},
                       {'description': 'Archiv-Sammlung'},
                       {'description': 'Konfiguration'}, )
-        f = self.response.forms['changeassemblyform']
+        f = self.response.forms['configureassemblyform']
         f['signup_end'] = '2000-02-22T01:00:00'
         self.submit(f)
         self.assertPresence("22.02.2000, 01:00:00")
@@ -377,38 +384,113 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertNonPresence("22.02.2000, 01:00:00")
         self.assertPresence("(Anmeldung nicht mehr möglich)")
 
-    # Use ferdinand since viktor is not a member and may not signup.
-    @as_users("ferdinand")
+    @storage
+    @as_users("viktor")
     def test_create_delete_assembly(self) -> None:
         presider_address = "presider@lists.cde-ev.de"
-        self._create_assembly(delta={'create_presider_list': True,
-                                     'presider_address': presider_address})
-        self.assertPresence("Häretiker", div='description')
-        self.assertPresence("Aprilscherz", div='notes')
-        self.assertPresence("Versammlungsleitungs-Mailingliste angelegt.",
-                            div="notifications")
-        self.assertPresence("Versammlungsleitungs-E-Mail-Adresse durch Adresse der"
-                            " neuen Mailingliste ersetzt.", div="notifications")
+        self._create_assembly(delta={'presider_address': presider_address,
+                                     'create_presider_list': True,
+                                     'create_attendee_list': True})
+        self.assertNotification("Versammlungsleitungs-Mailingliste angelegt.")
+        self.assertNotification("Versammlungsteilnehmer-Mailingliste angelegt.")
+        self.assertNotification("Angegebene Versammlungsleitungs-E-Mail-Adresse durch"
+                                " Adresse der neuen Mailingliste ersetzt.",
+                                ntype='info')
+
+        # create another assembly to test input validation
+        self.traverse("Versammlungen", "Versammlung anlegen")
+        f = self.response.forms['configureassemblyform']
+        f['title'] = "Drittes CdE-Konzil reloaded"
+        f['description'] = "Weil's so schön war."
+        f['notes'] = "Nicht schon wieder..."
+        f['shortname'] = "konzil3"  # same as before, to test ml address conflict
+        f['create_attendee_list'].checked = True
+        f['create_presider_list'].checked = True
+        self.submit(f, check_notification=False)
+        self.assertValidationError('signup_end',
+                                   "Muss ein valides Datum mit Uhrzeit sein.")
+        f['signup_end'] = "2222-9-1 00:00"
+        self.submit(f, check_notification=False)
+        self.assertValidationError(
+            'presider_ids',
+            "Darf nicht leer sein, damit eine Mailingliste erstellt werden kann")
+        user_archived = "Einige dieser Nutzer existieren nicht oder sind archiviert."
+        user_no_realm = "Einige dieser Nutzer sind keine Versammlungsnutzer."
+        f['presider_ids'] = USER_DICT['hades']['DB-ID']  # archived
+        self.submit(f, check_notification=False)
+        self.assertValidationError('presider_ids', user_archived)
+        self.assertNonPresence(user_no_realm)
+        f['presider_ids'] = USER_DICT['emilia']['DB-ID']  # event user
+        self.submit(f, check_notification=False)
+        self.assertValidationError('presider_ids', user_no_realm)
+        self.assertNonPresence(user_archived)
+        f['presider_ids'] = USER_DICT['werner']['DB-ID']
+        self.submit(f)
+        self.assertNotification(
+            "Mailingliste konzil3@lists.cde-ev.de existiert bereits.", 'info')
+        self.assertNotification(
+            "Mailingliste konzil3-leitung@lists.cde-ev.de existiert bereits.", 'info')
         self.assertNotIn('createpresiderlistform', self.response.forms)
-        f = self.response.forms['createattendeelistform']
+        self.assertNotIn('createattendeelistform', self.response.forms)
+        # log contains links
+        self.traverse("Versammlungs-Übersicht", "Log", "Drittes CdE-Konzil$")
+        self.traverse("Versammlungs-Übersicht", "Log", "Drittes CdE-Konzil reloaded$")
+
+        # delete one assembly
+        f = self.response.forms['deleteassemblyform']
+        self.submit(f, check_notification=False)
+        # TODO: there is no validation error near the checkbox
+        self.assertNotification("Validierung fehlgeschlagen.", 'error')
+        f['ack_delete'].checked = True
         self.submit(f)
-        self.assertPresence("Versammlungsteilnehmer-Mailingliste angelegt.",
-                            div="notifications")
-        self.traverse({'description': "Abstimmungen"})
+
+        # test deletion of other created assembly wtih some potential blockers
+        with open(self.testfile_dir / "form.pdf", 'rb') as datafile:
+            attachment = datafile.read()
+        bdata = {
+            'title': 'Müssen wir wirklich regeln...',
+            'vote_begin': "2222-12-12 00:00:00",
+            'vote_end': "2223-5-1 00:00:00",
+            'abs_quorum': "0",
+            'rel_quorum': "0",
+            'votes': "",
+        }
+        # Assemblies with ballots which started voting can not be deleted.
+        # Other ballots are ok
+        self.traverse("Drittes CdE-Konzil", "Abstimmungen")
         self.assertPresence("Es wurden noch keine Abstimmungen angelegt.")
-        self.traverse({'description': r"\sÜbersicht"})
-        # Make sure assemblies with attendees can be deleted
-        f = self.response.forms['signupform']
+        self._create_ballot(bdata, candidates=None)
+        self.assertTitle("Müssen wir wirklich regeln... (Drittes CdE-Konzil)")
+        # Make sure assemblies with attachments can be deleted
+        self.traverse("Dateien", "Datei hinzufügen")
+        f = self.response.forms['addattachmentform']
+        f['title'] = "Vorläufige Beschlussvorlage"
+        f['attachment'] = webtest.Upload("form", attachment, "application/octet-stream")
+        f['filename'] = "beschluss.pdf"
         self.submit(f)
+        self.assertPresence("Vorläufige Beschlussvorlage")
+        # Make sure assemblies with attendees can be deleted
+        with self.switch_user("ferdinand"):
+            self.traverse("Versammlungen", "Drittes CdE-Konzil")
+            self.submit(self.response.forms['signupform'])
+            self.assertNotification("Angemeldet.")
+        # now delete
+        self.traverse({'href': "/show", 'description': "Übersicht"})
         f = self.response.forms['deleteassemblyform']
         f['ack_delete'].checked = True
         self.submit(f)
         self.assertTitle("Versammlungen")
         self.assertNonPresence("Drittes CdE-Konzil")
+        # log does not contain links anymore
+        self.traverse("Log")
+        self.assertPresence("Drittes CdE-Konzil")
+        self.assertNoLink(content="Drittes CdE-Konzil")
 
-        self._create_assembly(delta={'presider_address': presider_address})
+        # no preiders are needed if no ml was requested
+        self._create_assembly(delta={'presider_address': presider_address,
+                                     'presider_ids': ""})
         self.traverse("Konfiguration")
-        f = self.response.forms['changeassemblyform']
+        f = self.response.forms['configureassemblyform']
         self.assertEqual(f['presider_address'].value, presider_address)
 
     @as_users("viktor")
@@ -416,19 +498,25 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.traverse("Versammlungen", "Archiv-Sammlung")
         self.assertTitle("Archiv-Sammlung")
 
-        self.assertPresence("Datei hinzufügen", div='attachmentspanel')
         self.submit(
             self.response.forms[f"removepresiderform{ USER_DICT['werner']['id'] }"])
         f = self.response.forms['createpresiderlistform']
-        self.assertIn('disabled', f.fields['submitform'][0].attrs)
+        self.assertInputHasAttr(f['submitform'], 'disabled')
+        f[ANTI_CSRF_TOKEN_NAME] = "evil"
         self.submit(f, check_notification=False)
-        self.assertPresence(
-            "Mailingliste kann nur mit Versammlungsleitern erstellt werden.",
-            div='notifications')
+        self.assertNotification("Der Anti-CSRF-Token wurde gefälscht.", 'error')
+        self.submit(self.response.forms['createpresiderlistform'],
+                    check_notification=False)
+        self.assertNotification(
+            "Mailingliste kann nur mit Versammlungsleitern erstellt werden.", 'error')
         f = self.response.forms['addpresidersform']
         f['presider_ids'] = USER_DICT['werner']['DB-ID']
         self.submit(f)
-        self.submit(self.response.forms['createattendeelistform'])
+        f = self.response.forms['createattendeelistform']
+        self.submit(f)
+        self.submit(f, check_notification=False)
+        self.assertNotification(
+            "Mailingliste archiv@lists.cde-ev.de existiert bereits.", 'info')
         self.submit(self.response.forms['createpresiderlistform'])
 
     @as_users("werner")
@@ -436,7 +524,6 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.traverse("Versammlungen", "Archiv-Sammlung")
         self.assertTitle("Archiv-Sammlung")
 
-        self.assertPresence("Datei hinzufügen", div='attachmentspanel')
         self.assertNotIn('addpresidersform', self.response.forms)
         self.assertNotIn('createattendeelistform', self.response.forms)
 
@@ -455,9 +542,21 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
                       {'description': 'Internationaler Kongress'},)
         self.assertTitle("Internationaler Kongress")
         f = self.response.forms['signupform']
+        f[ANTI_CSRF_TOKEN_NAME] = "evil"
+        self.submit(f, check_notification=False)
+        self.assertNotification("Der Anti-CSRF-Token wurde gefälscht.", 'error')
+        f = self.response.forms['signupform']
         self.submit(f)
         self.assertTitle("Internationaler Kongress")
         self.assertNotIn('signupform', self.response.forms)
+        self.submit(f, check_notification=False)
+        self.assertNotification("Bereits angemeldet.")
+        # assembly with running ballots
+        self.traverse("Versammlungs-Übersicht", "Kanonische Beispielversammlung")
+        self.assertNotIn('signupform', self.response.forms)
+        self.post('/assembly/assembly/2/signup', {})
+        self.assertTitle("Kanonische Beispielversammlung")
+        self.assertNotification("Der Anmeldezeitraum ist vorbei.", 'warning')
 
     @as_users("kalif")
     def test_no_signup(self) -> None:
@@ -474,31 +573,44 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertNonPresence("Kalif", div='attendees-list')
         # Valid request
         f = self.response.forms['addattendeeform']
-        f['persona_id'] = "DB-11-6"
+        f['persona_id'] = USER_DICT['kalif']['DB-ID']
         self.submit(f)
         self.assertTitle('Anwesenheitsliste (Archiv-Sammlung)')
         self.assertPresence("Kalif", div='attendees-list')
         # Archived user
-        f = self.response.forms['addattendeeform']
-        f['persona_id'] = "DB-8-6"
+        f['persona_id'] = USER_DICT['hades']['DB-ID']
         self.submit(f, check_notification=False)
         self.assertValidationError(
             "persona_id", "Dieser Benutzer existiert nicht oder ist archiviert.")
         # Member
-        f = self.response.forms['addattendeeform']
-        f['persona_id'] = "DB-2-7"
+        f['persona_id'] = USER_DICT['berta']['DB-ID']
         self.submit(f, check_notification=False)
         self.assertValidationError(
             "persona_id", "Mitglieder müssen sich selbst anmelden.")
         # Event user
-        f = self.response.forms['addattendeeform']
-        f['persona_id'] = "DB-5-1"
+        f['persona_id'] = USER_DICT['emilia']['DB-ID']
         self.submit(f, check_notification=False)
         self.assertValidationError(
             "persona_id", "Dieser Nutzer ist kein Versammlungsnutzer.")
-        # TODO: add a check for a non-existant user and an invalid DB-ID.
+        # Nonexistent user
+        f['persona_id'] = "DB-1000-6"
+        self.submit(f, check_notification=False)
+        self.assertValidationError(
+            'persona_id', "Dieser Benutzer existiert nicht oder ist archiviert.")
+        # Invalid DB-ID
+        f['persona_id'] = "DB-1000-X"
+        self.submit(f, check_notification=False)
+        self.assertValidationError('persona_id', "Checksumme stimmt nicht.")
+        # signup impossible after signup end
+        if self.user_in("ferdinand"):
+            self.get('/assembly/assembly/2/attendees')
+            self.assertNotIn('addattendeeform', self.response.forms)
+            self.post('/assembly/assembly/2/signup/external',
+                      {'persona_id': USER_DICT['kalif']['DB-ID']})
+            self.assertTitle("Anwesenheitsliste (Kanonische Beispielversammlung)")
+            self.assertNotification("Der Anmeldezeitraum ist vorbei.", 'warning')
 
-    @as_users("werner", "viktor", "kalif")
+    @as_users("werner", "kalif")
     def test_list_attendees(self) -> None:
         self.traverse({'description': 'Versammlungen'},
                       {'description': 'Internationaler Kongress'},
@@ -512,11 +624,17 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertNonPresence("Charly")
         if self.user_in('kalif'):
             self.assertNonPresence("Download")
-        elif self.user_in('viktor', 'werner'):
+        else:
             self.assertPresence("Download")
             self.traverse("TeX-Liste")
             for attendee in attendees:
                 self.assertIn(attendee, self.response.text)
+            self.get('/assembly/assembly/3/attendees')
+            self.assertTitle("Anwesenheitsliste (Archiv-Sammlung)")
+            self.assertPresence("Insgesamt 0 Anwesende", div='attendees-count')
+            self.traverse("TeX-Liste")
+            self.assertTitle("Anwesenheitsliste (Archiv-Sammlung)")
+            self.assertNotification("Leere Datei.")
 
     @storage
     @as_users("rowena")
@@ -537,42 +655,117 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
             self._create_assembly()
             self._signup()
             self.traverse({'description': 'Konfiguration'})
-            f = self.response.forms['changeassemblyform']
+            f = self.response.forms['configureassemblyform']
             f['signup_end'] = "2002-4-1 00:00:00"
             self.submit(f)
 
-            bdata = {
-                'title': 'Maximale Länge der Satzung',
-                'description': "Dann muss man halt eine alte Regel rauswerfen,"
-                               " wenn man eine neue will.",
-                'vote_begin': base_time + delta,
-                'vote_end': base_time + 3*delta,
-                'abs_quorum': "0",
-                'rel_quorum': "0",
-                'votes': "",
-                'notes': "Kein Aprilscherz!",
-            }
-            self._create_ballot(bdata, candidates=None)
-            self.assertTitle("Maximale Länge der Satzung (Drittes CdE-Konzil)")
+            for ballot_nr in (1, 2):
+                bdata = {
+                    'title': f'Maximale Länge der {ballot_nr}. Satzung',
+                    'description': "Dann muss man halt eine alte Regel rauswerfen,"
+                                   " wenn man eine neue will.",
+                    'vote_begin': base_time + delta,
+                    'vote_end': base_time + delta + 2*ballot_nr*delta,
+                    'abs_quorum': "0",
+                    'rel_quorum': "0",
+                    'votes': "",
+                    'notes': "Kein Aprilscherz!",
+                }
+                self._create_ballot(bdata, candidates=None)
+                self.assertTitle(f"{bdata['title']} (Drittes CdE-Konzil)")
 
-            frozen_time.tick(delta=4*delta)
-            self.traverse({'description': 'Abstimmungen'},
-                          {'description': 'Maximale Länge der Satzung'},
-                          {'description': 'Drittes CdE-Konzil'},)
+            frozen_time.tick(delta=2 * delta)
+            self.traverse({'href': "/show", 'description': "Übersicht"})
+            self.assertInputHasAttr(
+                self.response.forms['concludeassemblyform']['submitform'], 'disabled')
+            self.post('/assembly/assembly/1001/conclude', {'ack_conclude': True})
+            self.assertNotification(
+                "Kann Versammlung mit offenen Abstimmungen nicht abschließen.", 'error')
+
+            # upload an attachment to test after conclusion
+            with open(self.testfile_dir / "rechen.pdf", 'rb') as datafile:
+                attachment = datafile.read()
+            self.traverse("Dateien", "Datei hinzufügen")
+            f = self.response.forms['addattachmentform']
+            f['title'] = "Satzung mit weniger Paragrafen"
+            f['attachment'] = webtest.Upload("kurz.pdf", attachment,
+                                             "application/octet-stream")
+            self.submit(f)
+
+            # regression test for #2310
+            self.traverse("Abstimmungen", "Maximale Länge der 1. Satzung")
+            frozen_time.tick(delta=2 * delta)
+            # First ballot is concluded now, second still running. Ensure viewing
+            # the second concludes the first and navigation works
+            self.traverse("Nächste")
+            frozen_time.tick(delta=2 * delta)
+            self.traverse("Nächste", "Drittes CdE-Konzil")
+
+            # now the actual conclusion test
             self.assertTitle("Drittes CdE-Konzil")
             f = self.response.forms['concludeassemblyform']
+            self.submit(f, check_notification=False)
+            # TODO: there is no validation error near the checkbox
+            self.assertNotification("Validierung fehlgeschlagen.", 'error')
             f['ack_conclude'].checked = True
             self.submit(f)
+            self.assertNotIn('concludeassemblyform', self.response.forms)
+            self.submit(f, check_notification=False)
+            msg = "Die Versammlung ist bereits beendet."
+            self.assertNotification(msg, 'info')
             self.assertTitle("Drittes CdE-Konzil")
             # Presiders can no longer be changed
+            link = '/assembly/assembly/1001/presider/'
             self.assertNotIn("addpresidersform", self.response.forms)
-            self.assertNotIn("removepresiderform1", self.response.forms)
+            self.post(link + 'add', {'presider_ids': USER_DICT['werner']['DB-ID']})
+            self.assertTitle("Drittes CdE-Konzil")
+            self.assertNotification(msg, 'warning')
+            self.assertNotIn(f"removepresiderform{USER_DICT['werner']['id']}",
+                             self.response.forms)
+            self.post(link + 'remove', {'presider_id': USER_DICT['werner']['id']})
+            self.assertTitle("Drittes CdE-Konzil")
+            self.assertNotification(msg, 'warning')
+            # deletion is not possible
+            self.assertNotIn('deleteassemblyform', self.response.forms)
+            self.post('/assembly/assembly/1001/delete', {'ack_delete': True})
+            self.assertNotification(
+                "Versammlungen mit aktiven Abstimmungen können nicht gelöscht werden.",
+                'error')
+            # modification not possible anymore
+            self.traverse("Abstimmungen")  # first iteration starts at this page
+            for link in ('/assembly/assembly/1001/ballot/create',
+                         '/assembly/assembly/1001/change'):
+                self.assertNoLink(link)
+                self.get(link)
+                self.assertTitle("Drittes CdE-Konzil")
+                self.assertNotification(msg, 'warning')
+                self.post(link, {'title': "Jetzt ist Schluss!"})
+                self.assertTitle("Drittes CdE-Konzil")
+                self.assertNotification(msg, 'warning')
+            # attachments can no longer be changed
+            self.traverse("Dateien")
+            self.assertNotIn('addattachmentform', self.response.forms)
+            self.assertNotIn('addattachmentversionform', self.response.forms)
+            base_link = '/assembly/assembly/1001/attachment/'
+            for suffix, atype in (('add', "Datei"), ('1001/add', "Dateiversion")):
+                url = base_link + suffix
+                msg = f"Es kann keine {atype} mehr hinzugefügt werden, nachdem" \
+                      " die Versammlung gesperrt wurde."
+                self.get(url)
+                self.assertNotification(msg, 'error')
+                self.post(url, {'title': "Datei mit Verspätung"})
+                self.assertNotification(msg, 'error')
+            base_link += '1001/'
+            for suffix, atype in (('delete', "Datei"),
+                                  ('version/1/delete', "Dateiversion")):
+                url = base_link + suffix
+                self.post(url, {'attachment_ack_delete': True})
+                self.assertNotification(f"{atype} kann nicht gelöscht werden.", 'error')
 
     @storage
     @as_users("anton")
     def test_preferential_vote_result(self) -> None:
         self.get('/assembly/assembly/1/ballot/1/show')
-        self.follow()  # Redirect because ballot has not been tallied yet.
         self.assertTitle("Antwort auf die letzte aller Fragen "
                          "(Internationaler Kongress)")
         self.assertPresence("Nach dem Leben, dem Universum und dem ganzen Rest")
@@ -588,7 +781,6 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         f = self.response.forms['signupform']
         self.submit(f)
         self.get('/assembly/assembly/1/ballot/1/show')
-        self.follow()  # Redirect because ballot has not been tallied yet.
         self.assertTitle("Antwort auf die letzte aller Fragen "
                          "(Internationaler Kongress)")
         self.assertPresence("Nach dem Leben, dem Universum und dem ganzen Rest")
@@ -610,8 +802,6 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
             with self.subTest(ballot=ballot_id):
                 assembly = assemblies[ballot['assembly_id']]
                 self.get(f"/assembly/assembly/{assembly['id']}/ballot/{ballot_id}/show")
-                # redirect in case of tallying, extending etc
-                self.follow()
                 self.assertTitle(f"{ballot['title']} ({assembly['title']})")
 
                 # Check display of regular voting period.
@@ -657,7 +847,6 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
     @as_users("garcia")
     def test_show_ballot_without_attendance(self) -> None:
         self.get('/assembly/assembly/1/ballot/1/show')
-        self.follow()  # Redirect because ballot has not been tallied yet.
         self.assertTitle("Antwort auf die letzte aller Fragen "
                          "(Internationaler Kongress)")
         self.assertPresence("Nach dem Leben, dem Universum und dem ganzen Rest")
@@ -668,101 +857,283 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
     @storage
     @as_users("werner")
     def test_entity_ballot_simple(self) -> None:
-        self.traverse({'description': 'Versammlungen$'},
-                      {'description': 'Internationaler Kongress'},
-                      {'description': 'Abstimmungen'},)
-        self.assertTitle("Abstimmungen (Internationaler Kongress)")
-        self.assertNonPresence("Maximale Länge der Satzung")
-        self.assertNonPresence("Es wurden noch keine Abstimmungen angelegt")
-        bdata = {
-            'title': 'Maximale Länge der Satzung',
-            'description': "Dann muss man halt eine alte Regel rauswerfen,"
-                           " wenn man eine neue will.",
-            'vote_begin': "2222-4-1 00:00:00",
-            'vote_end': "2222-5-1 00:00:00",
-            'votes': "",
-            'notes': "Kein Aprilscherz!",
-        }
-        self._create_ballot(bdata, atitle="Internationaler Kongress")
-        self.traverse({'description': 'Bearbeiten'},)
-        f = self.response.forms['changeballotform']
-        self.assertEqual("Kein Aprilscherz!", f['notes'].value)
-        f['notes'] = "April, April!"
+        """This test covers all around creation, modification and deletion.
+        Everything related to voting and tallying is covered in own tests."""
+        # create ballot
+        self.traverse("Versammlungen", "Internationaler Kongress", "Abstimmungen",
+                      "Abstimmung anlegen")
+        self.assertTitle("Abstimmung anlegen (Internationaler Kongress)")
+        f = self.response.forms['configureballotform']
+        f['title'] = "Maximale Länge der Satzung"
+        f['description'] = ("Dann muss man halt eine alte Regel rauswerfen,"
+                            " wenn man eine neue will.")
         f['vote_begin'] = "2222-4-1 00:00:00"
-        f['vote_end'] = "2222-4-1 00:00:01"
-        self.submit(f)
-        # votes must be empty or a positive int
-        self.traverse({'description': 'Bearbeiten'}, )
-        f = self.response.forms['changeballotform']
-        f['votes'] = 0
+        f['vote_end'] = "2222-3-1 00:00:00"
+        f['notes'] = "_Kein_ Aprilscherz!"
         self.submit(f, check_notification=False)
-        self.assertValidationError('votes', message="Muss positiv sein.")
-        f['votes'] = 1
+        self.assertValidationError(
+            'vote_end',  "Darf nicht vor Abstimmungsbeginn liegen.")
+        f['vote_end'] = "2222-5-1 00:00:00"
         self.submit(f)
-        self.assertTitle("Maximale Länge der Satzung (Internationaler Kongress)")
-        self.traverse({'description': 'Bearbeiten'},)
-        f = self.response.forms['changeballotform']
-        self.assertEqual("April, April!", f['notes'].value)
-        self.traverse({'description': 'Abstimmungen'},)
-        self.assertTitle("Abstimmungen (Internationaler Kongress)")
-        self.assertPresence("Maximale Länge der Satzung")
-        self.traverse({'description': 'Maximale Länge der Satzung'},)
-        self.assertTitle("Maximale Länge der Satzung (Internationaler Kongress)")
-        f = self.response.forms['deleteballotform']
-        f['ack_delete'].checked = True
-        self.submit(f)
-        self.traverse({'description': 'Abstimmungen'},)
-        self.assertTitle("Abstimmungen (Internationaler Kongress)")
-        self.assertNonPresence("Maximale Länge der Satzung")
+
+        ballot_states = {
+            1: "Abstimmung ist beendet",
+            5: "Abstimmung läuft",
+            1001: "Abstimmung hat noch nicht begonnen",
+        }
+        comment_failed = "Kommentare sind nur für beendete Abstimmungen erlaubt."
+        change_failed = "Eine aktive Abstimmung kann nicht mehr verändert werden."
+        for ballot_id in ballot_states:
+            with self.subTest(ballot_states[ballot_id]):
+                ballot_title = (
+                    "Maximale Länge der Satzung"
+                    if ballot_id == 1001 else self.get_sample_datum(
+                        'assembly.ballots', ballot_id)['title'])
+                page_title = f"{ballot_title} (Internationaler Kongress)"
+                self.get(f'/assembly/assembly/1/ballot/{ballot_id}/show')
+                self.assertTitle(page_title)
+                self.assertPresence(ballot_states[ballot_id], div='status')
+
+                # commenting - only possible for past ballots
+                comment_url = f'/assembly/assembly/1/ballot/{ballot_id}/comment'
+                if ballot_id == 1:
+                    self.traverse({'href': comment_url, 'description': "Kommentieren"})
+                    self.assertTitle(f"{ballot_title} bearbeiten"
+                                     f" (Internationaler Kongress)")
+                    f = self.response.forms['commentballotform']
+                    comment = "War nur ein *Experiment*."
+                    f['comment'] = comment
+                    self.submit(f)
+                    self.assertTitle(page_title)
+                    self.assertPresence("Abstimmungskommentar")
+                    self.assertPresence("War nur ein Experiment.", div='comment')
+                    self.traverse("Kommentieren")
+                    f = self.response.forms['commentballotform']
+                    # check that the form is filled with the current comment
+                    self.assertEqual(comment, f['comment'].value)
+                    evil_comment = "Evil Eve war hier!"
+                    f['comment'] = evil_comment
+                    f[ANTI_CSRF_TOKEN_NAME] = "evil"
+                    self.submit(f, check_notification=False)
+                    self.assertTitle(f"{ballot_title} bearbeiten"
+                                     " (Internationaler Kongress)")
+                    self.assertNotification("Der Anti-CSRF-Token wurde gefälscht.",
+                                            'error')
+                    f = self.response.forms['commentballotform']
+                    self.assertEqual(evil_comment, f['comment'].value)
+                    f['comment'] = ""
+                    self.submit(f)
+                    self.assertTitle(page_title)
+                    self.assertNonPresence("Abstimmungskommentar")
+                else:
+                    self.assertNoLink(comment_url)
+                    self.get(comment_url)
+                    self.assertTitle(page_title)
+                    self.assertNotification(comment_failed, 'error')
+                    self.post(comment_url, {'comment': "Testkommentar"})
+                    self.assertTitle(page_title)
+                    self.assertNotification(comment_failed, 'error')
+
+                # modification and deletion - only possible for future ballots
+                self.assertTitle(page_title)
+                change_url = f'/assembly/assembly/1/ballot/{ballot_id}/change'
+                if ballot_id == 1001:
+                    # modification
+                    self.traverse({'href': change_url, 'description': "Bearbeiten"})
+                    f = self.response.forms['configureballotform']
+                    self.assertEqual("_Kein_ Aprilscherz!", f['notes'].value)
+                    f['notes'] = "April, April!"
+                    f['vote_begin'] = "2222-4-1 00:00:00"
+                    f['vote_end'] = "2222-4-1 00:00:01"
+                    # votes must be empty or a positive int
+                    f['votes'] = 0
+                    self.submit(f, check_notification=False)
+                    self.assertValidationError('votes', message="Muss positiv sein.")
+                    f['votes'] = 1
+                    self.submit(f)
+                    self.assertTitle(
+                        "Maximale Länge der Satzung (Internationaler Kongress)")
+                    self.assertPresence("April, April!", div='notes')
+                    # deletion
+                    f = self.response.forms['deleteballotform']
+                    self.submit(f, check_notification=False)
+                    # TODO: there is no validation error near the checkbox
+                    self.assertNotification("Validierung fehlgeschlagen.", 'error')
+                    f['ack_delete'].checked = True
+                    self.submit(f)
+                    self.assertTitle("Abstimmungen (Internationaler Kongress)")
+                    self.assertNonPresence("Maximale Länge der Satzung")
+                else:
+                    # modification
+                    self.assertNoLink(change_url)
+                    self.get(change_url)
+                    self.assertTitle(page_title)
+                    self.assertNotification(change_failed, 'warning')
+                    self.post(change_url, {'notes': "Störenfried war hier!"})
+                    self.assertTitle(page_title)
+                    self.assertNotification(change_failed, 'warning')
+                    # deletion
+                    self.assertNonPresence("Löschen")
+                    self.assertNotIn("deleteballotform", self.response.forms)
+                    self.post(f'/assembly/assembly/1/ballot/{ballot_id}/delete',
+                              {'ack_delete': True})
+                    self.assertNotification(
+                        "Eine aktive Abstimmung kann nicht gelöscht werden.", 'error')
+
+    @storage
+    @as_users("charly", "viktor")
+    def test_attachment_redirects(self) -> None:
+        # Test that accessing the latest version and the legacy urls redirect to the
+        # correct page.
+        # attachment_ids = set.union(*
+        attachment_ids = set.union(*[
+            self.assembly.list_attachments(self.key, assembly_id=assembly_id)
+            for assembly_id in self.assembly.list_assemblies(self.key)
+        ])
+
+        for attachment_id in attachment_ids:
+            self._test_one_attachment_redirect(attachment_id)
+
+    def _test_one_attachment_redirect(self, attachment_id: int) -> None:
+        attachment = self.assembly.get_attachment(self.key, attachment_id)
+        assembly_id = attachment['assembly_id']
+        ballot_ids = attachment['ballot_ids']
+        latest_version_nr = attachment['latest_version_nr']
+
+        # Use get via the app, to avoid following the redirects.
+        # Check that legacy urls with a version redirect to that version.
+        version_target = (f"/assembly/assembly/{assembly_id}"
+                          f"/attachment/{attachment_id}/version/{latest_version_nr}")
+
+        urls = (
+            # Shortcut that always redirects to the current version.
+            f"/assembly/assembly/{assembly_id}/attachment/{attachment_id}/latest",
+            # Legacy url with additional "/get".
+            f"/assembly/assembly/{assembly_id}"
+            f"/attachment/{attachment_id}/version/{latest_version_nr}/get",
+        ) + tuple(
+            # Legacy url with the ballot the attachment is linked to.
+            # This redirect should work with arbitrary ballot_ids in theory.
+            f"/assembly/assembly/{assembly_id}/ballot/{ballot_id}"
+            f"/attachment/{attachment_id}/version/{latest_version_nr}/get"
+            for ballot_id in ballot_ids
+        )
+        for url in urls:
+            with self.subTest(url=url):
+                self.assertRedirect(url, target_url=version_target)
+
+        # Check that urls without a version redirect to the latest version.
+        non_version_target = urls[0]
+
+        urls = (
+            # simple url for convenience
+            f"/assembly/assembly/{assembly_id}/attachment/{attachment_id}/",
+            # Legacy url that used to retrieve the latest version.
+            f"/assembly/assembly/{assembly_id}/attachment/{attachment_id}/get",
+        ) + tuple(
+            # Legacy url for retrieving the latest version of a ballot attachment.
+            f"/assembly/assembly/{assembly_id}/ballot/{ballot_id}"
+            f"/attachment/{attachment_id}/get"
+            for ballot_id in ballot_ids
+        )
+        for url in urls:
+            with self.subTest(url=url):
+                self.assertRedirect(url, target_url=non_version_target)
 
     @storage
     @as_users("werner")
-    def test_delete_ballot(self) -> None:
-        self.get("/assembly/assembly/1/ballot/2/show")
-        self.assertTitle("Farbe des Logos (Internationaler Kongress)")
-        self.assertPresence("Diese Abstimmung hat noch nicht begonnen.",
-                            div='status')
-        f = self.response.forms['deleteballotform']
-        f['ack_delete'].checked = True
+    def test_attachment(self) -> None:
+        with open(self.testfile_dir / "rechen.pdf", 'rb') as datafile:
+            data = datafile.read()
+
+        self.traverse("Versammlungen", "Archiv-Sammlung")
+
+        self.assertPresence("Rechenschaftsbericht", div="attachment1_version1")
+        self.assertPresence("Kassenprüferbericht 2 (Version 3)",
+                            div="attachment2_version3")
+        self.assertPresence("Liste der Kandidaten", div="attachment3_version1")
+
+        # Check file content.
+        saved_response = self.response
+        self.traverse({"href": "/assembly/assembly/3/attachment/1/latest"})
+        self.assertEqual(data, self.response.body)
+        self.response = saved_response
+
+        # Test Details link
+        self.traverse({"href": "assembly/assembly/3/attachments#attachment2_version3"})
+        self.assertTitle("Dateien (Archiv-Sammlung)")
+
+        self.assertPresence("Rechenschaftsbericht (Version 1)",
+                            div="attachment1_version1")
+        self.assertPresence("Kassenprüferbericht 2 (Version 3)",
+                            div="attachment2_version3")
+        self.assertPresence("Version 2 wurde gelöscht", div="attachment2_version2")
+        self.assertPresence("Kassenprüferbericht (Version 1)",
+                            div="attachment2_version1")
+        self.assertPresence("Liste der Kandidaten (Version 1)",
+                            div="attachment3_version1")
+
+        # remove Kassenprüferbericht:
+        base_link = '/assembly/assembly/3/attachment/2/'
+        # all-in-one is prohibited
+        self.post(base_link + 'delete', {'attachment_ack_delete': True})
+        self.assertNotification("Entferne erst alle bis auf eine Dateiversion,"
+                                " bevor Du die gesamte Datei entfernst.",
+                                'error')
+        # version 4 never existed
+        self.post(base_link + 'version/4/delete', {'attachment_ack_delete': True})
+        self.assertNotification("Diese Version der Datei existiert nicht.", 'error')
+        # version 2 is already deleted
+        self.post(base_link + 'version/2/delete', {'attachment_ack_delete': True})
+        self.assertNotification("Diese Version der Datei wurde bereits gelöscht.",
+                                'error')
+        # version 3 can be deleted
+        f = self.response.forms["removeattachmentversionform2_3"]
+        self.submit(f, check_notification=False)
+        # TODO: there is no validation error near the checkbox
+        self.assertNotification("Validierung fehlgeschlagen.", 'error')
+        f["attachment_ack_delete"] = True
         self.submit(f)
-        self.assertTitle("Abstimmungen (Internationaler Kongress)")
-        self.assertNonPresence("Farbe des Logos")
-        self.assertNonPresence("Zukünftige Abstimmungen")
-        self.traverse({"description": "Lieblingszahl"})
-        self.assertTitle("Lieblingszahl (Internationaler Kongress)")
-        self.assertPresence("Die Abstimmung läuft.", div='status')
-        self.assertNonPresence("Löschen")
-        self.assertNotIn("deleteballotform", self.response.forms)
-        self.get("/assembly/assembly/1/ballot/list")
-        self.traverse({"description": "Antwort auf die letzte aller Fragen"})
-        self.assertTitle(
-            "Antwort auf die letzte aller Fragen (Internationaler Kongress)")
-        self.assertPresence("Die Abstimmung ist beendet.", div='status')
-        self.assertNonPresence("Löschen")
-        self.assertNotIn("deleteballotform", self.response.forms)
+        self.assertPresence("Version 3 wurde gelöscht", div="attachment2_version3")
+        # version 3 is removed now
+        self.get(base_link + 'version/3/')
+        self.assertTitle("Dateien (Archiv-Sammlung)")
+        self.assertNotification("Datei wurde nicht gefunden.", 'error')
+        # deleting last version specifically is neither possible nor reachable via UI...
+        self.post(base_link + 'version/1/delete', {'attachment_ack_delete': True})
+        self.assertNotification(
+            "Die letzte verbleibende Version einer Datei kann nicht gelöscht werden.",
+            'error')
+        # ... since the button links to deletion of the whole attachment:
+        f = self.response.forms["deleteattachmentform2"]
+        self.submit(f, check_notification=False)
+        self.assertNotification("Validierung fehlgeschlagen.", 'error')
+        f["attachment_ack_delete"] = True
+        self.submit(f)
+        self.assertNonPresence("Kassenprüferbericht")
+
+        # check log
+        self.traverse("Log")
+        self.assertPresence("Anhangsversion entfernt", div="1-1001")
+        self.assertPresence("Kassenprüferbericht 2: Version 3", div="1-1001")
+        self.assertPresence("Anhang entfernt", div="2-1002")
+        self.assertPresence("Kassenprüferbericht", div="2-1002")
 
     @storage
-    @as_users("werner", "ferdinand")
-    def test_attachments(self) -> None:
+    @as_users("werner")
+    def test_attachment_ballot_linking(self) -> None:
         with open(self.testfile_dir / "form.pdf", 'rb') as datafile:
             data = datafile.read()
         self.traverse({'description': 'Versammlungen$'},
                       {'description': 'Internationaler Kongress'},
-                      {'description': 'Datei-Übersicht'})
-        if self.user_in(6):
-            f = self.response.forms['adminviewstoggleform']
-            self.submit(f, button="view_specifier", value="-assembly_presider")
-            self.assertTitle("Datei-Übersicht (Internationaler Kongress)")
-            self.assertPresence("Es wurden noch keine Dateien hochgeladen.")
-            f = self.response.forms['adminviewstoggleform']
-            self.submit(f, button="view_specifier", value="+assembly_presider")
-        self.traverse({'description': r"\sÜbersicht"},
-                      {'description': "Datei hinzufügen"})
+                      {'description': 'Dateien'})
+        self.assertTitle("Dateien (Internationaler Kongress)")
+        self.traverse("Datei hinzufügen")
         self.assertTitle("Datei hinzufügen (Internationaler Kongress)")
 
         # First try upload with invalid default filename
         f = self.response.forms['addattachmentform']
-        f['title'] = "Maßgebliche Beschlussvorlage"
+        f['title'] = "Vorläufige Beschlussvorlage"
         f['attachment'] = webtest.Upload("form….pdf", data, "application/octet-stream")
         self.submit(f, check_notification=False)
         self.assertValidationError(
@@ -772,68 +1143,133 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         f['attachment'] = webtest.Upload("form….pdf", data, "application/octet-stream")
         f['filename'] = "beschluss.pdf"
         self.submit(f)
-        self.assertTitle(
-            "Datei-Details (Internationaler Kongress) – Maßgebliche Beschlussvorlage")
+
+        self.assertTitle("Dateien (Internationaler Kongress)")
+        self.assertPresence(
+            "Vorläufige Beschlussvorlage", div="attachment1001_version1")
+        self.assertIn("deleteattachmentform1001", self.response.forms)
 
         # Check file content.
         saved_response = self.response
-        self.traverse({'description': 'Maßgebliche Beschlussvorlage'},)
+        self.traverse({'description': 'Vorläufige Beschlussvorlage'},)
         self.assertEqual(data, self.response.body)
         self.response = saved_response
 
-        # Change version data.
-        self.traverse({'href': '/assembly/assembly/1/attachment/1001/version/1/edit'})
-        self.assertTitle("Version bearbeiten (Internationaler Kongress)"
-                         " – Maßgebliche Beschlussvorlage (Version 1)")
-        f = self.response.forms['editattachmentversionform']
-        f['title'] = "Insignifikante Beschlussvorlage"
+        # Add a new version
+        self.traverse({"href": "/assembly/assembly/1/attachment/1001/add"})
+        self.assertNonPresence("Eine verknüpfte Abstimmung wurde bereits gesperrt",
+                               div='static-notifications', check_div=True)
+        f = self.response.forms['addattachmentversionform']
+        f['title'] = "Maßgebliche Beschlussvorlage"
         f['authors'] = "Der Vorstand"
+        f['attachment'] = webtest.Upload("form.pdf", data, "application/octet-stream")
         self.submit(f)
-        self.assertTitle("Datei-Details (Internationaler Kongress)"
-                         " – Insignifikante Beschlussvorlage")
-        self.traverse({'description': 'Datei-Verknüpfung ändern'})
-        self.assertTitle("Datei-Verknüpfung ändern (Internationaler Kongress)"
-                         " – Insignifikante Beschlussvorlage")
-        f = self.response.forms['changeattachmentlinkform']
-        f['new_ballot_id'] = 2
-        self.submit(f)
-        self.assertTitle("Datei-Details (Internationaler Kongress/Farbe des Logos)"
-                         " – Insignifikante Beschlussvorlage")
-        self.traverse({'description': 'Version hinzufügen'})
-        self.assertTitle("Version hinzufügen (Internationaler Kongress/Farbe des Logos)"
-                         " – Insignifikante Beschlussvorlage")
-        f = self.response.forms['addattachmentform']
-        f['title'] = "Alternative Beschlussvorlage"
-        f['authors'] = "Die Wahlleitung"
-        f['attachment'] = webtest.Upload("beschluss2.pdf", data + b'123',
-                                         "application/octet-stream")
-        self.submit(f)
-        self.assertTitle("Datei-Details (Internationaler Kongress/Farbe des Logos)"
-                         " – Alternative Beschlussvorlage")
-        self.assertPresence("Insignifikante Beschlussvorlage (Version 1)")
-        self.assertPresence("Alternative Beschlussvorlage (Version 2)")
-        f = self.response.forms['removeattachmentversionform1001_1']
-        f['attachment_ack_delete'].checked = True
-        self.submit(f)
-        self.assertTitle("Datei-Details (Internationaler Kongress/Farbe des Logos)"
-                         " – Alternative Beschlussvorlage")
-        self.assertPresence("Version 1 wurde gelöscht")
 
-        # Now check the attachment over view without the presider admin view.
-        self.traverse({'description': "Datei-Übersicht"})
-        if self.user_in(6):
-            f = self.response.forms['adminviewstoggleform']
-            self.submit(f, button="view_specifier", value="-assembly_presider")
-            self.assertTitle("Datei-Übersicht (Internationaler Kongress)")
-            self.assertPresence("Alternative Beschlussvorlage (Version 2)")
-            self.assertPresence("Version 1 wurde gelöscht.")
-            self.assertNonPresence("Es wurden noch keine Dateien hochgeladen.")
-            f = self.response.forms['adminviewstoggleform']
-            self.submit(f, button="view_specifier", value="+assembly_presider")
-        f = self.response.forms['deleteattachmentform1001']
-        f['attachment_ack_delete'].checked = True
+        self.assertTitle("Dateien (Internationaler Kongress)")
+        self.assertPresence(
+            "Vorläufige Beschlussvorlage", div="attachment1001_version1")
+        self.assertIn("removeattachmentversionform1001_1", self.response.forms)
+        self.assertPresence(
+            "Maßgebliche Beschlussvorlage", div="attachment1001_version2")
+        self.assertIn("removeattachmentversionform1001_2", self.response.forms)
+        self.assertNotIn("deleteattachmentform1001", self.response.forms)
+
+        # Link the attachment with a ballot
+        self.traverse("Abstimmungen", "Farbe des Logos")
+        self.assertPresence("Zu dieser Abstimmung gibt es noch keine Dateien.",
+                            div="attachments")
+        self.traverse("Bearbeiten")
+        f = self.response.forms["configureballotform"]
+        f["linked_attachments"] = ["1001"]
         self.submit(f)
+
         self.assertTitle("Farbe des Logos (Internationaler Kongress)")
+        self.assertPresence("Maßgebliche Beschlussvorlage (Version 2)",
+                            div="attachments")
+        # check that the correct version is linked
+        saved_response = self.response
+        self.traverse({"href": "/assembly/assembly/1/attachment/1001/version/2/"})
+        self.response = saved_response
+
+        # now start voting
+        base_time = now()
+        delta = datetime.timedelta(days=1)
+        with freezegun.freeze_time(base_time) as frozen_time:
+            self.traverse("Bearbeiten")
+            f = self.response.forms['configureballotform']
+            f['vote_begin'] = base_time + delta
+            f['vote_end'] = base_time + 3*delta
+            f['vote_extension_end'] = base_time + 5*delta
+            f['abs_quorum'] = "0"
+            f['rel_quorum'] = "100"
+            self.submit(f)
+            frozen_time.tick(delta=2*delta)
+
+            self.traverse("Abstimmungen", "Farbe des Logos")
+            self.assertTitle("Farbe des Logos (Internationaler Kongress)")
+            self.assertPresence("Die Abstimmung läuft.", div="ballot-status")
+            self.assertPresence("Maßgebliche Beschlussvorlage (Version 2, maßgeblich)",
+                                div="attachments")
+
+            # check that the attachment can not be deleted anymore
+            self.traverse(
+                {"href": "/assembly/assembly/1/attachments#attachment1001_version2"})
+            self.assertTitle("Dateien (Internationaler Kongress)")
+            self.assertNotIn("removeattachmentversionform1001_1", self.response.forms)
+            self.assertNotIn("removeattachmentversionform1001_2", self.response.forms)
+
+            # add a new version
+            self.traverse({"href": "/assembly/assembly/1/attachment/1001/add"})
+            self.assertNotification(
+                "Eine verknüpfte Abstimmung wurde bereits gesperrt, daher sollten nur"
+                " noch formale Korrekturen an dieser Datei vorgenommen werden. Die"
+                " maßgebliche Version bleibt für alle bereits begeonnen Abstimmungen"
+                " maßgeblich.",
+                ntype='warning',
+                static=True
+            )
+            f = self.response.forms['addattachmentversionform']
+            f['title'] = "Formal geänderte Beschlussvorlage"
+            f['attachment'] = webtest.Upload("form.pdf", data,
+                                             "application/octet-stream")
+            self.submit(f, check_notification=False)
+            self.assertValidationError("ack_creation", "Muss markiert sein.")
+            f = self.response.forms['addattachmentversionform']
+            f['title'] = "Formal geänderte Beschlussvorlage"
+            f['attachment'] = webtest.Upload("form.pdf", data,
+                                             "application/octet-stream")
+            f['ack_creation'] = True
+            self.submit(f)
+
+            self.assertTitle("Dateien (Internationaler Kongress)")
+            self.assertPresence(
+                "Vorläufige Beschlussvorlage", div="attachment1001_version1")
+            self.assertPresence(
+                "Maßgebliche Beschlussvorlage", div="attachment1001_version2")
+            self.assertPresence(
+                "Formal geänderte Beschlussvorlage", div="attachment1001_version3")
+            self.assertNotIn("removeattachmentversionform1001_3", self.response.forms)
+
+            # check the definitive version is still correct and the new version is shown
+            self.traverse("Abstimmungen", "Farbe des Logos")
+            self.assertPresence("Die Abstimmung läuft.", div="ballot-status")
+            self.assertPresence("Maßgebliche Beschlussvorlage (Version 2, maßgeblich)",
+                                div="attachments")
+            self.assertPresence(
+                "Formal geänderte Beschlussvorlage (Version 3)", div="attachments")
+
+        # check log
+        self.traverse("Log")
+        self.assertPresence("Anhang hinzugefügt", div="1-1001")
+        self.assertPresence("Vorläufige Beschlussvorlage", div="1-1001")
+        self.assertPresence("Anhangsversion hinzugefügt", div="2-1002")
+        self.assertPresence("Maßgebliche Beschlussvorlage: Version 2", div="2-1002")
+        self.assertPresence("Anhang mit Abstimmung verknüpft", div="5-1005")
+        self.assertPresence("Maßgebliche Beschlussvorlage (Farbe des Logos)",
+                            div="5-1005")
+        self.assertPresence("Anhangsversion hinzugefügt", div="8-1008")
+        self.assertPresence("Formal geänderte Beschlussvorlage: Version 3",
+                            div="8-1008")
 
     @storage
     @as_users("werner", "inga", "kalif")
@@ -855,6 +1291,22 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertTitle("Lieblingszahl (Internationaler Kongress)")
         f = self.response.forms['voteform']
         self.assertEqual("1=i>pi>e=0", f['vote'].value)
+        # test abstaining, with and without bar
+        f['vote'] = ""
+        self.submit(f)
+        self.assertPresence("Du hast Dich enthalten.", div='status')
+        f = self.response.forms['voteform']
+        self.assertEqual("e=pi=i=1=0", f['vote'].value)
+        self.traverse("Abstimmungen", "Wahl des Innenvorstand")
+        self.assertTitle("Wahl des Innenvorstand (Internationaler Kongress)")
+        self.submit(self.response.forms['voteform'])
+        f = self.response.forms['voteform']
+        self.assertEqual(f"Anton=Berta=Akira={ASSEMBLY_BAR_SHORTNAME}", f['vote'].value)
+        # invalid candidates - test validation errors
+        f['vote'] = "Werner>Anton"
+        self.submit(f, check_notification=False)
+        self.assertValidationError('vote', "Unerwartete Kandidaten gefunden.")
+        self.assertValidationError('vote', "Nicht alle Kandidaten vorhanden.")
 
     @storage
     @as_users("werner", "inga", "kalif")
@@ -924,6 +1376,23 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertEqual("E", f.get('vote', index=0).value)
         self.assertEqual(None, f.get('vote', index=1).value)
         self.assertEqual(None, f.get('vote', index=2).value)
+        # test error cases
+        f = self.response.forms['voteform']
+        f['vote'] = ["E", ASSEMBLY_BAR_SHORTNAME]
+        self.submit(f, check_notification=False)
+        self.assertNotification("Ablehnung kann nur exklusiv gewählt werden.", 'error')
+        f['vote'] = []
+        f.get('vote', index=0).force_value("")
+        f.get('vote', index=0).checked = True
+        self.submit(f, check_notification=False)
+        self.assertValidationError('vote', "Darf nicht leer sein.", index=0)
+        # ballot without _bar_
+        self.traverse("Abstimmungen", "Antrag zur DSGVO 2.0")
+        f = self.response.forms["voteform"]
+        f.get('vote', index=0).force_value(ASSEMBLY_BAR_SHORTNAME)
+        f.get('vote', index=0).checked = True
+        with self.assertRaisesRegex(ValueError, "Option not available"):
+            self.submit(f)
 
     @storage
     @as_users("werner")
@@ -1012,6 +1481,46 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
                       {'description': 'Ergebnisdatei herunterladen'},)
         with open(self.testfile_dir / "ballot_result.json", 'rb') as f:
             self.assertEqual(json.load(f), json.loads(self.response.body))
+        self.get('/assembly/assembly/1/ballot/3/result/download')  # running ballot
+        self.assertTitle("Bester Hof (Internationaler Kongress)")
+        self.assertNotification("Abstimmung noch nicht ausgezählt.", 'warning')
+
+    @storage
+    @as_users("kalif")
+    def test_late_voting(self) -> None:
+        # create a ballot shortly before its voting end
+        base_time = now()
+        delta = datetime.timedelta(seconds=42)
+        btitle = "Ganz kurzfristige Entscheidung"
+        bdata = {
+            'title': btitle,
+            'vote_begin': base_time + delta,
+            'vote_end': base_time + 3 * delta,
+            'votes': "2",
+        }
+        candidates = [
+            {'shortname': "y", 'title': "Ja!"},
+            {'shortname': "n", 'title': "Nein!"},
+        ]
+        with freezegun.freeze_time(base_time) as frozen_time:
+            # only presiders can create ballots
+            with self.switch_user('werner'):
+                self.traverse("Versammlungen", "Internationaler Kongress")
+                self._create_ballot(bdata, candidates)
+
+            # wait for voting to start then get vote form.
+            frozen_time.tick(delta=2 * delta)
+            self.traverse("Versammlungen", "Internationaler Kongress",
+                          "Abstimmungen", btitle)
+            f = self.response.forms["voteform"]
+            f["vote"] = ["y"]
+
+            # submit after voting period ended
+            frozen_time.tick(delta=2 * delta)
+            self.submit(f, check_notification=False)
+            self.assertPresence("Fehler! Abstimmung ist außerhalb"
+                                " des Abstimmungszeitraums",
+                                div='notifications')
 
     @storage
     @as_users("anton")
@@ -1021,8 +1530,6 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
             assembly = self.get_sample_datum(
                 'assembly.assemblies', ballot['assembly_id'])
             self.get(f'/assembly/assembly/{assembly["id"]}/ballot/{ballot_id}/result')
-            # redirect in case of tallying, extending etc
-            self.follow()
 
             # redirect to show_ballot if the ballot has not been tallied yet
             if ballot_id in self.BALLOT_STATES['tallied']:
@@ -1167,7 +1674,7 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
                           {'description': 'Internationaler Kongress'},
                           {'description': 'Abstimmungen'},
                           {'description': 'Abstimmung anlegen'},)
-            f = self.response.forms['createballotform']
+            f = self.response.forms['configureballotform']
             f['title'] = 'Maximale Länge der Verfassung'
             f['vote_begin'] = base_time + delta
             f['vote_end'] = base_time + 3*delta
@@ -1177,16 +1684,17 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
             f['votes'] = ""
             self.submit(f)
             self.assertTitle("Maximale Länge der Verfassung (Internationaler Kongress)")
+            ballot = self.assembly.get_ballot(self.key, 1001)
             self.assertPresence(
-                "Verlängerung bis 01.05.2037, 00:00:00, falls 10 Stimmen nicht "
-                "erreicht werden.", div='voting-period')
+                f"Verlängerung bis 01.05.2037, 00:00:00, falls {ballot['quorum']}"
+                f" Stimmen nicht erreicht werden.", div='voting-period')
 
             frozen_time.tick(delta=4*delta)
             self.traverse({'href': '/assembly/1/ballot/list'},
                           {'description': 'Maximale Länge der Verfassung'},)
             self.assertTitle("Maximale Länge der Verfassung (Internationaler Kongress)")
-            s = ("Wurde bis 01.05.2037, 00:00:00 verlängert, da 10 Stimmen nicht "
-                 "erreicht wurden.")
+            s = (f"Wurde bis 01.05.2037, 00:00:00 verlängert, da {ballot['quorum']}"
+                 f" Stimmen nicht erreicht wurden.")
             self.assertPresence(s, div='voting-period')
 
     @storage
@@ -1201,7 +1709,9 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertEqual("rot", f['shortname_6'].value)
         self.assertEqual("gelb", f['shortname_7'].value)
         self.assertEqual("gruen", f['shortname_8'].value)
-        self.assertNotIn("Dunkelaquamarin", f.fields)
+        self.assertNotIn(
+            "Dunkelaquamarin",
+            (field.value for field_list in f.fields.values() for field in field_list))
         f['create_-1'].checked = True
         f['shortname_-1'] = "aqua"
         f['title_-1'] = "Dunkelaquamarin"
@@ -1217,7 +1727,7 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.assertTitle("Farbe des Logos (Internationaler Kongress)")
         f = self.response.forms['candidatessummaryform']
         f['shortname_7'] = "gelb"
-        f['shortname_8'] = "_bar_"
+        f['shortname_8'] = ASSEMBLY_BAR_SHORTNAME
         self.submit(f, check_notification=False)
         self.assertValidationError(
             "shortname_8", "Darf nicht der Bezeichner der Ablehnungsoption sein")
@@ -1230,11 +1740,27 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
         self.submit(f)
 
         self.assertTitle("Farbe des Logos (Internationaler Kongress)")
+        f = self.response.forms['candidatessummaryform']
         self.assertEqual("rot", f['shortname_6'].value)
         self.assertEqual("lila", f['title_6'].value)
         self.assertEqual("farbe", f['shortname_8'].value)
         self.assertEqual("aqua", f['shortname_1001'].value)
-        self.assertNotIn("gelb", f.fields)
+        self.assertNotIn(
+            "gelb",
+            (field.value for field_list in f.fields.values() for field in field_list))
+
+    @storage
+    @as_users("werner")
+    def test_start_voting_button(self) -> None:
+        if not self.conf['CDEDB_DEV']:
+            self.skipTest("Not in development mode.")
+        link = '/assembly/assembly/1/ballot/2/show'
+        self.get(link)
+        self.submit(self.response.forms['startvotingform'])
+        self.assertPresence("Die Abstimmung läuft")
+        with freezegun.freeze_time(now() + datetime.timedelta(seconds=60)):
+            self.get(link)
+            self.assertPresence("Die Abstimmung ist beendet.")
 
     @storage
     @as_users("werner")
@@ -1262,18 +1788,15 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
 
     @storage
     def test_provide_secret(self) -> None:
-        self.login("anton")
-        self.logout()
         base_time = now()
         delta = datetime.timedelta(seconds=42)
         with freezegun.freeze_time(base_time,
                                    ignore=['cdedb.filter', 'icu']) as frozen_time:
-            user = USER_DICT["werner"]
-            self.login(user)
+            self.login('werner')
             self.traverse({'description': 'Versammlungen'},
                           {'description': 'Archiv-Sammlung'})
             # werner is no member, so he must signup external
-            secret = self._external_signup(user)
+            secret = self._external_signup('werner')
             # Create new ballot.
             bdata = {
                 'title': 'Maximale Länge der Verfassung',
@@ -1316,24 +1839,23 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
             f = self.response.forms['deleteballotform']
             f['ack_delete'].checked = True
             self.submit(f)
-
-            self.logout()
-            self.login("anton")
-            # Conclude assembly.
-            self.traverse({'description': "Versammlung"},
-                          {'description': "Archiv-Sammlung"})
-            self.assertTitle("Archiv-Sammlung")
-            f = self.response.forms['concludeassemblyform']
-            f['ack_conclude'].checked = True
+            self.traverse({'description': 'Abstimmungen'},
+                          {'description': 'Genauso wichtige Wahl'})
+            f = self.response.forms['deleteballotform']
+            f['ack_delete'].checked = True
             self.submit(f)
 
-            self.logout()
-            self.login(user)
+            # Conclude assembly.
+            with self.switch_user('viktor'):
+                self.traverse({'description': "Versammlung"},
+                              {'description': "Archiv-Sammlung"})
+                self.assertTitle("Archiv-Sammlung")
+                f = self.response.forms['concludeassemblyform']
+                f['ack_conclude'].checked = True
+                self.submit(f)
+
             # Own vote should be hidden now.
-            self.traverse({'description': "Versammlung"},
-                          {'description': 'Archiv-Sammlung'},
-                          {'description': 'Abstimmungen'},
-                          {'description': 'Maximale Länge der Verfassung'},
+            self.traverse({'description': 'Maximale Länge der Verfassung'},
                           {'description': 'Ergebnisdetails'})
             s = ("Die Versammlung wurde beendet. Das Abstimmungsverhalten einzelner"
                  " Nutzer ist nicht mehr aus der Datenbank auslesbar.")
@@ -1342,6 +1864,13 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
 
             # Provide the secret to retrieve the vote.
             f = self.response.forms['showoldvoteform']
+            f[ANTI_CSRF_TOKEN_NAME] = "evil"
+            self.submit(f, check_notification=False)
+            self.assertNotification("Der Anti-CSRF-Token wurde gefälscht.", 'error')
+            f = self.response.forms['showoldvoteform']
+            f['secret'] = "IForgotMySecret"
+            self.submit(f, check_notification=False)
+            self.assertValidationError('secret', "Ungültiges Geheimnis eingegeben.")
             f['secret'] = secret
             self.submit(f, check_notification=False)
             self.assertNonPresence("Die Versammlung wurde beendet und die "
@@ -1349,88 +1878,68 @@ class TestAssemblyFrontend(AssemblyTestHelpers):
             self.assertPresence("Du hast für die folgenden Kandidaten gestimmt: Ja",
                                 div='own-vote', exact=True)
 
-    @storage
-    def test_log(self) -> None:
-        # First: generate data
-        self.test_entity_ballot_simple()
-        self.logout()
-        self.test_conclude_assembly()
-        # test_tally_and_get_result
-        self.traverse({'description': 'Versammlungen'},
-                      {'description': 'Internationaler Kongress'},
-                      {'description': 'Abstimmungen'},
-                      {'description': 'Antwort auf die letzte aller Fragen'},)
-        self.logout()
-        self.test_extend()
-        self.logout()
-
-        # Now check it
-        self.login(USER_DICT['anton'])
-        self.traverse({'description': 'Versammlungen'},
-                      {'description': 'Log'})
-        self.assertTitle("Versammlungs-Log [1–26 von 26]")
-        self.assertNonPresence("LogCodes")
-        f = self.response.forms['logshowform']
-        codes = [const.AssemblyLogCodes.assembly_created.value,
-                 const.AssemblyLogCodes.assembly_changed.value,
-                 const.AssemblyLogCodes.ballot_created.value,
-                 const.AssemblyLogCodes.ballot_changed.value,
-                 const.AssemblyLogCodes.ballot_deleted.value,
-                 const.AssemblyLogCodes.ballot_tallied.value,
-                 const.AssemblyLogCodes.assembly_presider_added.value,
-                 ]
-        f['codes'] = codes
-        f['assembly_id'] = 1
-        self.submit(f)
-        self.assertTitle("Versammlungs-Log [1–8 von 8]")
-
-        self.logout()
-        self.login("werner")
-        self.traverse({'description': 'Versammlungen'},
-                      {'description': 'Drittes CdE-Konzil'},
-                      {'description': 'Log'})
-        self.assertTitle("Drittes CdE-Konzil: Log [1–8 von 8]")
-
-        f = self.response.forms['logshowform']
-        f['codes'] = codes
-        f['offset'] = 2
-        self.submit(f)
-        self.assertTitle("Drittes CdE-Konzil: Log [3–52 von 6]")
+            # providing secret for running ballot not possible
+            self.post('/assembly/assembly/1/ballot/3/result',
+                      {'secret': "-YZN1KWDfMLQ5Y8Q"})
+            self.assertTitle("Bester Hof (Internationaler Kongress)")
+            self.assertNotification("Abstimmung wurde noch nicht ausgezählt.", 'error')
 
 
 class TestMultiAssemblyFrontend(MultiAppFrontendTest, AssemblyTestHelpers):
     n = 2
 
     def test_presiders(self) -> None:
-        self.login("anton")
+        self.login("viktor")
         self._create_assembly()
         self._external_signup(USER_DICT["werner"])
         self.traverse(r"\sÜbersicht")
         self.assertPresence("Werner Wahlleitung", div='assembly-presiders')
+
         self.switch_app(1)
         self.login("werner")
         self.traverse("Versammlung", "Drittes CdE-Konzil", "Konfiguration")
-        f = self.response.forms['changeassemblyform']
+        f = self.response.forms['configureassemblyform']
         f['notes'] = "Werner war hier!"
         self.submit(f)
         self.assertTitle("Drittes CdE-Konzil")
         self.assertPresence("Werner war hier!", div='notes')
-        self.assertNotIn('removepresiderform23', self.response.forms)
+        self.assertNotIn(f"removepresiderform{USER_DICT['werner']['id']}",
+                         self.response.forms)
         self.traverse("Log")
+
         self.switch_app(0)
         self.traverse(r"\sÜbersicht")
         self.assertPresence("Werner war hier!", div='notes')
-        f = self.response.forms['removepresiderform23']
-        self.submit(f, verbose=True)
+        f = self.response.forms[f"removepresiderform{USER_DICT['werner']['id']}"]
+        f['presider_id'] = "ThisIsNoID"
+        self.submit(f, check_notification=False)
+        self.assertNotification("Validierung fehlgeschlagen.", 'error')
+        f = self.response.forms[f"removepresiderform{USER_DICT['werner']['id']}"]
+        self.submit(f)
+        self.assertNotIn(f"removepresiderform{USER_DICT['werner']['id']}",
+                         self.response.forms)
+        self.submit(f, check_notification=False)
+        self.assertNotification(
+            "Dieser Nutzer ist kein Wahlleiter für diese Versammlung.")
         self.assertNonPresence("Werner Wahlleitung", div='assembly-presiders',
                                check_div=False)
+
         self.switch_app(1)
         self.traverse(r"\sÜbersicht")
         self.assertNonPresence("Werner war hier!")
         self.assertNoLink("Konfiguration")
         self.assertNoLink("Log")
+
         self.switch_app(0)
         f = self.response.forms['addpresidersform']
+        # wrong format
+        f['presider_ids'] = "DB-1000-X...DB-0-abc"
+        self.submit(f, check_notification=False)
+        self.assertValidationError("presider_ids", "Falsches Format.")
+        # invalid id
+        f['presider_ids'] = "DB-1000-X"
+        self.submit(f, check_notification=False)
+        self.assertValidationError("presider_ids", "Checksumme stimmt nicht.")
         # Try non-existing user.
         f['presider_ids'] = "DB-1000-6"
         self.submit(f, check_notification=False)
@@ -1453,6 +1962,7 @@ class TestMultiAssemblyFrontend(MultiAppFrontendTest, AssemblyTestHelpers):
         f['presider_ids'] = USER_DICT["werner"]['DB-ID']
         self.submit(f)
         self.assertPresence("Werner Wahlleitung", div='assembly-presiders')
+
         self.switch_app(1)
         self.traverse(r"\sÜbersicht")
         self.assertPresence("Werner war hier!", div='notes')
