@@ -8,10 +8,13 @@ from typing import Dict, Optional
 import webtest
 
 import cdedb.database.constants as const
-from cdedb.common import ADMIN_VIEWS_COOKIE_NAME, CdEDBObject, get_hash
+from cdedb.common import (
+    ADMIN_VIEWS_COOKIE_NAME, IGNORE_WARNINGS_NAME, CdEDBObject, GenesisDecision,
+    get_hash,
+)
 from cdedb.query import QueryOperators
 from tests.common import (
-    FrontendTest, UserIdentifier, UserObject, USER_DICT, as_users, storage
+    USER_DICT, FrontendTest, UserIdentifier, UserObject, as_users, get_user, storage,
 )
 
 
@@ -93,8 +96,8 @@ class TestCoreFrontend(FrontendTest):
             self.assertPresence("Nutzer verwalten", div='adminshowuser-box')
             self.assertPresence("Platin-Lounge", div='moderator-box')
             # Check moderation notification
-            self.assertPresence("Mailman-Migration", div='moderator-box')
-            self.traverse({'href': '/ml/mailinglist/99/moderate',
+            self.assertPresence("Moderatoren-Liste", div='moderator-box')
+            self.traverse({'href': '/ml/mailinglist/12/moderate',
                            'description': "3 E-Mails"})
             self.traverse({'href': '/'})
             self.assertTitle("CdE-Datenbank")
@@ -191,6 +194,36 @@ class TestCoreFrontend(FrontendTest):
         self.assertPresence("Kampfbrief-Kommentare (geblockt)")
         self.assertNonPresence("Witz des Tages")
 
+    @as_users("anton")
+    def test_user_archived(self) -> None:
+        def _check_redirected_profile() -> None:
+            self.follow()
+            self.assertTitle("Hades Hell")
+            self.assertPresence("Account ist archiviert.", div='notifications')
+
+        self.get('/core/persona/8/events')
+        _check_redirected_profile()
+        self.get('/core/persona/8/mailinglists')
+        _check_redirected_profile()
+        self.get('/core/persona/8/history')
+        _check_redirected_profile()
+        self.get('/core/persona/8/adminchange')
+        _check_redirected_profile()
+        self.get('/core/persona/8/privileges')
+        _check_redirected_profile()
+        self.get('/core/persona/8/promote')
+        _check_redirected_profile()
+        self.get('/core/persona/8/membership/change')
+        _check_redirected_profile()
+        self.get('/core/persona/8/balance/change')
+        _check_redirected_profile()
+        self.get('/core/persona/8/foto/change')
+        _check_redirected_profile()
+        self.get('/core/persona/8/username/adminchange')
+        _check_redirected_profile()
+        self.post('/core/persona/8/activity/change', {'activity': False})
+        _check_redirected_profile()
+
     @as_users("emilia", "janis")
     def test_showuser_self(self) -> None:
         name = f"{self.user['given_names']} {self.user['family_name']}"
@@ -226,11 +259,17 @@ class TestCoreFrontend(FrontendTest):
                  "FN:Bertålotta Beispiel",
                  "N:Beispiel;Bertålotta;;Dr.;MdB",
                  "NICKNAME:Bertå",
-                 "TEL;TYPE=\"home,voice\":+49 (5432) 987654321",
-                 "TEL;TYPE=\"cell,voice\":0163/123456789",
+                 "TEL;TYPE=\"home,voice\":+495432987654321",
+                 "TEL;TYPE=\"cell,voice\":+4916312345678",
                  "END:VCARD"]
         for line in vcard:
             self.assertIn(line, self.response.text)
+
+        self.get("/core/self/show")
+        self.follow()
+        self.traverse("QR")
+        # our modal javascript in cdedb_helper.js relies on xml content type
+        self.assertEqual(self.response.content_type, "image/svg+xml")
 
     @as_users("vera")
     def test_vcard_cde_admin(self) -> None:
@@ -350,6 +389,10 @@ class TestCoreFrontend(FrontendTest):
         expectation = (11,)
         reality = tuple(e['id'] for e in self.response.json['personas'])
         self.assertEqual(expectation, reality)
+        self.get('/core/persona/select?kind=other&phrase=@exam')
+        self.assertEqual({}, self.response.json)
+        self.get('/core/persona/select?kind=ml_user&phrase=@exam&aux=other')
+        self.assertEqual({}, self.response.json)
 
     @as_users("quintus")
     def test_selectpersona_two(self) -> None:
@@ -359,14 +402,21 @@ class TestCoreFrontend(FrontendTest):
         reality = tuple(e['id'] for e in self.response.json['personas'])
         self.assertEqual(expectation, reality)
 
-    @as_users("berta", "martin", "nina", "rowena", "vera", "viktor", "werner", "annika")
+    @as_users("berta", "martin", "nina", "rowena", "vera", "viktor", "werner", "annika",
+              "katarina")
     def test_selectpersona_403(self) -> None:
-        # These can not be done by Berta no matter what.
-        if not self.user_in("vera"):
+        if not self.user_in("vera", "katarina"):
             self.get('/core/persona/select?kind=admin_persona&phrase=@exam',
                      status=403)
             self.assertTitle('403: Forbidden')
+            self.get('/core/persona/select?kind=cde_user&phrase=@exam',
+                     status=403)
+            self.assertTitle('403: Forbidden')
             self.get('/core/persona/select?kind=past_event_user&phrase=@exam',
+                     status=403)
+            self.assertTitle('403: Forbidden')
+        if not self.user_in("katarina", "viktor", "werner"):
+            self.get('/core/persona/select?kind=assembly_user&phrase=@exam',
                      status=403)
             self.assertTitle('403: Forbidden')
         if not self.user_in("viktor", "werner"):
@@ -381,7 +431,7 @@ class TestCoreFrontend(FrontendTest):
             self.get('/core/persona/select?kind=pure_ml_user&phrase=@exam',
                      status=403)
             self.assertTitle('403: Forbidden')
-        if not self.user_in("annika", "berta"):
+        if not self.user_in("annika", "berta", "katarina"):
             self.get('/core/persona/select?kind=event_user&phrase=@exam',
                      status=403)
             self.assertTitle('403: Forbidden')
@@ -429,6 +479,8 @@ class TestCoreFrontend(FrontendTest):
         expectation = (1, 2, 5)
         reality = tuple(e['id'] for e in self.response.json['personas'])
         self.assertEqual(expectation, reality)
+        self.get('/core/persona/select'
+                '?kind=ml_subscriber&phrase=@exam', status=400)
         self.get('/core/persona/select'
                  '?kind=ml_subscriber&phrase=inga&aux=9')
         expectation = (9,)
@@ -487,6 +539,63 @@ class TestCoreFrontend(FrontendTest):
         expectation = (1, 2, 3)
         reality = tuple(e['id'] for e in self.response.json['personas'])
         self.assertEqual(expectation, reality)
+
+    @as_users("katarina")
+    def test_selectpersona_auditor(self) -> None:
+        self.get('/core/persona/select?kind=admin_persona&phrase=din')
+        expectation = {
+            'personas': [
+                {
+                    'id': 4,
+                    'name': 'Daniel Dino',
+                },
+                {
+                    'id': 6,
+                    'name': 'Ferdinand Findus',
+                },
+            ],
+        }
+        self.assertEqual(expectation, self.response.json)
+        self.get('/core/persona/select?kind=ml_user&phrase=@exam')
+        expectation = (1, 2, 3, 4, 5, 6, 7, 9, 10, 11, 13, 14)
+        reality = tuple(e['id'] for e in self.response.json['personas'])
+        self.assertEqual(expectation[:self.conf["NUM_PREVIEW_PERSONAS"]], reality)
+        self.get('/core/persona/select?kind=event_user&phrase=bert')
+        expectation = (2,)
+        reality = tuple(e['id'] for e in self.response.json['personas'])
+        self.assertEqual(expectation, reality)
+        self.get('/core/persona/select?kind=past_event_user&phrase=emil')
+        expectation = (5,)
+        reality = tuple(e['id'] for e in self.response.json['personas'])
+        self.assertEqual(expectation, reality)
+        self.get('/core/persona/select?kind=assembly_user&phrase=kalif')
+        expectation = (11,)
+        reality = tuple(e['id'] for e in self.response.json['personas'])
+        self.assertEqual(expectation, reality)
+        self.get('/core/persona/select?kind=pure_assembly_user&phrase=kal', status=403)
+        self.get('/core/persona/select?kind=pure_ml_user&phrase=@exam', status=403)
+        for ml_id in self.ml.list_mailinglists(self.key):
+            self.get(f'/core/persona/select?kind=ml_subscriber'
+                     f'&phrase=@exam&aux={ml_id}', status=403)
+
+    @as_users("paul")
+    def test_selectpersona_ids(self) -> None:
+        self.get('/core/persona/select?kind=admin_persona&phrase=DB-2-7')
+        expectation = (2,)
+        reality = tuple(e['id'] for e in self.response.json['personas'])
+        self.assertEqual(expectation, reality)
+        self.get('/core/persona/select?kind=ml_user&phrase=14')
+        expectation = (14,)
+        reality = tuple(e['id'] for e in self.response.json['personas'])
+        self.assertEqual(expectation, reality)
+
+    @as_users("quintus")
+    def test_selectpersona_ids_unprivileged(self) -> None:
+        self.get('/core/persona/select?kind=admin_persona&phrase=DB-2-7')
+        reality = tuple(e['id'] for e in self.response.json['personas'])
+        self.assertEqual(tuple(), reality)
+        self.get('/core/persona/select?kind=cde_user&phrase=14')
+        self.assertEqual({}, self.response.json)
 
     @as_users("vera")
     def test_adminshowuser_advanced(self) -> None:
@@ -691,11 +800,34 @@ class TestCoreFrontend(FrontendTest):
         new_password = 'krce84#(=kNO3xb'
         self.traverse({'description': self.user['display_name']},
                       {'description': 'Passwort ändern'})
+
+        # non-matching password
+        f = self.response.forms['passwordchangeform']
+        f['old_password'] = 'wrongpassword'
+        f['new_password'] = new_password
+        f['new_password2'] = 'something else'
+        self.submit(f, check_notification=False)
+        self.assertValidationError('new_password', "Passwörter stimmen nicht überein.",
+                                   notification="Passwörter stimmen nicht überein.")
+        self.assertValidationError('new_password2', "Passwörter stimmen nicht überein.",
+                                   notification="Passwörter stimmen nicht überein.")
+
+        # wrong old password
+        f = self.response.forms['passwordchangeform']
+        f['old_password'] = 'wrongpassword'
+        f['new_password'] = new_password
+        f['new_password2'] = new_password
+        self.submit(f, check_notification=False)
+        self.assertValidationError('old_password', "Passwort falsch.",
+                                   notification="Passwort ist falsch.")
+
+        # everything correct
         f = self.response.forms['passwordchangeform']
         f['old_password'] = self.user['password']
         f['new_password'] = new_password
         f['new_password2'] = new_password
         self.submit(f)
+
         self.logout()
         self.assertNonPresence(self.user['display_name'])
         self.login(self.user)
@@ -730,7 +862,6 @@ class TestCoreFrontend(FrontendTest):
                         continue
                     link = self.fetch_link()
                     self.get(link)
-                    self.follow()
                     self.assertTitle("Neues Passwort setzen")
                     f = self.response.forms['passwordresetform']
                     f['new_password'] = val
@@ -763,15 +894,22 @@ class TestCoreFrontend(FrontendTest):
         link = self.fetch_link()
         # First reset should work
         self.get(link)
-        self.follow()
         self.assertTitle("Neues Passwort setzen")
+        f = self.response.forms['passwordresetform']
+        f['new_password'] = new_password
+        f['new_password2'] = 'something else'
+        self.submit(f, check_notification=False)
+        self.assertPresence("Passwörter stimmen nicht überein.", div='notifications')
+        self.assertValidationError('new_password', "Passwörter stimmen nicht überein.",
+                                   notification="Passwörter stimmen nicht überein.")
+        self.assertValidationError('new_password2', "Passwörter stimmen nicht überein.",
+                                   notification="Passwörter stimmen nicht überein.")
         f = self.response.forms['passwordresetform']
         f['new_password'] = new_password
         f['new_password2'] = new_password
         self.submit(f)
         # Second reset with same link should fail
         self.get(link)
-        self.follow()
         self.assertTitle("Neues Passwort setzen")
         f = self.response.forms['passwordresetform']
         f['new_password'] = new_password
@@ -791,7 +929,6 @@ class TestCoreFrontend(FrontendTest):
         link = self.fetch_link()
         self.logout()
         self.get(link)
-        self.follow()
         self.assertTitle("Neues Passwort setzen")
         f = self.response.forms['passwordresetform']
         f['new_password'] = new_password
@@ -837,6 +974,10 @@ class TestCoreFrontend(FrontendTest):
         self.assertValidationError(
             "new_username", "Muss sich von der aktuellen E-Mail-Adresse unterscheiden.")
         self.assertNonPresence("E-Mail abgeschickt!", div="notifications")
+        # Now with taken username
+        f['new_username'] = "charly@example.cde"
+        self.submit(f, check_notification=False)
+        self.assertValidationError("new_username", "E-Mail-Adresse bereits vorhanden.")
         # Now with new username
         new_username = "zelda@example.cde"
         f = self.response.forms['usernamechangeform']
@@ -900,21 +1041,23 @@ class TestCoreFrontend(FrontendTest):
         self.get('/core/search/user')
         save = self.response
         self.response = save.click(description="Alle Admins")
-        self.assertPresence("Ergebnis [14]", div='query-results')
+        self.assertPresence("Ergebnis [16]", div='query-results')
         self.assertPresence("Akira", div='query-result')
         self.assertPresence("Anton Armin A.", div='query-result')
+        self.assertPresence("Annika", div='query-result')
         self.assertPresence("Beispiel", div='query-result')
+        self.assertPresence("Quintus", div='query-result')
         self.assertPresence("Farin", div='query-result')
         self.assertPresence("Findus", div='query-result')
         self.assertPresence("Generalis", div='query-result')
+        self.assertPresence("Inga", div='query-result')
+        self.assertPresence("Katarina", div='query-result')
         self.assertPresence("Meister", div='query-result')
-        self.assertPresence("Olaf", div='query-result')
-        self.assertPresence("Panther", div='query-result')
-        self.assertPresence("Quintus", div='query-result')
         self.assertPresence("Neubauer", div='query-result')
         self.assertPresence("Olafson", div='query-result')
-        self.assertPresence("Vera", div='query-result')
+        self.assertPresence("Panther", div='query-result')
         self.assertPresence("Viktor", div='query-result')
+        self.assertPresence("Vera", div='query-result')
 
     def test_privilege_change(self) -> None:
         # Grant new admin privileges.
@@ -965,8 +1108,9 @@ class TestCoreFrontend(FrontendTest):
         f['is_finance_admin'] = True
         f['notes'] = "Berta ist jetzt Praktikant der Finanz Vorstände."
         self.submit(f, check_notification=False)
-        self.assertPresence("Nur CdE Admins können Finanz Admin werden.",
-                            div='notifications')
+        self.assertValidationError(
+            "is_finance_admin",
+            "Diese Rolle kann nicht an nicht-CdE-Admin vergeben werden.")
         f['is_cde_admin'] = True
         f['notes'] = "Dann ist Berta jetzt eben CdE und Finanz Admin."
         self.submit(f)
@@ -1020,6 +1164,10 @@ class TestCoreFrontend(FrontendTest):
         other_user_name = "berta"
         self.admin_view_profile(other_user_name)
         f = self.response.forms["invalidatepasswordform"]
+        f["confirm_username"] = "something else"
+        self.submit(f, check_notification=False)
+        self.assertValidationError('confirm_username',
+                                   "Bitte gib die Emailadresse des Nutzers an.")
         f["confirm_username"] = USER_DICT[other_user_name]["username"]
         self.submit(f)
         self.logout()
@@ -1180,6 +1328,8 @@ class TestCoreFrontend(FrontendTest):
         self.traverse({'description': self.user['display_name']},
                       {'description': 'Profilbild ändern'})
         f = self.response.forms['setfotoform']
+        self.submit(f, check_notification=False)
+        self.assertValidationError('foto', "Darf nicht leer sein.")
         with open(self.testfile_dir / "picture.png", 'rb') as datafile:
             data = datafile.read()
         my_hash = get_hash(data)
@@ -1227,7 +1377,7 @@ class TestCoreFrontend(FrontendTest):
                 f[field].checked = True
         self.submit(f)
         self.assertTitle("Allgemeine Nutzerverwaltung")
-        self.assertPresence("Ergebnis [13]", div='query-results')
+        self.assertPresence("Ergebnis [14]", div='query-results')
         self.assertPresence("Jalapeño", div='query-result')
 
     @as_users("vera")
@@ -1260,6 +1410,12 @@ class TestCoreFrontend(FrontendTest):
         self.traverse({'description': 'Archivsuche'})
         self.assertTitle("Archivsuche")
         f = self.response.forms['queryform']
+        self.submit(f)
+        self.assertPresence("Ergebnis [2]", div='query-results')
+        self.assertPresence("Hell", div='query-result')
+        self.assertPresence("Lost", div='query-result')
+        self.assertNonPresence("N/A", div='query-result')
+
         f['qop_given_names'] = QueryOperators.match.value
         f['qval_given_names'] = 'des'
         for field in f.fields:
@@ -1269,23 +1425,6 @@ class TestCoreFrontend(FrontendTest):
         self.assertTitle("Archivsuche")
         self.assertPresence("Ergebnis [1]", div='query-results')
         self.assertPresence("Hell", div='query-result')
-
-    @as_users("vera")
-    def test_archived_user_search2(self) -> None:
-        self.traverse({'description': 'Archivsuche'})
-        self.assertTitle("Archivsuche")
-        f = self.response.forms['queryform']
-        f['qval_birthday'] = '31.12.2000'
-        f['qop_birthday'] = QueryOperators.less.value
-        for field in f.fields:
-            if field and field.startswith('qsel_'):
-                f[field].checked = True
-        self.submit(f)
-        self.assertTitle("Archivsuche")
-        self.assertPresence("Ergebnis [2]", div='query-results')
-        self.assertPresence("Hell", div='query-result')
-        self.assertPresence("Lost", div='query-result')
-        self.assertPresence("N/A", div='query-result')
 
     @as_users("vera")
     def test_show_archived_user(self) -> None:
@@ -1305,12 +1444,13 @@ class TestCoreFrontend(FrontendTest):
         f = self.response.forms['archivepersonaform']
         f['ack_delete'].checked = True
         self.submit(f, check_notification=False)
-        self.assertValidationError(
-            "note", "Darf nicht leer sein",
-            notification="Archivierungsnotiz muss angegeben werden.")
+        self.assertValidationError("note", "Darf nicht leer sein")
         self.assertTitle("Charly Clown")
         self.assertNonPresence("Der Benutzer ist archiviert.")
         self.assertPresence("Zirkusstadt", div='address')
+        f = self.response.forms['archivepersonaform']
+        f['ack_delete'].checked = False
+        self.submit(f, check_notification=False)
         f = self.response.forms['archivepersonaform']
         f['ack_delete'].checked = True
         f['note'] = "Archived for testing."
@@ -1330,6 +1470,8 @@ class TestCoreFrontend(FrontendTest):
         self.admin_view_profile('hades', check=False)
         self.assertTitle("Hades Hell")
         self.assertPresence("Der Benutzer ist archiviert.", div='archived')
+        f = self.response.forms['purgepersonaform']
+        self.submit(f, check_notification=False)
         f = self.response.forms['purgepersonaform']
         f['ack_delete'].checked = True
         self.submit(f)
@@ -1524,16 +1666,33 @@ class TestCoreFrontend(FrontendTest):
         self.assertTitle("Bereichsänderung für Emilia E. Eventis")
         f = self.response.forms['realmselectionform']
         self.assertNotIn("event", f['target_realm'].options)
+        f['target_realm'].force_value("event")
+        self.submit(f)
+        self.assertPresence("Keine Änderung erforderlich.", div='notifications')
         f['target_realm'] = "cde"
         self.submit(f)
         self.assertTitle("Bereichsänderung für Emilia E. Eventis")
         f = self.response.forms['promotionform']
         self.submit(f, check_notification=False)
+        f = self.response.forms['promotionform']
+        f['pevent_id'] = 2
+        self.assertPresence("Die Kursauswahl wird angezeigt, nachdem")
+        f['is_orga'] = True
         self.assertValidationError('change_note', "Darf nicht leer sein.")
-        f['change_note'] = change_note = "Hat an einer Akademie teilgenommen."
+        f['change_note'] = change_note = "Hat eine Akademie organisiert."
+        self.submit(f, check_notification=False)
+        f = self.response.forms['promotionform']
+        self.assertNonPresence("Die Kursauswahl wird angezeigt, nachdem")
+        f['pcourse_id'] = ''
+        self.submit(f, check_notification=False)
+        # ignore phone number ValidationWarning
+        # TODO list this warnings anywhere?
+        # self.assertValidationWarning("mobile", "Telefonnummer scheint invalide zu")
+        f[IGNORE_WARNINGS_NAME].checked = True
         self.submit(f)
         self.assertTitle("Emilia E. Eventis")
         self.assertPresence("0,00 €", div='balance')
+        self.assertPresence("Geburtstagsfete (Orga)", div="past-events")
         self.assertCheckbox(True, "paper_expuls_checkbox")
         self.assertNonPresence("CdE-Mitglied", div="cde-membership")
         self.assertNonPresence("Probemitgliedschaft", div="cde-membership")
@@ -1550,13 +1709,21 @@ class TestCoreFrontend(FrontendTest):
         self.submit(f)
         self.assertTitle("Bereichsänderung für Nina Neubauer")
         f = self.response.forms['promotionform']
+        f['pevent_id'] = 1
         f['trial_member'].checked = True
         f['change_note'] = "Per Vorstandsbeschluss aufgenommen."
+        self.submit(f, check_notification=False)
+        f = self.response.forms['promotionform']
+        f['pcourse_id'] = 1
+        f['is_instructor'] = True
         self.submit(f)
         self.assertTitle("Nina Neubauer")
         self.assertPresence("0,00 €", div='balance')
         self.assertPresence("CdE-Mitglied", div="cde-membership")
         self.assertPresence("Probemitgliedschaft", div="cde-membership")
+        self.assertPresence("PfingstAkademie", div="past-events")
+        self.assertPresence("Swish", div="past-events")
+        self.assertPresence("Kursleiter", div="past-events")
 
         # check for correct welcome mail
         mail = self.fetch_mail_content()
@@ -1583,7 +1750,7 @@ class TestCoreFrontend(FrontendTest):
         self.assertTitle("Bereichsänderung für Kalif Karabatschi")
         # Now, do it right
         f['birthday'] = "21.6.1977"
-        f['gender'] = 1
+        f['gender'] = const.Genders.female
         f['change_note'] = "Komplizierte Aufnahme"
         self.submit(f)
         self.assertTitle("Kalif Karabatschi")
@@ -1595,6 +1762,7 @@ class TestCoreFrontend(FrontendTest):
     @as_users("vera")
     def test_ignore_warnings_postal_code(self) -> None:
         self.admin_view_profile("vera")
+
         self.traverse({'description': 'Bearbeiten \\(normal\\)'})
         f = self.response.forms['changedataform']
         f['postal_code'] = "11111"
@@ -1603,8 +1771,10 @@ class TestCoreFrontend(FrontendTest):
         self.assertPresence("Ungültige Postleitzahl")
         self.assertPresence("Warnungen ignorieren")
         f = self.response.forms['changedataform']
-        self.submit(f, button="ignore_warnings")
+        f[IGNORE_WARNINGS_NAME].checked = True
+        self.submit(f)
         self.assertTitle("Vera Verwaltung")
+
         self.traverse({'description': 'Bearbeiten \\(mit Adminrechten\\)'})
         f = self.response.forms['changedataform']
         self.assertNonPresence("Warnungen ignorieren")
@@ -1612,7 +1782,9 @@ class TestCoreFrontend(FrontendTest):
         self.assertPresence("Ungültige Postleitzahl")
         self.assertPresence("Warnungen ignorieren")
         f = self.response.forms['changedataform']
-        self.submit(f, button="ignore_warnings")
+        f[IGNORE_WARNINGS_NAME].checked = True
+        self.submit(f)
+
         self.get("/core/genesis/request")
         self.assertTitle("Account anfordern")
         f = self.response.forms['genesisform']
@@ -1631,26 +1803,28 @@ class TestCoreFrontend(FrontendTest):
         self.assertPresence("Ungültige Postleitzahl")
         self.assertPresence("Warnungen ignorieren")
         f = self.response.forms['genesisform']
-        self.submit(f, button="ignore_warnings")
+        f[IGNORE_WARNINGS_NAME].checked = True
+        self.submit(f)
         link = self.fetch_link()
         self.get(link)
-        self.follow()
-        self.traverse({'description': 'Accountanfragen'},
-                      {'description': 'Details'},
-                      {'description': 'Bearbeiten'})
+
+        self.traverse('Accountanfragen', 'Details', 'Accountanfrage bearbeiten')
         f = self.response.forms['genesismodifyform']
         self.assertNonPresence("Warnungen ignorieren")
         self.submit(f, check_notification=False)
         self.assertPresence("Ungültige Postleitzahl")
         self.assertPresence("Warnungen ignorieren")
         f = self.response.forms['genesismodifyform']
-        self.submit(f, button="ignore_warnings")
-        f = self.response.forms['genesiseventapprovalform']
+        f[IGNORE_WARNINGS_NAME].checked = True
         self.submit(f)
+        f = self.response.forms['genesisdecisionform']
+        self.submit(f, button="decision", value=str(GenesisDecision.approve))
 
     def _genesis_request(self, data: CdEDBObject, realm: Optional[str] = None) -> None:
         if realm:
             self.get('/core/genesis/request?realm=' + realm)
+        elif self.user:
+            self.get('/core/genesis/request')
         else:
             self.get('/')
             self.traverse({'description': 'Account anfordern'})
@@ -1661,7 +1835,6 @@ class TestCoreFrontend(FrontendTest):
         self.submit(f)
         link = self.fetch_link()
         self.get(link)
-        self.follow()
 
     ML_GENESIS_DATA_NO_REALM: CdEDBObject = {
         'given_names': "Zelda", 'family_name': "Zeruda-Hime",
@@ -1670,7 +1843,7 @@ class TestCoreFrontend(FrontendTest):
 
     EVENT_GENESIS_DATA = ML_GENESIS_DATA.copy()
     EVENT_GENESIS_DATA.update({
-        'realm': "event", 'gender': const.Genders.female.value,
+        'realm': "event", 'gender': const.Genders.female,
         'birthday': "1987-06-05", 'address': "An der Eiche", 'postal_code': "12345",
         'location': "Marcuria", 'country': "AQ"
     })
@@ -1722,9 +1895,13 @@ class TestCoreFrontend(FrontendTest):
             div='no-ml-request')
         self.traverse({'href': '/core/genesis/1001/show'})
         self.assertTitle("Accountanfrage von Zelda Zeruda-Hime")
-        f = self.response.forms['genesiseventapprovalform']
-        self.submit(f)
+        f = self.response.forms['genesisdecisionform']
+        self.submit(f, button="decision", value=str(GenesisDecision.approve))
         link = self.fetch_link()
+        self.submit(f, button="decision", value=str(GenesisDecision.approve),
+                    check_notification=False)
+        self.assertPresence("Emailadresse bereits vergeben.", div="notifications")
+        self.assertTitle("Accountanfrage von Zelda Zeruda-Hime")
         self.logout()
         self.get(link)
         self.assertTitle("Neues Passwort setzen")
@@ -1750,7 +1927,7 @@ class TestCoreFrontend(FrontendTest):
         user = USER_DICT['vera']
         self._genesis_request(self.ML_GENESIS_DATA_NO_REALM, realm='ml')
         self.login(user)
-        self.traverse({'description': 'Accountanfrage'})
+        self.traverse('Accountanfragen')
         self.assertTitle("Accountanfragen")
         self.assertPresence("zelda@example.cde", div='request-1001')
         self.assertPresence(
@@ -1758,8 +1935,9 @@ class TestCoreFrontend(FrontendTest):
             div='no-event-request')
         self.assertNonPresence(
             "Aktuell stehen keine Mailinglisten-Account-Anfragen zur Bestätigung aus.")
-        f = self.response.forms['genesismlapprovalform1']
-        self.submit(f)
+        self.traverse("Details")
+        f = self.response.forms['genesisdecisionform']
+        self.submit(f, button="decision", value=str(GenesisDecision.approve))
         link = self.fetch_link()
         self.logout()
         self.get(link)
@@ -1789,6 +1967,9 @@ class TestCoreFrontend(FrontendTest):
         f = self.response.forms['genesisform']
         for field, entry in self.CDE_GENESIS_DATA.items():
             f[field] = entry
+        self.submit(f, check_notification=False)
+        self.assertValidationError('attachment')
+        f = self.response.forms['genesisform']
         f['birth_name'] = "Ganondorf"
         f['notes'] = ""  # Do not send this to test upload permanance.
         with open(self.testfile_dir / "form.pdf", 'rb') as datafile:
@@ -1807,7 +1988,6 @@ class TestCoreFrontend(FrontendTest):
         self.submit(f)
         link = self.fetch_link()
         self.get(link)
-        self.follow()
         self.login(USER_DICT["vera"])
         self.traverse({'href': '/core'},
                       {'href': '/core/genesis/list'})
@@ -1823,13 +2003,24 @@ class TestCoreFrontend(FrontendTest):
         self.traverse({'href': '/core/genesis/1001/show'})
         self.assertPresence("Ganondorf")
         self.assertNonPresence("Zickzack")
+        self.assertNonPresence("PfingstAkademie")
         self.traverse({'href': '/core/genesis/1001/modify'})
         self.assertTitle("Accountanfrage bearbeiten")
         f = self.response.forms['genesismodifyform']
         f['birth_name'] = "Zickzack"
+        f['pevent_id'] = 1
         self.submit(f)
         self.assertPresence("Zickzack")
         self.assertNonPresence("Ganondorf")
+        self.assertPresence("PfingstAkademie 2014")
+
+        self.traverse({'href': '/core/genesis/1001/modify'})
+        self.assertTitle("Accountanfrage bearbeiten")
+        f = self.response.forms['genesismodifyform']
+        f['pcourse_id'] = 2
+        self.submit(f)
+        self.assertPresence("Goethe")
+
         self.traverse({'href': '/core/genesis/1001/modify'})
         self.assertTitle("Accountanfrage bearbeiten")
         f = self.response.forms['genesismodifyform']
@@ -1867,8 +2058,8 @@ class TestCoreFrontend(FrontendTest):
             "Aktuell stehen keine CdE-Mitglieds-Account-Anfragen zur Bestätigung aus.")
         self.traverse({'href': '/core/genesis/1001/show'})
         self.assertTitle("Accountanfrage von Zelda Zeruda-Hime")
-        f = self.response.forms['genesiseventapprovalform']
-        self.submit(f)
+        f = self.response.forms['genesisdecisionform']
+        self.submit(f, button="decision", value=str(GenesisDecision.approve))
         link = self.fetch_link()
         self.traverse({'href': '^/$'})
         f = self.response.forms['adminshowuserform']
@@ -1900,6 +2091,8 @@ class TestCoreFrontend(FrontendTest):
         self.assertCheckbox(True, "paper_expuls_checkbox")
         self.assertPresence("12345")
         self.assertPresence("Zickzack")
+        self.assertPresence("PfingstAkademie 2014")
+        self.assertPresence("Goethe")
         self.traverse({'href': '/cde'})
         self.assertTitle('CdE-Mitgliederbereich')
         self.traverse({'description': 'Sonstiges'})
@@ -1916,6 +2109,13 @@ class TestCoreFrontend(FrontendTest):
         # Submit twice
         self.submit(f, check_notification=False)
         self.assertPresence("Bestätigungsmail erneut versendet.",
+                            div="notifications")
+        link = self.fetch_link()
+        self.get(link)
+        self.follow()
+        # Submit thrice
+        self.submit(f, check_notification=False)
+        self.assertPresence("Deine Anfrage wartet derzeit auf Bestätigung.",
                             div="notifications")
         self.get('/')
         self.traverse({'description': 'Account anfordern'})
@@ -1934,9 +2134,16 @@ class TestCoreFrontend(FrontendTest):
         self.login(user)
         self.traverse({'description': 'Account-Log'})
         f = self.response.forms['logshowform']
-        f['codes'] = [20]
+        f['codes'] = [const.CoreLogCodes.genesis_request]
         self.submit(f)
         self.assertTitle("Account-Log [1–1 von 1]")
+        self.admin_view_profile("hades")
+        self.traverse("Account wiederherstellen")
+        f = self.response.forms['dearchivepersonaform']
+        f['new_username'] = self.ML_GENESIS_DATA_NO_REALM['username']
+        self.submit(f, check_notification=False)
+        self.assertValidationError(
+            'new_username', "Ein Benutzer mit dieser E-Mail-Adresse existiert bereits.")
 
     def test_genesis_verification_mail_resend(self) -> None:
         self.get('/')
@@ -1996,16 +2203,16 @@ class TestCoreFrontend(FrontendTest):
         self.traverse({'description': 'Accountanfrage'},
                       {'description': 'Details'})
         self.assertTitle("Accountanfrage von Zelda Zeruda-Hime")
-        self.traverse({'description': 'Bearbeiten'})
+        self.traverse('Accountanfrage bearbeiten')
         f = self.response.forms['genesismodifyform']
         f['family_name'] = "Zeruda"
         self.submit(f)
         self.assertTitle("Accountanfrage von Zelda Zeruda")
 
-        self.traverse({'description': 'Bearbeiten'})
+        self.traverse('Accountanfrage bearbeiten')
         f = self.response.forms['genesismodifyform']
         f['realm'] = "event"
-        f['gender'] = "1"
+        f['gender'] = const.Genders.female
         f['birthday'] = "1987-06-05"
         f['address'] = "An der Eiche"
         f['postal_code'] = "12345"
@@ -2015,13 +2222,205 @@ class TestCoreFrontend(FrontendTest):
         self.assertPresence("An der Eiche", div='address')
         self.assertPresence("Antarktis", div='address')
 
-        self.traverse({'description': 'Bearbeiten'})
+        self.traverse('Accountanfrage bearbeiten')
         f = self.response.forms['genesismodifyform']
         f['birthday'] = "1987-06-05"
         self.submit(f)
+
+        # Check that we cannot assign a username that is already taken.
+        f['username'] = self.user['username']
+        self.submit(f, check_notification=False)
+        self.assertValidationError("username", "Emailadresse bereits vergeben.")
+        self.traverse("Abbrechen")
+
         self.assertTitle("Accountanfrage von Zelda Zeruda")
-        f = self.response.forms['genesiseventapprovalform']
+        f = self.response.forms['genesisdecisionform']
+        self.submit(f, button="decision", value=str(GenesisDecision.approve),
+                    check_notification=False)
+
+    def _create_genesis_doppelganger(self, user: UserIdentifier = None) -> UserObject:
+        user = get_user(user or self.user)
+        # Create a new request almost identical to the current user.
+        dg_data = {k: user[k] for k in self.ML_GENESIS_DATA if k in user}
+        dg_data["username"] = "notme@example.cde"
+        dg_data["notes"] = "Bestimmt jemand anderes1"
+        dg_data["realm"] = "ml"
+        self._genesis_request(dg_data)
+        return dg_data
+
+    @as_users("vera")
+    def test_genesis_doppelganger(self) -> None:
+        dg_data = self._create_genesis_doppelganger()
+        log_expectation = [
+            {
+                'code': const.CoreLogCodes.genesis_request,
+                'change_note': dg_data['username'],
+            },
+            {
+                'code': const.CoreLogCodes.genesis_verified,
+                'change_note': dg_data['username'],
+            },
+        ]
+
+        self.traverse("Accountanfragen", "Details")
+        self.assertTitle(f"Accountanfrage von {self.user['given_names']}"
+                         f" {self.user['family_name']}")
+        self.assertPresence("Ähnliche Accounts")
+        self.assertPresence(self.user['username'], div="doppelgangers")
+        f = self.response.forms['genesisdecisionform']
+        # Rejection causes info not success notification.
+        self.submit(f, button="decision", value=str(GenesisDecision.deny),
+                    check_notification=False, verbose=True)
+        self.assertPresence("Anfrage abgewiesen", div="notifications")
+        log_expectation.append({
+            'code': const.CoreLogCodes.genesis_rejected,
+            'change_note': dg_data['username'],
+        })
+
+        # Create two almost identical requests, approve one and check that the second
+        # one finds a doppelgänger.
+        self._genesis_request(self.EVENT_GENESIS_DATA)
+        log_expectation.extend([
+            {
+                'code': const.CoreLogCodes.genesis_request,
+                'change_note': self.EVENT_GENESIS_DATA['username'],
+            },
+            {
+                'code': const.CoreLogCodes.genesis_verified,
+                'change_note': self.EVENT_GENESIS_DATA['username'],
+            },
+        ])
+        alternate_username = "notzelda@example.cde"
+        self._genesis_request(
+            dict(self.EVENT_GENESIS_DATA, username=alternate_username))
+        log_expectation.extend([
+            {
+                'code': const.CoreLogCodes.genesis_request,
+                'change_note': alternate_username,
+            },
+            {
+                'code': const.CoreLogCodes.genesis_verified,
+                'change_note': alternate_username,
+            },
+        ])
+
+        # Approve the first request.
+        self.traverse("Accountanfragen", "Details")
+        self.assertTitle(f"Accountanfrage von {self.EVENT_GENESIS_DATA['given_names']}"
+                         f" {self.EVENT_GENESIS_DATA['family_name']}")
+        self.assertNonPresence("Ähnliche Accounts")
+        self.assertNonPresence("Wiederherstellen", div="genesisdecisionform")
+        f = self.response.forms['genesisdecisionform']
+        self.submit(f, button="decision", value=str(GenesisDecision.approve),
+                    check_notification=False)
+        new_persona_id = 1001
+        log_expectation.extend([
+            {
+                'code': const.CoreLogCodes.persona_creation,
+                'persona_id': new_persona_id,
+            },
+            {
+                'code': const.CoreLogCodes.genesis_approved,
+                'change_note': self.EVENT_GENESIS_DATA['username'],
+                'persona_id': new_persona_id,
+            },
+            {
+                'code': const.CoreLogCodes.password_reset_cookie,
+                'persona_id': new_persona_id,
+            }
+        ])
+
+        # Check that the first username exists and the alternate one does not.
+        self.assertTrue(self.core.verify_existence(
+            self.key, self.EVENT_GENESIS_DATA['username']))
+        self.assertFalse(self.core.verify_existence(
+            self.key, alternate_username, include_genesis=False))
+
+        # Check that the second request finds the new account as a doppelgänger and
+        # update that account to the alternate username.
+        self.traverse("Accountanfragen", "Details")
+        self.assertTitle(f"Accountanfrage von {self.EVENT_GENESIS_DATA['given_names']}"
+                         f" {self.EVENT_GENESIS_DATA['family_name']}")
+        self.assertPresence("Ähnliche Accounts")
+        self.assertPresence(self.EVENT_GENESIS_DATA['username'], div="doppelgangers")
+
+        # Check that a cde genesis request cannot be merged into a non-cde account.
+        self.traverse("Accountanfrage bearbeiten")
+        f = self.response.forms['genesismodifyform']
+        f['realm'] = "cde"
         self.submit(f)
+        f = self.response.forms['genesisdecisionform']
+        # Set persona_id to the value of the second radio button.
+        f['persona_id'] = f['persona_id'].options[1][0]
+        self.submit(f, button="decision", value=str(GenesisDecision.update),
+                    check_notification=False)
+        self.assertPresence("Ungültiger Benutzer für Aktualisierung."
+                            " Füge zunächst folgenden Bereich hinzu: cde.",
+                            div="notifications")
+        # Repair the request.
+        self.traverse("Accountanfrage bearbeiten")
+        f = self.response.forms['genesismodifyform']
+        f['realm'] = "event"
+        self.submit(f)
+
+        # Check that approving the request fails if a persona is selected.
+        f = self.response.forms['genesisdecisionform']
+        f['persona_id'] = f['persona_id'].options[1][0]
+        self.submit(f, button="decision", value=str(GenesisDecision.approve),
+                    check_notification=False)
+        self.assertPresence("Existierender Account ausgewählt,"
+                            " aber Accountanfrage bestätigt.", div="notifications")
+
+        # Now merge the genesis request into the existing account.
+        # Submit without selecting doppelganger.
+        f = self.response.forms['genesisdecisionform']
+        f['persona_id'] = ""
+        self.submit(f, button="decision", value=str(GenesisDecision.update),
+                    check_notification=False)
+        self.assertPresence("Kein Account ausgewählt.", div="notifications")
+        # Now for real.
+        f['persona_id'] = f['persona_id'].options[1][0]
+        self.submit(f, button="decision", value=str(GenesisDecision.update))
+        self.assertPresence("Benutzer aktualisiert", div="notifications")
+        log_expectation.extend([
+            {
+                'code': const.CoreLogCodes.genesis_merged,
+                'persona_id': new_persona_id,
+            },
+            {
+                'code': const.CoreLogCodes.username_change,
+                'change_note': alternate_username,
+                'persona_id': new_persona_id,
+            },
+            {
+                'code': const.CoreLogCodes.password_reset_cookie,
+                'persona_id': new_persona_id,
+            }
+        ])
+
+        self.assertLogEqual(
+            log_expectation, realm="core", offset=len(self.get_sample_data("core.log")))
+
+        # Check that the first username no longer exists, but the alternate one does.
+        self.assertFalse(self.core.verify_existence(
+            self.key, self.EVENT_GENESIS_DATA['username']))
+        self.assertTrue(self.core.verify_existence(
+            self.key, alternate_username))
+
+    @as_users("vera")
+    def test_genesis_dearchive_doppelganger(self) -> None:
+        hades = get_user("hades")
+        self._create_genesis_doppelganger(hades)
+        self.traverse("Accountanfragen", "Details")
+        self.assertTitle(f"Accountanfrage von {hades['given_names']}"
+                         f" {hades['family_name']}")
+        self.assertPresence(f"{hades['given_names']} {hades['family_name']}",
+                            div="doppelgangers")
+        self.assertPresence("(archiviert)", div="doppelgangers")
+        f = self.response.forms['genesisdecisionform']
+        f['persona_id'] = hades['id']
+        self.submit(f, button="decision", value=str(GenesisDecision.update))
+        self.assertPresence("Benutzer aktualisiert.", div="notifications")
 
     def test_resolve_api(self) -> None:
         at = urllib.parse.quote_plus('@')
@@ -2044,6 +2443,12 @@ class TestCoreFrontend(FrontendTest):
             "is_member": True,
             "id": 1,
             "username": "anton@example.cde",
+        })
+        self.get(
+            '/core/api/resolve?username=antonatexample.cde',
+            headers={'X-CdEDB-API-token': 'a1o2e3u4i5d6h7t8n9s0'})
+        self.assertEqual(self.response.json, {
+            'error':  ["('username', ValueError('Must be a valid email address.'))"]
         })
         self.get('/core/api/resolve', status=403)
 
@@ -2070,18 +2475,18 @@ class TestCoreFrontend(FrontendTest):
 
         # approve the account requests
         self.login(user)
-        self.traverse({'description': 'Accountanfragen'})
-        f = self.response.forms['genesismlapprovalform1']
-        self.submit(f)
-        logs.append((1005, const.CoreLogCodes.genesis_approved))
-        logs.append((1006, const.CoreLogCodes.persona_creation))
+        self.traverse("Accountanfragen", "Details")
+        f = self.response.forms['genesisdecisionform']
+        self.submit(f, button="decision", value=str(GenesisDecision.approve))
+        logs.append((1005, const.CoreLogCodes.persona_creation))
+        logs.append((1006, const.CoreLogCodes.genesis_approved))
         logs.append((1007, const.CoreLogCodes.password_reset_cookie))
 
-        self.traverse({'href': 'core/genesis/1002/show'})
-        f = self.response.forms['genesiseventapprovalform']
-        self.submit(f)
-        logs.append((1008, const.CoreLogCodes.genesis_approved))
-        logs.append((1009, const.CoreLogCodes.persona_creation))
+        self.traverse("Details")
+        f = self.response.forms['genesisdecisionform']
+        self.submit(f, button="decision", value=str(GenesisDecision.approve))
+        logs.append((1008, const.CoreLogCodes.persona_creation))
+        logs.append((1009, const.CoreLogCodes.genesis_approved))
         logs.append((1010, const.CoreLogCodes.password_reset_cookie))
 
         # make janis assembly user
@@ -2115,3 +2520,19 @@ class TestCoreFrontend(FrontendTest):
         self.assertPresence(promotion_change_note)
         self.assertPresence("zelda@example.cde")
         self.assertPresence("bertalotta@example.cde")
+
+    @as_users("katarina")
+    def test_auditor(self) -> None:
+        realm_logs = {
+            "Index": ("Account-Log", "Nutzerdaten-Log",),
+            "Mitglieder": ("CdE-Log", "Finanz-Log", "Verg.-Veranstaltungen-Log",),
+            "Veranstaltungen": ("Log",),
+            "Mailinglisten": ("Log",),
+            "Versammlungen": ("Log",),
+        }
+        for realm, logs in realm_logs.items():
+            self.traverse(realm, *logs, realm)
+            self._click_admin_view_button("Kassenprüfer")
+            for log in logs:
+                self.assertNonPresence(log, div="sidebar-navigation")
+            self._click_admin_view_button("Kassenprüfer")
