@@ -4,15 +4,23 @@
 import csv
 import re
 import unittest.mock
-from typing import Any, List, Tuple
+from typing import Any, List, Set, Tuple
+
+import webtest
 
 import cdedb.database.constants as const
 from cdedb.common import ADMIN_VIEWS_COOKIE_NAME, CdEDBObject
-from cdedb.devsamples import MockHeldMessage, HELD_MESSAGE_SAMPLE
+from cdedb.devsamples import HELD_MESSAGE_SAMPLE, MockHeldMessage
 from cdedb.frontend.common import CustomCSVDialect
 from cdedb.ml_type_aux import CdeLokalMailinglist
 from cdedb.query import QueryOperators
 from tests.common import USER_DICT, FrontendTest, as_users, prepsql
+
+
+def _get_registration_part_stati(f: webtest.Form) -> Set[const.RegistrationPartStati]:
+    return set(filter(
+        None, (f.get("registration_stati", index=i).value
+               for i in range(len(const.RegistrationPartStati)))))
 
 
 class TestMlFrontend(FrontendTest):
@@ -389,7 +397,7 @@ class TestMlFrontend(FrontendTest):
         # Test if moderation hints work
         self.assertPresence("Mailman-Migration", div="mailinglist-99-row")
         self.assertPresence("Mitglieder (Opt-in)", div="mailinglist-99-row")
-        self.assertPresence("3", div="mailinglist-99-row")
+        self.assertPresence("0", div="mailinglist-99-row")
         self.assertPresence("0 Abonnenten. 1 Moderator.", div="mailinglist-99-row")
         # Test that events are shown
         self.assertPresence("Große Testakademie 2222")
@@ -426,7 +434,7 @@ class TestMlFrontend(FrontendTest):
         self.assertPresence("Mitgliedermailinglisten")
         self.assertPresence("Aktivenforum 2001", div="mailinglist-7-row")
         self.assertPresence("Mitglieder (Opt-in)", div="mailinglist-7-row")
-        self.assertPresence("3", div="mailinglist-7-row")
+        self.assertPresence("0", div="mailinglist-7-row")
         self.assertPresence("1 Abonnent. 2 Moderatoren.", div="mailinglist-7-row")
         if self.user_in('berta'):
             self.assertPresence("Veranstaltungslisten")
@@ -768,18 +776,18 @@ class TestMlFrontend(FrontendTest):
         self.traverse({'description': 'Mailingliste anlegen'})
         self.assertTitle("Mailingliste anlegen")
         f = self.response.forms['selectmltypeform']
-        f['ml_type'] = const.MailinglistTypes.member_mandatory.value
+        f['ml_type'] = const.MailinglistTypes.member_mandatory
         self.submit(f)
         f = self.response.forms['createlistform']
         self.assertEqual(f['maxsize'].value, '64')
         f['title'] = "Munkelwand"
-        f['mod_policy'] = 1
-        f['attachment_policy'] = 2
-        f['subject_prefix'] = "[munkel]"
+        f['mod_policy'] = const.ModerationPolicy.unmoderated
+        f['attachment_policy'] = const.AttachmentPolicy.pdf_only
+        f['subject_prefix'] = "munkel"
         f['maxsize'] = 512
         f['is_active'].checked = True
         f['notes'] = "Noch mehr Gemunkel."
-        f['domain'] = 1
+        f['domain'] = const.MailinglistDomain.lists
         f['local_part'] = 'munkelwand'
         # Check that there must be some moderators
         errormsg = "Darf nicht leer sein."
@@ -820,7 +828,7 @@ class TestMlFrontend(FrontendTest):
         # Check if the attachment policy hint works
         self.assertPresence("Admin-Team")
         f = self.response.forms['changelistform']
-        f['domain'] = const.MailinglistDomain.testmail.value
+        f['domain'] = const.MailinglistDomain.testmail
         f['maxsize'] = "intentionally no valid maxsize"
         self.submit(f, check_notification=False)
         self.assertValidationError("maxsize", "Ungültige Eingabe für eine Ganzzahl.")
@@ -831,7 +839,7 @@ class TestMlFrontend(FrontendTest):
         self.assertTitle("Werbung")
         self.traverse({'href': '/ml/mailinglist/2/change'}, )
         self.assertPresence("Admin-Team")
-        f['domain'] = const.MailinglistDomain.lists.value
+        f['domain'] = const.MailinglistDomain.lists
         self.submit(f)
         self.assertTitle("Werbung")
         self.traverse({'href': '/ml/mailinglist/2/change'}, )
@@ -896,11 +904,11 @@ class TestMlFrontend(FrontendTest):
 
         for ml_type in const.MailinglistTypes:
             with self.subTest(ml_type=ml_type):
-                f['ml_type'] = ml_type.value
+                f['ml_type'] = ml_type
                 f['event_id'] = event_id
                 f['registration_stati'] = [
-                    const.RegistrationPartStati.participant.value,
-                    const.RegistrationPartStati.waitlist.value,
+                    const.RegistrationPartStati.participant,
+                    const.RegistrationPartStati.waitlist,
                 ]
                 f['assembly_id'] = assembly_id
                 # no ml type should allow event *and* assembly fields to be set
@@ -950,12 +958,17 @@ class TestMlFrontend(FrontendTest):
         self.get("/ml/mailinglist/9/change")
         self.assertTitle("Teilnehmer-Liste – Konfiguration")
         f = self.response.forms['changelistform']
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
-        self.assertEqual({"2", "4", None}, tmp)
-        f['registration_stati'] = [3, 5]
+        reality = _get_registration_part_stati(f)
+        expectation = {str(const.RegistrationPartStati.participant),
+                       str(const.RegistrationPartStati.guest)}
+        self.assertEqual(expectation, reality)
+        f['registration_stati'] = [const.RegistrationPartStati.waitlist,
+                                   const.RegistrationPartStati.cancelled]
         self.submit(f)
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
-        self.assertEqual({"3", "5", None}, tmp)
+        reality = _get_registration_part_stati(f)
+        expectation = {str(const.RegistrationPartStati.waitlist),
+                       str(const.RegistrationPartStati.cancelled)}
+        self.assertEqual(expectation, reality)
 
     @as_users("nina")
     def test_delete_ml(self) -> None:
@@ -1045,9 +1058,9 @@ class TestMlFrontend(FrontendTest):
                 self.login("anton")
                 mdata = {
                     'title': 'TestAkaList',
-                    'ml_type': const.MailinglistTypes.event_associated.value,
+                    'ml_type': const.MailinglistTypes.event_associated,
                     'local_part': 'testaka',
-                    'domain': const.MailinglistDomain.aka.value,
+                    'domain': const.MailinglistDomain.aka,
                     'event_id': "1",
                     'moderators': user['DB-ID'],
                 }
@@ -1083,7 +1096,6 @@ class TestMlFrontend(FrontendTest):
         self.assertTitle("Klatsch und Tratsch")
         link = self.fetch_link()
         self.get(link)
-        self.follow()
         self.assertTitle("Klatsch und Tratsch")
         self.assertIn('unsubscribeform', self.response.forms)
         self.assertPresence('pepper@example.cde')
@@ -1116,27 +1128,21 @@ class TestMlFrontend(FrontendTest):
         f = self.response.forms['addsubscriberform']
         f['subscriber_ids'] = USER_DICT["inga"]["DB-ID"]
         self.submit(f, check_notification=False)
-        self.assertIn("alert alert-danger", self.response.text)
-        self.assertPresence(
-            "Der Nutzer hat aktuell eine Abonnement-Anfrage gestellt.",
-            div="notifications")
+        self.assertNotification(
+            "Der Nutzer hat aktuell eine Abonnement-Anfrage gestellt.", 'error')
         # as mod subscriber
         self.traverse({'href': '/ml/mailinglist/4/management/advanced'}, )
         f = self.response.forms['addmodsubscriberform']
         f['modsubscriber_ids'] = USER_DICT["inga"]["DB-ID"]
         self.submit(f, check_notification=False)
-        self.assertIn("alert alert-danger", self.response.text)
-        self.assertPresence(
-            "Der Nutzer hat aktuell eine Abonnement-Anfrage gestellt.",
-            div="notifications")
+        self.assertNotification(
+            "Der Nutzer hat aktuell eine Abonnement-Anfrage gestellt.", 'error')
         # as mod unsubscribe
         f = self.response.forms['addmodunsubscriberform']
         f['modunsubscriber_ids'] = USER_DICT["inga"]["DB-ID"]
         self.submit(f, check_notification=False)
-        self.assertIn("alert alert-danger", self.response.text)
-        self.assertPresence(
-            "Der Nutzer hat aktuell eine Abonnement-Anfrage gestellt.",
-            div="notifications")
+        self.assertNotification(
+            "Der Nutzer hat aktuell eine Abonnement-Anfrage gestellt.", 'error')
 
         # testing: mod subscribe and unsubscribe
         # add already subscribed user as mod subscribe
@@ -1151,15 +1157,12 @@ class TestMlFrontend(FrontendTest):
         self.traverse({'href': '/ml/mailinglist/4/management'}, )
         f = self.response.forms['removesubscriberform1']
         self.submit(f, check_notification=False)
-        self.assertIn("alert alert-danger", self.response.text)
-        self.assertPresence("Der Nutzer ist aktuell fixierter Abonnent.",
-                            div="notifications")
+        self.assertNotification("Der Nutzer ist aktuell fixierter Abonnent.", 'error')
         # try to add a mod unsubscribed user
         f = self.response.forms['addsubscriberform']
         f['subscriber_ids'] = USER_DICT["garcia"]["DB-ID"]
         self.submit(f, check_notification=False)
-        self.assertIn("alert alert-danger", self.response.text)
-        self.assertPresence("Der Nutzer ist aktuell blockiert.", div="notifications")
+        self.assertNotification("Der Nutzer ist aktuell blockiert.", 'error')
 
     @as_users("berta", "janis")
     def test_moderator_access(self) -> None:
@@ -1192,13 +1195,13 @@ class TestMlFrontend(FrontendTest):
         f['event_id'].force_value(1)
         f['is_active'].force_value(False)
         # these properties can be changed by full moderators only
-        f['registration_stati'] = [const.RegistrationPartStati.guest.value]
+        f['registration_stati'] = [const.RegistrationPartStati.guest]
         # these properties can be changed by every moderator
         f['description'] = "Wir machen Party!"
         f['notes'] = "Nur geladene Gäste."
-        f['mod_policy'] = const.ModerationPolicy.unmoderated.value
+        f['mod_policy'] = const.ModerationPolicy.unmoderated
         f['subject_prefix'] = "party"
-        f['attachment_policy'] = const.AttachmentPolicy.allow.value
+        f['attachment_policy'] = const.AttachmentPolicy.allow
         f['maxsize'] = 1111
         self.submit(f)
 
@@ -1211,21 +1214,21 @@ class TestMlFrontend(FrontendTest):
         self.assertEqual(str(old_ml['event_id']), f['event_id'].value)
 
         # ... these have only changed if the moderator is privileged ...
-        reality = {f.get("registration_stati", index=i).value for i in range(7)}
+        reality = _get_registration_part_stati(f)
         if self.user_in('berta'):
-            expectation = {None, str(const.RegistrationPartStati.guest.value)}
+            expectation = {str(const.RegistrationPartStati.guest)}
         else:
-            expectation = {str(status)
-                           for status in old_ml['registration_stati']} | {None}
+            expectation = {str(const.RegistrationPartStati(status))
+                           for status in old_ml['registration_stati']}
         self.assertEqual(expectation, reality)
 
         # ... and these have changed.
         self.assertEqual("Wir machen Party!", f['description'].value)
         self.assertEqual("Nur geladene Gäste.", f['notes'].value)
-        self.assertEqual(str(const.ModerationPolicy.unmoderated.value),
+        self.assertEqual(str(const.ModerationPolicy.unmoderated),
                          f['mod_policy'].value)
         self.assertEqual("party", f['subject_prefix'].value)
-        self.assertEqual(str(const.AttachmentPolicy.allow.value),
+        self.assertEqual(str(const.AttachmentPolicy.allow),
                          f['attachment_policy'].value)
         self.assertEqual("1111", f['maxsize'].value)
 
@@ -1325,7 +1328,7 @@ class TestMlFrontend(FrontendTest):
         self.traverse({"description": "Mailinglisten"},
                       {"description": "Mailingliste anlegen"})
         f = self.response.forms['selectmltypeform']
-        f['ml_type'] = const.MailinglistTypes.cdelokal.value
+        f['ml_type'] = const.MailinglistTypes.cdelokal
         self.assertEqual(len(f['ml_type'].options), 2)
         self.submit(f)
         f = self.response.forms['createlistform']
@@ -1335,7 +1338,7 @@ class TestMlFrontend(FrontendTest):
         f['description'] = "If anyone else lives here, please come by, " \
                            "I am lonely."
         f['local_part'] = "littlewhinging"
-        f['domain'] = const.MailinglistDomain.cdelokal.value
+        f['domain'] = const.MailinglistDomain.cdelokal
         self.assertEqual(len(f['domain'].options),
                          len(CdeLokalMailinglist.domains))
         moderator = USER_DICT["berta"]
@@ -1348,28 +1351,28 @@ class TestMlFrontend(FrontendTest):
     def test_1342(self) -> None:
         self.get("/ml/mailinglist/60/change")
         f = self.response.forms['changelistform']
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
+        reality = _get_registration_part_stati(f)
         sample_data_stati = set(
-            str(x) for x in self.get_sample_data(
+            str(const.RegistrationPartStati(x)) for x in self.get_sample_data(
                 "ml.mailinglists", (60,), ("registration_stati",)
             )[60]["registration_stati"])
-        self.assertEqual(sample_data_stati | {None}, tmp)
-        stati = [const.RegistrationPartStati.waitlist.value,
-                 const.RegistrationPartStati.guest.value]
+        self.assertEqual(sample_data_stati, reality)
+        stati = [const.RegistrationPartStati.waitlist,
+                 const.RegistrationPartStati.guest]
         f['registration_stati'] = stati
         self.submit(f)
         self.traverse({"description": "Konfiguration"})
         f = self.response.forms['changelistform']
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
-        self.assertEqual({str(x) for x in stati} | {None}, tmp)
+        reality = _get_registration_part_stati(f)
+        self.assertEqual({str(x) for x in stati}, reality)
 
-        stati = [const.RegistrationPartStati.not_applied.value]
+        stati = [const.RegistrationPartStati.not_applied]
         f['registration_stati'] = stati
         self.submit(f)
         self.traverse({"description": "Konfiguration"})
         f = self.response.forms['changelistform']
-        tmp = {f.get("registration_stati", index=i).value for i in range(7)}
-        self.assertEqual({str(x) for x in stati} | {None}, tmp)
+        reality = _get_registration_part_stati(f)
+        self.assertEqual({str(x) for x in stati}, reality)
 
     @staticmethod
     def _prepare_moderation_mock(client_class: unittest.mock.Mock) -> Tuple[
@@ -1381,6 +1384,7 @@ class TestMlFrontend(FrontendTest):
         mmlist.moderate_message.return_value = moderation_response
         client = client_class.return_value
         client.get_held_messages.return_value = messages
+        client.get_held_message_count.return_value = len(messages)
         client.get_list_safe.return_value = mmlist
 
         return messages, mmlist, client
@@ -1403,19 +1407,25 @@ class TestMlFrontend(FrontendTest):
         self.assertPresence("Finanzbericht")
         self.assertPresence("Verschwurbelung")
         self.assertPresence("unerwartetes Erbe")
+        mmlist.get_held_message.return_value = messages[0]
         client.get_held_messages.return_value = messages[1:]
+        client.get_held_message_count.return_value = len(messages[1:])
         f = self.response.forms['msg1']
         self.submit(f, button='action', value='accept')
         self.assertNonPresence("Finanzbericht")
         self.assertPresence("Verschwurbelung")
         self.assertPresence("unerwartetes Erbe")
+        mmlist.get_held_message.return_value = messages[1]
         client.get_held_messages.return_value = messages[2:]
+        client.get_held_message_count.return_value = len(messages[2:])
         f = self.response.forms['msg2']
         self.submit(f, button='action', value='reject')
         self.assertNonPresence("Finanzbericht")
         self.assertNonPresence("Verschwurbelung")
         self.assertPresence("unerwartetes Erbe")
+        mmlist.get_held_message.return_value = messages[2]
         client.get_held_messages.return_value = messages[3:]
+        client.get_held_message_count.return_value = len(messages[3:])
         f = self.response.forms['msg3']
         self.submit(f, button='action', value='discard')
         self.assertNonPresence("Finanzbericht")
@@ -1430,6 +1440,17 @@ class TestMlFrontend(FrontendTest):
         self.assertEqual(
             mmlist.moderate_message.call_args_list,
             [umcall(1, 'accept'), umcall(2, 'reject'), umcall(3, 'discard')])
+
+        self.traverse("Log")
+        self.assertPresence("Nachricht akzeptiert", div="1-1001")
+        self.assertPresence("Nachricht zurückgewiesen", div="2-1002")
+        self.assertPresence("Nachricht verworfen", div="3-1003")
+        self.assertPresence("kassenwart@example.cde / Finanzbericht / Spam score: —",
+                            div="1-1001")
+        self.assertPresence("illuminati@example.cde / Verschwurbelung"
+                            " / Spam score: 1.108", div="2-1002")
+        self.assertPresence("nigerian_prince@example.cde / unerwartetes Erbe"
+                            " / Spam score: 2.725", div="3-1003")
 
     @unittest.mock.patch("cdedb.frontend.common.CdEMailmanClient")
     @as_users("anton")
@@ -1450,6 +1471,8 @@ class TestMlFrontend(FrontendTest):
         self.assertPresence("Finanzbericht")
         self.assertPresence("Verschwurbelung")
         self.assertPresence("unerwartetes Erbe")
+        # placeholder
+        mmlist.get_held_message.return_value = messages[0]
         client.get_held_messages.return_value = []
         f = self.response.forms['moderateallform']
         self.submit(f, button='action', value='accept')
@@ -1485,6 +1508,8 @@ class TestMlFrontend(FrontendTest):
         self.assertPresence("Finanzbericht")
         self.assertPresence("Verschwurbelung")
         self.assertPresence("unerwartetes Erbe")
+        # placeholder
+        mmlist.get_held_message.return_value = messages[0]
         client.get_held_messages.return_value = []
         f = self.response.forms['moderateallform']
         self.submit(f, button='action', value='discard')
@@ -1518,6 +1543,7 @@ class TestMlFrontend(FrontendTest):
         self.assertTitle("Mailman-Migration – Nachrichtenmoderation")
         self.assertPresence("Finanzbericht")
         self.assertPresence("kassenwart@example.cde")
+        mmlist.get_held_message.return_value = messages[0]
         client.get_held_messages.return_value = messages[1:]
         f = self.response.forms['msg1']
         self.submit(f, button='action', value='whitelist')
@@ -1540,7 +1566,11 @@ class TestMlFrontend(FrontendTest):
         self.assertTitle("Mailinglisten-Log [1–9 von 9]")
         self.assertNonPresence("LogCodes")
         f = self.response.forms['logshowform']
-        f['codes'] = [10, 11, 20, 21, 22]
+        f['codes'] = [const.MlLogCodes.moderator_added,
+                      const.MlLogCodes.moderator_removed,
+                      const.MlLogCodes.subscription_requested,
+                      const.MlLogCodes.subscribed,
+                      const.MlLogCodes.subscription_changed]
         f['mailinglist_id'] = 4
 
         self.submit(f)
